Add Asset manager and remove GL code from launcher
