+++ src/smol_gl.cpp
@@ -1,0 +1,1297 @@
+
+#define SMOL_GL_DEFINE_EXTERN
+#include <smol/smol_gl.h>
+
+#if defined(SMOL_ENGINE_IMPLEMENTATION) && defined(SMOL_PLATFORM_WINDOWS)
+namespace smol
+{
+  static HMODULE opengl32Dll = GetModuleHandleA("OpenGL32.dll");
+  static void* win32GetGLFunctionPtr(char* name)
+  {
+      void* addr = GetProcAddress(opengl32Dll, name);
+      if(! addr) addr = wglGetProcAddress(name);
+      return addr;
+  }
+
+#define SMOL_GETGLPROC(param) win32GetGLFunctionPtr(#param)
+
+  void getOpenGLFunctionPointers()
+  {
+    glCullFace = (PFNGLCULLFACEPROC) SMOL_GETGLPROC(glCullFace); 
+    glFrontFace = (PFNGLFRONTFACEPROC) SMOL_GETGLPROC(glFrontFace); 
+    glHint = (PFNGLHINTPROC) SMOL_GETGLPROC(glHint); 
+    glLineWidth = (PFNGLLINEWIDTHPROC) SMOL_GETGLPROC(glLineWidth); 
+    glPointSize = (PFNGLPOINTSIZEPROC) SMOL_GETGLPROC(glPointSize); 
+    glPolygonMode = (PFNGLPOLYGONMODEPROC) SMOL_GETGLPROC(glPolygonMode); 
+    glScissor = (PFNGLSCISSORPROC) SMOL_GETGLPROC(glScissor); 
+    glTexParameterf = (PFNGLTEXPARAMETERFPROC) SMOL_GETGLPROC(glTexParameterf); 
+    glTexParameterfv = (PFNGLTEXPARAMETERFVPROC) SMOL_GETGLPROC(glTexParameterfv); 
+    glTexParameteri = (PFNGLTEXPARAMETERIPROC) SMOL_GETGLPROC(glTexParameteri); 
+    glTexParameteriv = (PFNGLTEXPARAMETERIVPROC) SMOL_GETGLPROC(glTexParameteriv); 
+    glTexImage1D = (PFNGLTEXIMAGE1DPROC) SMOL_GETGLPROC(glTexImage1D); 
+    glTexImage2D = (PFNGLTEXIMAGE2DPROC) SMOL_GETGLPROC(glTexImage2D); 
+    glDrawBuffer = (PFNGLDRAWBUFFERPROC) SMOL_GETGLPROC(glDrawBuffer); 
+    glClear = (PFNGLCLEARPROC) SMOL_GETGLPROC(glClear); 
+    glClearColor = (PFNGLCLEARCOLORPROC) SMOL_GETGLPROC(glClearColor); 
+    glClearStencil = (PFNGLCLEARSTENCILPROC) SMOL_GETGLPROC(glClearStencil); 
+    glClearDepth = (PFNGLCLEARDEPTHPROC) SMOL_GETGLPROC(glClearDepth); 
+    glStencilMask = (PFNGLSTENCILMASKPROC) SMOL_GETGLPROC(glStencilMask); 
+    glColorMask = (PFNGLCOLORMASKPROC) SMOL_GETGLPROC(glColorMask); 
+    glDepthMask = (PFNGLDEPTHMASKPROC) SMOL_GETGLPROC(glDepthMask); 
+    glDisable = (PFNGLDISABLEPROC) SMOL_GETGLPROC(glDisable); 
+    glEnable = (PFNGLENABLEPROC) SMOL_GETGLPROC(glEnable); 
+    glFinish = (PFNGLFINISHPROC) SMOL_GETGLPROC(glFinish); 
+    glFlush = (PFNGLFLUSHPROC) SMOL_GETGLPROC(glFlush); 
+    glBlendFunc = (PFNGLBLENDFUNCPROC) SMOL_GETGLPROC(glBlendFunc); 
+    glLogicOp = (PFNGLLOGICOPPROC) SMOL_GETGLPROC(glLogicOp); 
+    glStencilFunc = (PFNGLSTENCILFUNCPROC) SMOL_GETGLPROC(glStencilFunc); 
+    glStencilOp = (PFNGLSTENCILOPPROC) SMOL_GETGLPROC(glStencilOp); 
+    glDepthFunc = (PFNGLDEPTHFUNCPROC) SMOL_GETGLPROC(glDepthFunc); 
+    glPixelStoref = (PFNGLPIXELSTOREFPROC) SMOL_GETGLPROC(glPixelStoref); 
+    glPixelStorei = (PFNGLPIXELSTOREIPROC) SMOL_GETGLPROC(glPixelStorei); 
+    glReadBuffer = (PFNGLREADBUFFERPROC) SMOL_GETGLPROC(glReadBuffer); 
+    glReadPixels = (PFNGLREADPIXELSPROC) SMOL_GETGLPROC(glReadPixels); 
+    glGetBooleanv = (PFNGLGETBOOLEANVPROC) SMOL_GETGLPROC(glGetBooleanv); 
+    glGetDoublev = (PFNGLGETDOUBLEVPROC) SMOL_GETGLPROC(glGetDoublev); 
+    glGetError = (PFNGLGETERRORPROC) SMOL_GETGLPROC(glGetError); 
+    glGetFloatv = (PFNGLGETFLOATVPROC) SMOL_GETGLPROC(glGetFloatv); 
+    glGetIntegerv = (PFNGLGETINTEGERVPROC) SMOL_GETGLPROC(glGetIntegerv); 
+    glGetString = (PFNGLGETSTRINGPROC) SMOL_GETGLPROC(glGetString);
+    glGetTexImage = (PFNGLGETTEXIMAGEPROC) SMOL_GETGLPROC(glGetTexImage); 
+    glGetTexParameterfv = (PFNGLGETTEXPARAMETERFVPROC) SMOL_GETGLPROC(glGetTexParameterfv); 
+    glGetTexParameteriv = (PFNGLGETTEXPARAMETERIVPROC) SMOL_GETGLPROC(glGetTexParameteriv); 
+    glGetTexLevelParameterfv = (PFNGLGETTEXLEVELPARAMETERFVPROC) SMOL_GETGLPROC(glGetTexLevelParameterfv); 
+    glGetTexLevelParameteriv = (PFNGLGETTEXLEVELPARAMETERIVPROC) SMOL_GETGLPROC(glGetTexLevelParameteriv); 
+    glIsEnabled = (PFNGLISENABLEDPROC) SMOL_GETGLPROC(glIsEnabled); 
+    glDepthRange = (PFNGLDEPTHRANGEPROC) SMOL_GETGLPROC(glDepthRange); 
+    glViewport = (PFNGLVIEWPORTPROC) SMOL_GETGLPROC(glViewport); 
+    glDrawArrays = (PFNGLDRAWARRAYSPROC) SMOL_GETGLPROC(glDrawArrays); 
+    glDrawElements = (PFNGLDRAWELEMENTSPROC) SMOL_GETGLPROC(glDrawElements); 
+    glGetPointerv = (PFNGLGETPOINTERVPROC) SMOL_GETGLPROC(glGetPointerv); 
+    glPolygonOffset = (PFNGLPOLYGONOFFSETPROC) SMOL_GETGLPROC(glPolygonOffset); 
+    glCopyTexImage1D = (PFNGLCOPYTEXIMAGE1DPROC) SMOL_GETGLPROC(glCopyTexImage1D); 
+    glCopyTexImage2D = (PFNGLCOPYTEXIMAGE2DPROC) SMOL_GETGLPROC(glCopyTexImage2D); 
+    glCopyTexSubImage1D = (PFNGLCOPYTEXSUBIMAGE1DPROC) SMOL_GETGLPROC(glCopyTexSubImage1D); 
+    glCopyTexSubImage2D = (PFNGLCOPYTEXSUBIMAGE2DPROC) SMOL_GETGLPROC(glCopyTexSubImage2D); 
+    glTexSubImage1D = (PFNGLTEXSUBIMAGE1DPROC) SMOL_GETGLPROC(glTexSubImage1D); 
+    glTexSubImage2D = (PFNGLTEXSUBIMAGE2DPROC) SMOL_GETGLPROC(glTexSubImage2D); 
+    glBindTexture = (PFNGLBINDTEXTUREPROC) SMOL_GETGLPROC(glBindTexture); 
+    glDeleteTextures = (PFNGLDELETETEXTURESPROC) SMOL_GETGLPROC(glDeleteTextures); 
+    glGenTextures = (PFNGLGENTEXTURESPROC) SMOL_GETGLPROC(glGenTextures); 
+    glIsTexture = (PFNGLISTEXTUREPROC) SMOL_GETGLPROC(glIsTexture); 
+    glDrawRangeElements = (PFNGLDRAWRANGEELEMENTSPROC) SMOL_GETGLPROC(glDrawRangeElements); 
+    glTexImage3D = (PFNGLTEXIMAGE3DPROC) SMOL_GETGLPROC(glTexImage3D); 
+    glTexSubImage3D = (PFNGLTEXSUBIMAGE3DPROC) SMOL_GETGLPROC(glTexSubImage3D); 
+    glCopyTexSubImage3D = (PFNGLCOPYTEXSUBIMAGE3DPROC) SMOL_GETGLPROC(glCopyTexSubImage3D); 
+    glActiveTexture = (PFNGLACTIVETEXTUREPROC) SMOL_GETGLPROC(glActiveTexture); 
+    glSampleCoverage = (PFNGLSAMPLECOVERAGEPROC) SMOL_GETGLPROC(glSampleCoverage); 
+    glCompressedTexImage3D = (PFNGLCOMPRESSEDTEXIMAGE3DPROC) SMOL_GETGLPROC(glCompressedTexImage3D); 
+    glCompressedTexImage2D = (PFNGLCOMPRESSEDTEXIMAGE2DPROC) SMOL_GETGLPROC(glCompressedTexImage2D); 
+    glCompressedTexImage1D = (PFNGLCOMPRESSEDTEXIMAGE1DPROC) SMOL_GETGLPROC(glCompressedTexImage1D); 
+    glCompressedTexSubImage3D = (PFNGLCOMPRESSEDTEXSUBIMAGE3DPROC) SMOL_GETGLPROC(glCompressedTexSubImage3D); 
+    glCompressedTexSubImage2D = (PFNGLCOMPRESSEDTEXSUBIMAGE2DPROC) SMOL_GETGLPROC(glCompressedTexSubImage2D); 
+    glCompressedTexSubImage1D = (PFNGLCOMPRESSEDTEXSUBIMAGE1DPROC) SMOL_GETGLPROC(glCompressedTexSubImage1D); 
+    glGetCompressedTexImage = (PFNGLGETCOMPRESSEDTEXIMAGEPROC) SMOL_GETGLPROC(glGetCompressedTexImage); 
+    glBlendFuncSeparate = (PFNGLBLENDFUNCSEPARATEPROC) SMOL_GETGLPROC(glBlendFuncSeparate); 
+    glMultiDrawArrays = (PFNGLMULTIDRAWARRAYSPROC) SMOL_GETGLPROC(glMultiDrawArrays); 
+    glMultiDrawElements = (PFNGLMULTIDRAWELEMENTSPROC) SMOL_GETGLPROC(glMultiDrawElements); 
+    glPointParameterf = (PFNGLPOINTPARAMETERFPROC) SMOL_GETGLPROC(glPointParameterf); 
+    glPointParameterfv = (PFNGLPOINTPARAMETERFVPROC) SMOL_GETGLPROC(glPointParameterfv); 
+    glPointParameteri = (PFNGLPOINTPARAMETERIPROC) SMOL_GETGLPROC(glPointParameteri); 
+    glPointParameteriv = (PFNGLPOINTPARAMETERIVPROC) SMOL_GETGLPROC(glPointParameteriv); 
+    glBlendColor = (PFNGLBLENDCOLORPROC) SMOL_GETGLPROC(glBlendColor); 
+    glBlendEquation = (PFNGLBLENDEQUATIONPROC) SMOL_GETGLPROC(glBlendEquation); 
+    glGenQueries = (PFNGLGENQUERIESPROC) SMOL_GETGLPROC(glGenQueries); 
+    glDeleteQueries = (PFNGLDELETEQUERIESPROC) SMOL_GETGLPROC(glDeleteQueries); 
+    glIsQuery = (PFNGLISQUERYPROC) SMOL_GETGLPROC(glIsQuery); 
+    glBeginQuery = (PFNGLBEGINQUERYPROC) SMOL_GETGLPROC(glBeginQuery); 
+    glEndQuery = (PFNGLENDQUERYPROC) SMOL_GETGLPROC(glEndQuery); 
+    glGetQueryiv = (PFNGLGETQUERYIVPROC) SMOL_GETGLPROC(glGetQueryiv); 
+    glGetQueryObjectiv = (PFNGLGETQUERYOBJECTIVPROC) SMOL_GETGLPROC(glGetQueryObjectiv); 
+    glGetQueryObjectuiv = (PFNGLGETQUERYOBJECTUIVPROC) SMOL_GETGLPROC(glGetQueryObjectuiv); 
+    glBindBuffer = (PFNGLBINDBUFFERPROC) SMOL_GETGLPROC(glBindBuffer); 
+    glDeleteBuffers = (PFNGLDELETEBUFFERSPROC) SMOL_GETGLPROC(glDeleteBuffers); 
+    glGenBuffers = (PFNGLGENBUFFERSPROC) SMOL_GETGLPROC(glGenBuffers); 
+    glIsBuffer = (PFNGLISBUFFERPROC) SMOL_GETGLPROC(glIsBuffer); 
+    glBufferData = (PFNGLBUFFERDATAPROC) SMOL_GETGLPROC(glBufferData); 
+    glBufferSubData = (PFNGLBUFFERSUBDATAPROC) SMOL_GETGLPROC(glBufferSubData); 
+    glGetBufferSubData = (PFNGLGETBUFFERSUBDATAPROC) SMOL_GETGLPROC(glGetBufferSubData); 
+    glMapBuffer = (PFNGLMAPBUFFERPROC) SMOL_GETGLPROC(glMapBuffer); 
+    glUnmapBuffer = (PFNGLUNMAPBUFFERPROC) SMOL_GETGLPROC(glUnmapBuffer); 
+    glGetBufferParameteriv = (PFNGLGETBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetBufferParameteriv); 
+    glGetBufferPointerv = (PFNGLGETBUFFERPOINTERVPROC) SMOL_GETGLPROC(glGetBufferPointerv); 
+    glBlendEquationSeparate = (PFNGLBLENDEQUATIONSEPARATEPROC) SMOL_GETGLPROC(glBlendEquationSeparate); 
+    glDrawBuffers = (PFNGLDRAWBUFFERSPROC) SMOL_GETGLPROC(glDrawBuffers); 
+    glStencilOpSeparate = (PFNGLSTENCILOPSEPARATEPROC) SMOL_GETGLPROC(glStencilOpSeparate); 
+    glStencilFuncSeparate = (PFNGLSTENCILFUNCSEPARATEPROC) SMOL_GETGLPROC(glStencilFuncSeparate); 
+    glStencilMaskSeparate = (PFNGLSTENCILMASKSEPARATEPROC) SMOL_GETGLPROC(glStencilMaskSeparate); 
+    glAttachShader = (PFNGLATTACHSHADERPROC) SMOL_GETGLPROC(glAttachShader); 
+    glBindAttribLocation = (PFNGLBINDATTRIBLOCATIONPROC) SMOL_GETGLPROC(glBindAttribLocation); 
+    glCompileShader = (PFNGLCOMPILESHADERPROC) SMOL_GETGLPROC(glCompileShader); 
+    glCreateProgram = (PFNGLCREATEPROGRAMPROC) SMOL_GETGLPROC(glCreateProgram); 
+    glCreateShader = (PFNGLCREATESHADERPROC) SMOL_GETGLPROC(glCreateShader); 
+    glDeleteProgram = (PFNGLDELETEPROGRAMPROC) SMOL_GETGLPROC(glDeleteProgram); 
+    glDeleteShader = (PFNGLDELETESHADERPROC) SMOL_GETGLPROC(glDeleteShader); 
+    glDetachShader = (PFNGLDETACHSHADERPROC) SMOL_GETGLPROC(glDetachShader); 
+    glDisableVertexAttribArray = (PFNGLDISABLEVERTEXATTRIBARRAYPROC) SMOL_GETGLPROC(glDisableVertexAttribArray); 
+    glEnableVertexAttribArray = (PFNGLENABLEVERTEXATTRIBARRAYPROC) SMOL_GETGLPROC(glEnableVertexAttribArray); 
+    glGetActiveAttrib = (PFNGLGETACTIVEATTRIBPROC) SMOL_GETGLPROC(glGetActiveAttrib); 
+    glGetActiveUniform = (PFNGLGETACTIVEUNIFORMPROC) SMOL_GETGLPROC(glGetActiveUniform); 
+    glGetAttachedShaders = (PFNGLGETATTACHEDSHADERSPROC) SMOL_GETGLPROC(glGetAttachedShaders); 
+    glGetAttribLocation = (PFNGLGETATTRIBLOCATIONPROC) SMOL_GETGLPROC(glGetAttribLocation); 
+    glGetProgramiv = (PFNGLGETPROGRAMIVPROC) SMOL_GETGLPROC(glGetProgramiv); 
+    glGetProgramInfoLog = (PFNGLGETPROGRAMINFOLOGPROC) SMOL_GETGLPROC(glGetProgramInfoLog); 
+    glGetShaderiv = (PFNGLGETSHADERIVPROC) SMOL_GETGLPROC(glGetShaderiv); 
+    glGetShaderInfoLog = (PFNGLGETSHADERINFOLOGPROC) SMOL_GETGLPROC(glGetShaderInfoLog); 
+    glGetShaderSource = (PFNGLGETSHADERSOURCEPROC) SMOL_GETGLPROC(glGetShaderSource); 
+    glGetUniformLocation = (PFNGLGETUNIFORMLOCATIONPROC) SMOL_GETGLPROC(glGetUniformLocation); 
+    glGetUniformfv = (PFNGLGETUNIFORMFVPROC) SMOL_GETGLPROC(glGetUniformfv); 
+    glGetUniformiv = (PFNGLGETUNIFORMIVPROC) SMOL_GETGLPROC(glGetUniformiv); 
+    glGetVertexAttribdv = (PFNGLGETVERTEXATTRIBDVPROC) SMOL_GETGLPROC(glGetVertexAttribdv); 
+    glGetVertexAttribfv = (PFNGLGETVERTEXATTRIBFVPROC) SMOL_GETGLPROC(glGetVertexAttribfv); 
+    glGetVertexAttribiv = (PFNGLGETVERTEXATTRIBIVPROC) SMOL_GETGLPROC(glGetVertexAttribiv); 
+    glGetVertexAttribPointerv = (PFNGLGETVERTEXATTRIBPOINTERVPROC) SMOL_GETGLPROC(glGetVertexAttribPointerv); 
+    glIsProgram = (PFNGLISPROGRAMPROC) SMOL_GETGLPROC(glIsProgram); 
+    glIsShader = (PFNGLISSHADERPROC) SMOL_GETGLPROC(glIsShader); 
+    glLinkProgram = (PFNGLLINKPROGRAMPROC) SMOL_GETGLPROC(glLinkProgram); 
+    glShaderSource = (PFNGLSHADERSOURCEPROC) SMOL_GETGLPROC(glShaderSource); 
+    glUseProgram = (PFNGLUSEPROGRAMPROC) SMOL_GETGLPROC(glUseProgram); 
+    glUniform1f = (PFNGLUNIFORM1FPROC) SMOL_GETGLPROC(glUniform1f); 
+    glUniform2f = (PFNGLUNIFORM2FPROC) SMOL_GETGLPROC(glUniform2f); 
+    glUniform3f = (PFNGLUNIFORM3FPROC) SMOL_GETGLPROC(glUniform3f); 
+    glUniform4f = (PFNGLUNIFORM4FPROC) SMOL_GETGLPROC(glUniform4f); 
+    glUniform1i = (PFNGLUNIFORM1IPROC) SMOL_GETGLPROC(glUniform1i); 
+    glUniform2i = (PFNGLUNIFORM2IPROC) SMOL_GETGLPROC(glUniform2i); 
+    glUniform3i = (PFNGLUNIFORM3IPROC) SMOL_GETGLPROC(glUniform3i); 
+    glUniform4i = (PFNGLUNIFORM4IPROC) SMOL_GETGLPROC(glUniform4i); 
+    glUniform1fv = (PFNGLUNIFORM1FVPROC) SMOL_GETGLPROC(glUniform1fv); 
+    glUniform2fv = (PFNGLUNIFORM2FVPROC) SMOL_GETGLPROC(glUniform2fv); 
+    glUniform3fv = (PFNGLUNIFORM3FVPROC) SMOL_GETGLPROC(glUniform3fv); 
+    glUniform4fv = (PFNGLUNIFORM4FVPROC) SMOL_GETGLPROC(glUniform4fv); 
+    glUniform1iv = (PFNGLUNIFORM1IVPROC) SMOL_GETGLPROC(glUniform1iv); 
+    glUniform2iv = (PFNGLUNIFORM2IVPROC) SMOL_GETGLPROC(glUniform2iv); 
+    glUniform3iv = (PFNGLUNIFORM3IVPROC) SMOL_GETGLPROC(glUniform3iv); 
+    glUniform4iv = (PFNGLUNIFORM4IVPROC) SMOL_GETGLPROC(glUniform4iv); 
+    glUniformMatrix2fv = (PFNGLUNIFORMMATRIX2FVPROC) SMOL_GETGLPROC(glUniformMatrix2fv); 
+    glUniformMatrix3fv = (PFNGLUNIFORMMATRIX3FVPROC) SMOL_GETGLPROC(glUniformMatrix3fv); 
+    glUniformMatrix4fv = (PFNGLUNIFORMMATRIX4FVPROC) SMOL_GETGLPROC(glUniformMatrix4fv); 
+    glValidateProgram = (PFNGLVALIDATEPROGRAMPROC) SMOL_GETGLPROC(glValidateProgram); 
+    glVertexAttrib1d = (PFNGLVERTEXATTRIB1DPROC) SMOL_GETGLPROC(glVertexAttrib1d); 
+    glVertexAttrib1dv = (PFNGLVERTEXATTRIB1DVPROC) SMOL_GETGLPROC(glVertexAttrib1dv); 
+    glVertexAttrib1f = (PFNGLVERTEXATTRIB1FPROC) SMOL_GETGLPROC(glVertexAttrib1f); 
+    glVertexAttrib1fv = (PFNGLVERTEXATTRIB1FVPROC) SMOL_GETGLPROC(glVertexAttrib1fv); 
+    glVertexAttrib1s = (PFNGLVERTEXATTRIB1SPROC) SMOL_GETGLPROC(glVertexAttrib1s); 
+    glVertexAttrib1sv = (PFNGLVERTEXATTRIB1SVPROC) SMOL_GETGLPROC(glVertexAttrib1sv); 
+    glVertexAttrib2d = (PFNGLVERTEXATTRIB2DPROC) SMOL_GETGLPROC(glVertexAttrib2d); 
+    glVertexAttrib2dv = (PFNGLVERTEXATTRIB2DVPROC) SMOL_GETGLPROC(glVertexAttrib2dv); 
+    glVertexAttrib2f = (PFNGLVERTEXATTRIB2FPROC) SMOL_GETGLPROC(glVertexAttrib2f); 
+    glVertexAttrib2fv = (PFNGLVERTEXATTRIB2FVPROC) SMOL_GETGLPROC(glVertexAttrib2fv); 
+    glVertexAttrib2s = (PFNGLVERTEXATTRIB2SPROC) SMOL_GETGLPROC(glVertexAttrib2s); 
+    glVertexAttrib2sv = (PFNGLVERTEXATTRIB2SVPROC) SMOL_GETGLPROC(glVertexAttrib2sv); 
+    glVertexAttrib3d = (PFNGLVERTEXATTRIB3DPROC) SMOL_GETGLPROC(glVertexAttrib3d); 
+    glVertexAttrib3dv = (PFNGLVERTEXATTRIB3DVPROC) SMOL_GETGLPROC(glVertexAttrib3dv); 
+    glVertexAttrib3f = (PFNGLVERTEXATTRIB3FPROC) SMOL_GETGLPROC(glVertexAttrib3f); 
+    glVertexAttrib3fv = (PFNGLVERTEXATTRIB3FVPROC) SMOL_GETGLPROC(glVertexAttrib3fv); 
+    glVertexAttrib3s = (PFNGLVERTEXATTRIB3SPROC) SMOL_GETGLPROC(glVertexAttrib3s); 
+    glVertexAttrib3sv = (PFNGLVERTEXATTRIB3SVPROC) SMOL_GETGLPROC(glVertexAttrib3sv); 
+    glVertexAttrib4Nbv = (PFNGLVERTEXATTRIB4NBVPROC) SMOL_GETGLPROC(glVertexAttrib4Nbv); 
+    glVertexAttrib4Niv = (PFNGLVERTEXATTRIB4NIVPROC) SMOL_GETGLPROC(glVertexAttrib4Niv); 
+    glVertexAttrib4Nsv = (PFNGLVERTEXATTRIB4NSVPROC) SMOL_GETGLPROC(glVertexAttrib4Nsv); 
+    glVertexAttrib4Nub = (PFNGLVERTEXATTRIB4NUBPROC) SMOL_GETGLPROC(glVertexAttrib4Nub); 
+    glVertexAttrib4Nubv = (PFNGLVERTEXATTRIB4NUBVPROC) SMOL_GETGLPROC(glVertexAttrib4Nubv); 
+    glVertexAttrib4Nuiv = (PFNGLVERTEXATTRIB4NUIVPROC) SMOL_GETGLPROC(glVertexAttrib4Nuiv); 
+    glVertexAttrib4Nusv = (PFNGLVERTEXATTRIB4NUSVPROC) SMOL_GETGLPROC(glVertexAttrib4Nusv); 
+    glVertexAttrib4bv = (PFNGLVERTEXATTRIB4BVPROC) SMOL_GETGLPROC(glVertexAttrib4bv); 
+    glVertexAttrib4d = (PFNGLVERTEXATTRIB4DPROC) SMOL_GETGLPROC(glVertexAttrib4d); 
+    glVertexAttrib4dv = (PFNGLVERTEXATTRIB4DVPROC) SMOL_GETGLPROC(glVertexAttrib4dv); 
+    glVertexAttrib4f = (PFNGLVERTEXATTRIB4FPROC) SMOL_GETGLPROC(glVertexAttrib4f); 
+    glVertexAttrib4fv = (PFNGLVERTEXATTRIB4FVPROC) SMOL_GETGLPROC(glVertexAttrib4fv); 
+    glVertexAttrib4iv = (PFNGLVERTEXATTRIB4IVPROC) SMOL_GETGLPROC(glVertexAttrib4iv); 
+    glVertexAttrib4s = (PFNGLVERTEXATTRIB4SPROC) SMOL_GETGLPROC(glVertexAttrib4s); 
+    glVertexAttrib4sv = (PFNGLVERTEXATTRIB4SVPROC) SMOL_GETGLPROC(glVertexAttrib4sv); 
+    glVertexAttrib4ubv = (PFNGLVERTEXATTRIB4UBVPROC) SMOL_GETGLPROC(glVertexAttrib4ubv); 
+    glVertexAttrib4uiv = (PFNGLVERTEXATTRIB4UIVPROC) SMOL_GETGLPROC(glVertexAttrib4uiv); 
+    glVertexAttrib4usv = (PFNGLVERTEXATTRIB4USVPROC) SMOL_GETGLPROC(glVertexAttrib4usv); 
+    glVertexAttribPointer = (PFNGLVERTEXATTRIBPOINTERPROC) SMOL_GETGLPROC(glVertexAttribPointer); 
+    glUniformMatrix2x3fv = (PFNGLUNIFORMMATRIX2X3FVPROC) SMOL_GETGLPROC(glUniformMatrix2x3fv); 
+    glUniformMatrix3x2fv = (PFNGLUNIFORMMATRIX3X2FVPROC) SMOL_GETGLPROC(glUniformMatrix3x2fv); 
+    glUniformMatrix2x4fv = (PFNGLUNIFORMMATRIX2X4FVPROC) SMOL_GETGLPROC(glUniformMatrix2x4fv); 
+    glUniformMatrix4x2fv = (PFNGLUNIFORMMATRIX4X2FVPROC) SMOL_GETGLPROC(glUniformMatrix4x2fv); 
+    glUniformMatrix3x4fv = (PFNGLUNIFORMMATRIX3X4FVPROC) SMOL_GETGLPROC(glUniformMatrix3x4fv); 
+    glUniformMatrix4x3fv = (PFNGLUNIFORMMATRIX4X3FVPROC) SMOL_GETGLPROC(glUniformMatrix4x3fv); 
+    glColorMaski = (PFNGLCOLORMASKIPROC) SMOL_GETGLPROC(glColorMaski); 
+    glGetBooleani_v = (PFNGLGETBOOLEANI_VPROC) SMOL_GETGLPROC(glGetBooleani_v); 
+    glGetIntegeri_v = (PFNGLGETINTEGERI_VPROC) SMOL_GETGLPROC(glGetIntegeri_v); 
+    glEnablei = (PFNGLENABLEIPROC) SMOL_GETGLPROC(glEnablei); 
+    glDisablei = (PFNGLDISABLEIPROC) SMOL_GETGLPROC(glDisablei); 
+    glIsEnabledi = (PFNGLISENABLEDIPROC) SMOL_GETGLPROC(glIsEnabledi); 
+    glBeginTransformFeedback = (PFNGLBEGINTRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glBeginTransformFeedback); 
+    glEndTransformFeedback = (PFNGLENDTRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glEndTransformFeedback); 
+    glBindBufferRange = (PFNGLBINDBUFFERRANGEPROC) SMOL_GETGLPROC(glBindBufferRange); 
+    glBindBufferBase = (PFNGLBINDBUFFERBASEPROC) SMOL_GETGLPROC(glBindBufferBase); 
+    glTransformFeedbackVaryings = (PFNGLTRANSFORMFEEDBACKVARYINGSPROC) SMOL_GETGLPROC(glTransformFeedbackVaryings); 
+    glGetTransformFeedbackVarying = (PFNGLGETTRANSFORMFEEDBACKVARYINGPROC) SMOL_GETGLPROC(glGetTransformFeedbackVarying); 
+    glClampColor = (PFNGLCLAMPCOLORPROC) SMOL_GETGLPROC(glClampColor); 
+    glBeginConditionalRender = (PFNGLBEGINCONDITIONALRENDERPROC) SMOL_GETGLPROC(glBeginConditionalRender); 
+    glEndConditionalRender = (PFNGLENDCONDITIONALRENDERPROC) SMOL_GETGLPROC(glEndConditionalRender); 
+    glVertexAttribIPointer = (PFNGLVERTEXATTRIBIPOINTERPROC) SMOL_GETGLPROC(glVertexAttribIPointer); 
+    glGetVertexAttribIiv = (PFNGLGETVERTEXATTRIBIIVPROC) SMOL_GETGLPROC(glGetVertexAttribIiv); 
+    glGetVertexAttribIuiv = (PFNGLGETVERTEXATTRIBIUIVPROC) SMOL_GETGLPROC(glGetVertexAttribIuiv); 
+    glVertexAttribI1i = (PFNGLVERTEXATTRIBI1IPROC) SMOL_GETGLPROC(glVertexAttribI1i); 
+    glVertexAttribI2i = (PFNGLVERTEXATTRIBI2IPROC) SMOL_GETGLPROC(glVertexAttribI2i); 
+    glVertexAttribI3i = (PFNGLVERTEXATTRIBI3IPROC) SMOL_GETGLPROC(glVertexAttribI3i); 
+    glVertexAttribI4i = (PFNGLVERTEXATTRIBI4IPROC) SMOL_GETGLPROC(glVertexAttribI4i); 
+    glVertexAttribI1ui = (PFNGLVERTEXATTRIBI1UIPROC) SMOL_GETGLPROC(glVertexAttribI1ui); 
+    glVertexAttribI2ui = (PFNGLVERTEXATTRIBI2UIPROC) SMOL_GETGLPROC(glVertexAttribI2ui); 
+    glVertexAttribI3ui = (PFNGLVERTEXATTRIBI3UIPROC) SMOL_GETGLPROC(glVertexAttribI3ui); 
+    glVertexAttribI4ui = (PFNGLVERTEXATTRIBI4UIPROC) SMOL_GETGLPROC(glVertexAttribI4ui); 
+    glVertexAttribI1iv = (PFNGLVERTEXATTRIBI1IVPROC) SMOL_GETGLPROC(glVertexAttribI1iv); 
+    glVertexAttribI2iv = (PFNGLVERTEXATTRIBI2IVPROC) SMOL_GETGLPROC(glVertexAttribI2iv); 
+    glVertexAttribI3iv = (PFNGLVERTEXATTRIBI3IVPROC) SMOL_GETGLPROC(glVertexAttribI3iv); 
+    glVertexAttribI4iv = (PFNGLVERTEXATTRIBI4IVPROC) SMOL_GETGLPROC(glVertexAttribI4iv); 
+    glVertexAttribI1uiv = (PFNGLVERTEXATTRIBI1UIVPROC) SMOL_GETGLPROC(glVertexAttribI1uiv); 
+    glVertexAttribI2uiv = (PFNGLVERTEXATTRIBI2UIVPROC) SMOL_GETGLPROC(glVertexAttribI2uiv); 
+    glVertexAttribI3uiv = (PFNGLVERTEXATTRIBI3UIVPROC) SMOL_GETGLPROC(glVertexAttribI3uiv); 
+    glVertexAttribI4uiv = (PFNGLVERTEXATTRIBI4UIVPROC) SMOL_GETGLPROC(glVertexAttribI4uiv); 
+    glVertexAttribI4bv = (PFNGLVERTEXATTRIBI4BVPROC) SMOL_GETGLPROC(glVertexAttribI4bv); 
+    glVertexAttribI4sv = (PFNGLVERTEXATTRIBI4SVPROC) SMOL_GETGLPROC(glVertexAttribI4sv); 
+    glVertexAttribI4ubv = (PFNGLVERTEXATTRIBI4UBVPROC) SMOL_GETGLPROC(glVertexAttribI4ubv); 
+    glVertexAttribI4usv = (PFNGLVERTEXATTRIBI4USVPROC) SMOL_GETGLPROC(glVertexAttribI4usv); 
+    glGetUniformuiv = (PFNGLGETUNIFORMUIVPROC) SMOL_GETGLPROC(glGetUniformuiv); 
+    glBindFragDataLocation = (PFNGLBINDFRAGDATALOCATIONPROC) SMOL_GETGLPROC(glBindFragDataLocation); 
+    glGetFragDataLocation = (PFNGLGETFRAGDATALOCATIONPROC) SMOL_GETGLPROC(glGetFragDataLocation); 
+    glUniform1ui = (PFNGLUNIFORM1UIPROC) SMOL_GETGLPROC(glUniform1ui); 
+    glUniform2ui = (PFNGLUNIFORM2UIPROC) SMOL_GETGLPROC(glUniform2ui); 
+    glUniform3ui = (PFNGLUNIFORM3UIPROC) SMOL_GETGLPROC(glUniform3ui); 
+    glUniform4ui = (PFNGLUNIFORM4UIPROC) SMOL_GETGLPROC(glUniform4ui); 
+    glUniform1uiv = (PFNGLUNIFORM1UIVPROC) SMOL_GETGLPROC(glUniform1uiv); 
+    glUniform2uiv = (PFNGLUNIFORM2UIVPROC) SMOL_GETGLPROC(glUniform2uiv); 
+    glUniform3uiv = (PFNGLUNIFORM3UIVPROC) SMOL_GETGLPROC(glUniform3uiv); 
+    glUniform4uiv = (PFNGLUNIFORM4UIVPROC) SMOL_GETGLPROC(glUniform4uiv); 
+    glTexParameterIiv = (PFNGLTEXPARAMETERIIVPROC) SMOL_GETGLPROC(glTexParameterIiv); 
+    glTexParameterIuiv = (PFNGLTEXPARAMETERIUIVPROC) SMOL_GETGLPROC(glTexParameterIuiv); 
+    glGetTexParameterIiv = (PFNGLGETTEXPARAMETERIIVPROC) SMOL_GETGLPROC(glGetTexParameterIiv); 
+    glGetTexParameterIuiv = (PFNGLGETTEXPARAMETERIUIVPROC) SMOL_GETGLPROC(glGetTexParameterIuiv); 
+    glClearBufferiv = (PFNGLCLEARBUFFERIVPROC) SMOL_GETGLPROC(glClearBufferiv); 
+    glClearBufferuiv = (PFNGLCLEARBUFFERUIVPROC) SMOL_GETGLPROC(glClearBufferuiv); 
+    glClearBufferfv = (PFNGLCLEARBUFFERFVPROC) SMOL_GETGLPROC(glClearBufferfv); 
+    glClearBufferfi = (PFNGLCLEARBUFFERFIPROC) SMOL_GETGLPROC(glClearBufferfi); 
+    glGetStringi = (PFNGLGETSTRINGIPROC) SMOL_GETGLPROC(glGetStringi);
+    glIsRenderbuffer = (PFNGLISRENDERBUFFERPROC) SMOL_GETGLPROC(glIsRenderbuffer); 
+    glBindRenderbuffer = (PFNGLBINDRENDERBUFFERPROC) SMOL_GETGLPROC(glBindRenderbuffer); 
+    glDeleteRenderbuffers = (PFNGLDELETERENDERBUFFERSPROC) SMOL_GETGLPROC(glDeleteRenderbuffers); 
+    glGenRenderbuffers = (PFNGLGENRENDERBUFFERSPROC) SMOL_GETGLPROC(glGenRenderbuffers); 
+    glRenderbufferStorage = (PFNGLRENDERBUFFERSTORAGEPROC) SMOL_GETGLPROC(glRenderbufferStorage); 
+    glGetRenderbufferParameteriv = (PFNGLGETRENDERBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetRenderbufferParameteriv); 
+    glIsFramebuffer = (PFNGLISFRAMEBUFFERPROC) SMOL_GETGLPROC(glIsFramebuffer); 
+    glBindFramebuffer = (PFNGLBINDFRAMEBUFFERPROC) SMOL_GETGLPROC(glBindFramebuffer); 
+    glDeleteFramebuffers = (PFNGLDELETEFRAMEBUFFERSPROC) SMOL_GETGLPROC(glDeleteFramebuffers); 
+    glGenFramebuffers = (PFNGLGENFRAMEBUFFERSPROC) SMOL_GETGLPROC(glGenFramebuffers); 
+    glCheckFramebufferStatus = (PFNGLCHECKFRAMEBUFFERSTATUSPROC) SMOL_GETGLPROC(glCheckFramebufferStatus); 
+    glFramebufferTexture1D = (PFNGLFRAMEBUFFERTEXTURE1DPROC) SMOL_GETGLPROC(glFramebufferTexture1D); 
+    glFramebufferTexture2D = (PFNGLFRAMEBUFFERTEXTURE2DPROC) SMOL_GETGLPROC(glFramebufferTexture2D); 
+    glFramebufferTexture3D = (PFNGLFRAMEBUFFERTEXTURE3DPROC) SMOL_GETGLPROC(glFramebufferTexture3D); 
+    glFramebufferRenderbuffer = (PFNGLFRAMEBUFFERRENDERBUFFERPROC) SMOL_GETGLPROC(glFramebufferRenderbuffer); 
+    glGetFramebufferAttachmentParameteriv = (PFNGLGETFRAMEBUFFERATTACHMENTPARAMETERIVPROC) SMOL_GETGLPROC(glGetFramebufferAttachmentParameteriv); 
+    glGenerateMipmap = (PFNGLGENERATEMIPMAPPROC) SMOL_GETGLPROC(glGenerateMipmap); 
+    glBlitFramebuffer = (PFNGLBLITFRAMEBUFFERPROC) SMOL_GETGLPROC(glBlitFramebuffer); 
+    glRenderbufferStorageMultisample = (PFNGLRENDERBUFFERSTORAGEMULTISAMPLEPROC) SMOL_GETGLPROC(glRenderbufferStorageMultisample); 
+    glFramebufferTextureLayer = (PFNGLFRAMEBUFFERTEXTURELAYERPROC) SMOL_GETGLPROC(glFramebufferTextureLayer); 
+    glMapBufferRange = (PFNGLMAPBUFFERRANGEPROC) SMOL_GETGLPROC(glMapBufferRange); 
+    glFlushMappedBufferRange = (PFNGLFLUSHMAPPEDBUFFERRANGEPROC) SMOL_GETGLPROC(glFlushMappedBufferRange); 
+    glBindVertexArray = (PFNGLBINDVERTEXARRAYPROC) SMOL_GETGLPROC(glBindVertexArray); 
+    glDeleteVertexArrays = (PFNGLDELETEVERTEXARRAYSPROC) SMOL_GETGLPROC(glDeleteVertexArrays); 
+    glGenVertexArrays = (PFNGLGENVERTEXARRAYSPROC) SMOL_GETGLPROC(glGenVertexArrays); 
+    glIsVertexArray = (PFNGLISVERTEXARRAYPROC) SMOL_GETGLPROC(glIsVertexArray); 
+    glDrawArraysInstanced = (PFNGLDRAWARRAYSINSTANCEDPROC) SMOL_GETGLPROC(glDrawArraysInstanced); 
+    glDrawElementsInstanced = (PFNGLDRAWELEMENTSINSTANCEDPROC) SMOL_GETGLPROC(glDrawElementsInstanced); 
+    glTexBuffer = (PFNGLTEXBUFFERPROC) SMOL_GETGLPROC(glTexBuffer); 
+    glPrimitiveRestartIndex = (PFNGLPRIMITIVERESTARTINDEXPROC) SMOL_GETGLPROC(glPrimitiveRestartIndex); 
+    glCopyBufferSubData = (PFNGLCOPYBUFFERSUBDATAPROC) SMOL_GETGLPROC(glCopyBufferSubData); 
+    glGetUniformIndices = (PFNGLGETUNIFORMINDICESPROC) SMOL_GETGLPROC(glGetUniformIndices); 
+    glGetActiveUniformsiv = (PFNGLGETACTIVEUNIFORMSIVPROC) SMOL_GETGLPROC(glGetActiveUniformsiv); 
+    glGetActiveUniformName = (PFNGLGETACTIVEUNIFORMNAMEPROC) SMOL_GETGLPROC(glGetActiveUniformName); 
+    glGetUniformBlockIndex = (PFNGLGETUNIFORMBLOCKINDEXPROC) SMOL_GETGLPROC(glGetUniformBlockIndex); 
+    glGetActiveUniformBlockiv = (PFNGLGETACTIVEUNIFORMBLOCKIVPROC) SMOL_GETGLPROC(glGetActiveUniformBlockiv); 
+    glGetActiveUniformBlockName = (PFNGLGETACTIVEUNIFORMBLOCKNAMEPROC) SMOL_GETGLPROC(glGetActiveUniformBlockName); 
+    glUniformBlockBinding = (PFNGLUNIFORMBLOCKBINDINGPROC) SMOL_GETGLPROC(glUniformBlockBinding); 
+    glDrawElementsBaseVertex = (PFNGLDRAWELEMENTSBASEVERTEXPROC) SMOL_GETGLPROC(glDrawElementsBaseVertex); 
+    glDrawRangeElementsBaseVertex = (PFNGLDRAWRANGEELEMENTSBASEVERTEXPROC) SMOL_GETGLPROC(glDrawRangeElementsBaseVertex); 
+    glDrawElementsInstancedBaseVertex = (PFNGLDRAWELEMENTSINSTANCEDBASEVERTEXPROC) SMOL_GETGLPROC(glDrawElementsInstancedBaseVertex); 
+    glMultiDrawElementsBaseVertex = (PFNGLMULTIDRAWELEMENTSBASEVERTEXPROC) SMOL_GETGLPROC(glMultiDrawElementsBaseVertex); 
+    glProvokingVertex = (PFNGLPROVOKINGVERTEXPROC) SMOL_GETGLPROC(glProvokingVertex); 
+    glFenceSync = (PFNGLFENCESYNCPROC) SMOL_GETGLPROC(glFenceSync); 
+    glIsSync = (PFNGLISSYNCPROC) SMOL_GETGLPROC(glIsSync); 
+    glDeleteSync = (PFNGLDELETESYNCPROC) SMOL_GETGLPROC(glDeleteSync); 
+    glClientWaitSync = (PFNGLCLIENTWAITSYNCPROC) SMOL_GETGLPROC(glClientWaitSync); 
+    glWaitSync = (PFNGLWAITSYNCPROC) SMOL_GETGLPROC(glWaitSync); 
+    glGetInteger64v = (PFNGLGETINTEGER64VPROC) SMOL_GETGLPROC(glGetInteger64v); 
+    glGetSynciv = (PFNGLGETSYNCIVPROC) SMOL_GETGLPROC(glGetSynciv); 
+    glGetInteger64i_v = (PFNGLGETINTEGER64I_VPROC) SMOL_GETGLPROC(glGetInteger64i_v); 
+    glGetBufferParameteri64v = (PFNGLGETBUFFERPARAMETERI64VPROC) SMOL_GETGLPROC(glGetBufferParameteri64v); 
+    glFramebufferTexture = (PFNGLFRAMEBUFFERTEXTUREPROC) SMOL_GETGLPROC(glFramebufferTexture); 
+    glTexImage2DMultisample = (PFNGLTEXIMAGE2DMULTISAMPLEPROC) SMOL_GETGLPROC(glTexImage2DMultisample); 
+    glTexImage3DMultisample = (PFNGLTEXIMAGE3DMULTISAMPLEPROC) SMOL_GETGLPROC(glTexImage3DMultisample); 
+    glGetMultisamplefv = (PFNGLGETMULTISAMPLEFVPROC) SMOL_GETGLPROC(glGetMultisamplefv); 
+    glSampleMaski = (PFNGLSAMPLEMASKIPROC) SMOL_GETGLPROC(glSampleMaski); 
+    glBindFragDataLocationIndexed = (PFNGLBINDFRAGDATALOCATIONINDEXEDPROC) SMOL_GETGLPROC(glBindFragDataLocationIndexed); 
+    glGetFragDataIndex = (PFNGLGETFRAGDATAINDEXPROC) SMOL_GETGLPROC(glGetFragDataIndex); 
+    glGenSamplers = (PFNGLGENSAMPLERSPROC) SMOL_GETGLPROC(glGenSamplers); 
+    glDeleteSamplers = (PFNGLDELETESAMPLERSPROC) SMOL_GETGLPROC(glDeleteSamplers); 
+    glIsSampler = (PFNGLISSAMPLERPROC) SMOL_GETGLPROC(glIsSampler); 
+    glBindSampler = (PFNGLBINDSAMPLERPROC) SMOL_GETGLPROC(glBindSampler); 
+    glSamplerParameteri = (PFNGLSAMPLERPARAMETERIPROC) SMOL_GETGLPROC(glSamplerParameteri); 
+    glSamplerParameteriv = (PFNGLSAMPLERPARAMETERIVPROC) SMOL_GETGLPROC(glSamplerParameteriv); 
+    glSamplerParameterf = (PFNGLSAMPLERPARAMETERFPROC) SMOL_GETGLPROC(glSamplerParameterf); 
+    glSamplerParameterfv = (PFNGLSAMPLERPARAMETERFVPROC) SMOL_GETGLPROC(glSamplerParameterfv); 
+    glSamplerParameterIiv = (PFNGLSAMPLERPARAMETERIIVPROC) SMOL_GETGLPROC(glSamplerParameterIiv); 
+    glSamplerParameterIuiv = (PFNGLSAMPLERPARAMETERIUIVPROC) SMOL_GETGLPROC(glSamplerParameterIuiv); 
+    glGetSamplerParameteriv = (PFNGLGETSAMPLERPARAMETERIVPROC) SMOL_GETGLPROC(glGetSamplerParameteriv); 
+    glGetSamplerParameterIiv = (PFNGLGETSAMPLERPARAMETERIIVPROC) SMOL_GETGLPROC(glGetSamplerParameterIiv); 
+    glGetSamplerParameterfv = (PFNGLGETSAMPLERPARAMETERFVPROC) SMOL_GETGLPROC(glGetSamplerParameterfv); 
+    glGetSamplerParameterIuiv = (PFNGLGETSAMPLERPARAMETERIUIVPROC) SMOL_GETGLPROC(glGetSamplerParameterIuiv); 
+    glQueryCounter = (PFNGLQUERYCOUNTERPROC) SMOL_GETGLPROC(glQueryCounter); 
+    glGetQueryObjecti64v = (PFNGLGETQUERYOBJECTI64VPROC) SMOL_GETGLPROC(glGetQueryObjecti64v); 
+    glGetQueryObjectui64v = (PFNGLGETQUERYOBJECTUI64VPROC) SMOL_GETGLPROC(glGetQueryObjectui64v); 
+    glVertexAttribDivisor = (PFNGLVERTEXATTRIBDIVISORPROC) SMOL_GETGLPROC(glVertexAttribDivisor); 
+    glVertexAttribP1ui = (PFNGLVERTEXATTRIBP1UIPROC) SMOL_GETGLPROC(glVertexAttribP1ui); 
+    glVertexAttribP1uiv = (PFNGLVERTEXATTRIBP1UIVPROC) SMOL_GETGLPROC(glVertexAttribP1uiv); 
+    glVertexAttribP2ui = (PFNGLVERTEXATTRIBP2UIPROC) SMOL_GETGLPROC(glVertexAttribP2ui); 
+    glVertexAttribP2uiv = (PFNGLVERTEXATTRIBP2UIVPROC) SMOL_GETGLPROC(glVertexAttribP2uiv); 
+    glVertexAttribP3ui = (PFNGLVERTEXATTRIBP3UIPROC) SMOL_GETGLPROC(glVertexAttribP3ui); 
+    glVertexAttribP3uiv = (PFNGLVERTEXATTRIBP3UIVPROC) SMOL_GETGLPROC(glVertexAttribP3uiv); 
+    glVertexAttribP4ui = (PFNGLVERTEXATTRIBP4UIPROC) SMOL_GETGLPROC(glVertexAttribP4ui); 
+    glVertexAttribP4uiv = (PFNGLVERTEXATTRIBP4UIVPROC) SMOL_GETGLPROC(glVertexAttribP4uiv); 
+    glMinSampleShading = (PFNGLMINSAMPLESHADINGPROC) SMOL_GETGLPROC(glMinSampleShading); 
+    glBlendEquationi = (PFNGLBLENDEQUATIONIPROC) SMOL_GETGLPROC(glBlendEquationi); 
+    glBlendEquationSeparatei = (PFNGLBLENDEQUATIONSEPARATEIPROC) SMOL_GETGLPROC(glBlendEquationSeparatei); 
+    glBlendFunci = (PFNGLBLENDFUNCIPROC) SMOL_GETGLPROC(glBlendFunci); 
+    glBlendFuncSeparatei = (PFNGLBLENDFUNCSEPARATEIPROC) SMOL_GETGLPROC(glBlendFuncSeparatei); 
+    glDrawArraysIndirect = (PFNGLDRAWARRAYSINDIRECTPROC) SMOL_GETGLPROC(glDrawArraysIndirect); 
+    glDrawElementsIndirect = (PFNGLDRAWELEMENTSINDIRECTPROC) SMOL_GETGLPROC(glDrawElementsIndirect); 
+    glUniform1d = (PFNGLUNIFORM1DPROC) SMOL_GETGLPROC(glUniform1d); 
+    glUniform2d = (PFNGLUNIFORM2DPROC) SMOL_GETGLPROC(glUniform2d); 
+    glUniform3d = (PFNGLUNIFORM3DPROC) SMOL_GETGLPROC(glUniform3d); 
+    glUniform4d = (PFNGLUNIFORM4DPROC) SMOL_GETGLPROC(glUniform4d); 
+    glUniform1dv = (PFNGLUNIFORM1DVPROC) SMOL_GETGLPROC(glUniform1dv); 
+    glUniform2dv = (PFNGLUNIFORM2DVPROC) SMOL_GETGLPROC(glUniform2dv); 
+    glUniform3dv = (PFNGLUNIFORM3DVPROC) SMOL_GETGLPROC(glUniform3dv); 
+    glUniform4dv = (PFNGLUNIFORM4DVPROC) SMOL_GETGLPROC(glUniform4dv); 
+    glUniformMatrix2dv = (PFNGLUNIFORMMATRIX2DVPROC) SMOL_GETGLPROC(glUniformMatrix2dv); 
+    glUniformMatrix3dv = (PFNGLUNIFORMMATRIX3DVPROC) SMOL_GETGLPROC(glUniformMatrix3dv); 
+    glUniformMatrix4dv = (PFNGLUNIFORMMATRIX4DVPROC) SMOL_GETGLPROC(glUniformMatrix4dv); 
+    glUniformMatrix2x3dv = (PFNGLUNIFORMMATRIX2X3DVPROC) SMOL_GETGLPROC(glUniformMatrix2x3dv); 
+    glUniformMatrix2x4dv = (PFNGLUNIFORMMATRIX2X4DVPROC) SMOL_GETGLPROC(glUniformMatrix2x4dv); 
+    glUniformMatrix3x2dv = (PFNGLUNIFORMMATRIX3X2DVPROC) SMOL_GETGLPROC(glUniformMatrix3x2dv); 
+    glUniformMatrix3x4dv = (PFNGLUNIFORMMATRIX3X4DVPROC) SMOL_GETGLPROC(glUniformMatrix3x4dv); 
+    glUniformMatrix4x2dv = (PFNGLUNIFORMMATRIX4X2DVPROC) SMOL_GETGLPROC(glUniformMatrix4x2dv); 
+    glUniformMatrix4x3dv = (PFNGLUNIFORMMATRIX4X3DVPROC) SMOL_GETGLPROC(glUniformMatrix4x3dv); 
+    glGetUniformdv = (PFNGLGETUNIFORMDVPROC) SMOL_GETGLPROC(glGetUniformdv); 
+    glGetSubroutineUniformLocation = (PFNGLGETSUBROUTINEUNIFORMLOCATIONPROC) SMOL_GETGLPROC(glGetSubroutineUniformLocation); 
+    glGetSubroutineIndex = (PFNGLGETSUBROUTINEINDEXPROC) SMOL_GETGLPROC(glGetSubroutineIndex); 
+    glGetActiveSubroutineUniformiv = (PFNGLGETACTIVESUBROUTINEUNIFORMIVPROC) SMOL_GETGLPROC(glGetActiveSubroutineUniformiv); 
+    glGetActiveSubroutineUniformName = (PFNGLGETACTIVESUBROUTINEUNIFORMNAMEPROC) SMOL_GETGLPROC(glGetActiveSubroutineUniformName); 
+    glGetActiveSubroutineName = (PFNGLGETACTIVESUBROUTINENAMEPROC) SMOL_GETGLPROC(glGetActiveSubroutineName); 
+    glUniformSubroutinesuiv = (PFNGLUNIFORMSUBROUTINESUIVPROC) SMOL_GETGLPROC(glUniformSubroutinesuiv); 
+    glGetUniformSubroutineuiv = (PFNGLGETUNIFORMSUBROUTINEUIVPROC) SMOL_GETGLPROC(glGetUniformSubroutineuiv); 
+    glGetProgramStageiv = (PFNGLGETPROGRAMSTAGEIVPROC) SMOL_GETGLPROC(glGetProgramStageiv); 
+    glPatchParameteri = (PFNGLPATCHPARAMETERIPROC) SMOL_GETGLPROC(glPatchParameteri); 
+    glPatchParameterfv = (PFNGLPATCHPARAMETERFVPROC) SMOL_GETGLPROC(glPatchParameterfv); 
+    glBindTransformFeedback = (PFNGLBINDTRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glBindTransformFeedback); 
+    glDeleteTransformFeedbacks = (PFNGLDELETETRANSFORMFEEDBACKSPROC) SMOL_GETGLPROC(glDeleteTransformFeedbacks); 
+    glGenTransformFeedbacks = (PFNGLGENTRANSFORMFEEDBACKSPROC) SMOL_GETGLPROC(glGenTransformFeedbacks); 
+    glIsTransformFeedback = (PFNGLISTRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glIsTransformFeedback); 
+    glPauseTransformFeedback = (PFNGLPAUSETRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glPauseTransformFeedback); 
+    glResumeTransformFeedback = (PFNGLRESUMETRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glResumeTransformFeedback); 
+    glDrawTransformFeedback = (PFNGLDRAWTRANSFORMFEEDBACKPROC) SMOL_GETGLPROC(glDrawTransformFeedback); 
+    glDrawTransformFeedbackStream = (PFNGLDRAWTRANSFORMFEEDBACKSTREAMPROC) SMOL_GETGLPROC(glDrawTransformFeedbackStream); 
+    glBeginQueryIndexed = (PFNGLBEGINQUERYINDEXEDPROC) SMOL_GETGLPROC(glBeginQueryIndexed); 
+    glEndQueryIndexed = (PFNGLENDQUERYINDEXEDPROC) SMOL_GETGLPROC(glEndQueryIndexed); 
+    glGetQueryIndexediv = (PFNGLGETQUERYINDEXEDIVPROC) SMOL_GETGLPROC(glGetQueryIndexediv); 
+    glReleaseShaderCompiler = (PFNGLRELEASESHADERCOMPILERPROC) SMOL_GETGLPROC(glReleaseShaderCompiler); 
+    glShaderBinary = (PFNGLSHADERBINARYPROC) SMOL_GETGLPROC(glShaderBinary); 
+    glGetShaderPrecisionFormat = (PFNGLGETSHADERPRECISIONFORMATPROC) SMOL_GETGLPROC(glGetShaderPrecisionFormat); 
+    glDepthRangef = (PFNGLDEPTHRANGEFPROC) SMOL_GETGLPROC(glDepthRangef); 
+    glClearDepthf = (PFNGLCLEARDEPTHFPROC) SMOL_GETGLPROC(glClearDepthf); 
+    glGetProgramBinary = (PFNGLGETPROGRAMBINARYPROC) SMOL_GETGLPROC(glGetProgramBinary); 
+    glProgramBinary = (PFNGLPROGRAMBINARYPROC) SMOL_GETGLPROC(glProgramBinary); 
+    glProgramParameteri = (PFNGLPROGRAMPARAMETERIPROC) SMOL_GETGLPROC(glProgramParameteri); 
+    glUseProgramStages = (PFNGLUSEPROGRAMSTAGESPROC) SMOL_GETGLPROC(glUseProgramStages); 
+    glActiveShaderProgram = (PFNGLACTIVESHADERPROGRAMPROC) SMOL_GETGLPROC(glActiveShaderProgram); 
+    glCreateShaderProgramv = (PFNGLCREATESHADERPROGRAMVPROC) SMOL_GETGLPROC(glCreateShaderProgramv); 
+    glBindProgramPipeline = (PFNGLBINDPROGRAMPIPELINEPROC) SMOL_GETGLPROC(glBindProgramPipeline); 
+    glDeleteProgramPipelines = (PFNGLDELETEPROGRAMPIPELINESPROC) SMOL_GETGLPROC(glDeleteProgramPipelines); 
+    glGenProgramPipelines = (PFNGLGENPROGRAMPIPELINESPROC) SMOL_GETGLPROC(glGenProgramPipelines); 
+    glIsProgramPipeline = (PFNGLISPROGRAMPIPELINEPROC) SMOL_GETGLPROC(glIsProgramPipeline); 
+    glGetProgramPipelineiv = (PFNGLGETPROGRAMPIPELINEIVPROC) SMOL_GETGLPROC(glGetProgramPipelineiv); 
+    glProgramUniform1i = (PFNGLPROGRAMUNIFORM1IPROC) SMOL_GETGLPROC(glProgramUniform1i); 
+    glProgramUniform1iv = (PFNGLPROGRAMUNIFORM1IVPROC) SMOL_GETGLPROC(glProgramUniform1iv); 
+    glProgramUniform1f = (PFNGLPROGRAMUNIFORM1FPROC) SMOL_GETGLPROC(glProgramUniform1f); 
+    glProgramUniform1fv = (PFNGLPROGRAMUNIFORM1FVPROC) SMOL_GETGLPROC(glProgramUniform1fv); 
+    glProgramUniform1d = (PFNGLPROGRAMUNIFORM1DPROC) SMOL_GETGLPROC(glProgramUniform1d); 
+    glProgramUniform1dv = (PFNGLPROGRAMUNIFORM1DVPROC) SMOL_GETGLPROC(glProgramUniform1dv); 
+    glProgramUniform1ui = (PFNGLPROGRAMUNIFORM1UIPROC) SMOL_GETGLPROC(glProgramUniform1ui); 
+    glProgramUniform1uiv = (PFNGLPROGRAMUNIFORM1UIVPROC) SMOL_GETGLPROC(glProgramUniform1uiv); 
+    glProgramUniform2i = (PFNGLPROGRAMUNIFORM2IPROC) SMOL_GETGLPROC(glProgramUniform2i); 
+    glProgramUniform2iv = (PFNGLPROGRAMUNIFORM2IVPROC) SMOL_GETGLPROC(glProgramUniform2iv); 
+    glProgramUniform2f = (PFNGLPROGRAMUNIFORM2FPROC) SMOL_GETGLPROC(glProgramUniform2f); 
+    glProgramUniform2fv = (PFNGLPROGRAMUNIFORM2FVPROC) SMOL_GETGLPROC(glProgramUniform2fv); 
+    glProgramUniform2d = (PFNGLPROGRAMUNIFORM2DPROC) SMOL_GETGLPROC(glProgramUniform2d); 
+    glProgramUniform2dv = (PFNGLPROGRAMUNIFORM2DVPROC) SMOL_GETGLPROC(glProgramUniform2dv); 
+    glProgramUniform2ui = (PFNGLPROGRAMUNIFORM2UIPROC) SMOL_GETGLPROC(glProgramUniform2ui); 
+    glProgramUniform2uiv = (PFNGLPROGRAMUNIFORM2UIVPROC) SMOL_GETGLPROC(glProgramUniform2uiv); 
+    glProgramUniform3i = (PFNGLPROGRAMUNIFORM3IPROC) SMOL_GETGLPROC(glProgramUniform3i); 
+    glProgramUniform3iv = (PFNGLPROGRAMUNIFORM3IVPROC) SMOL_GETGLPROC(glProgramUniform3iv); 
+    glProgramUniform3f = (PFNGLPROGRAMUNIFORM3FPROC) SMOL_GETGLPROC(glProgramUniform3f); 
+    glProgramUniform3fv = (PFNGLPROGRAMUNIFORM3FVPROC) SMOL_GETGLPROC(glProgramUniform3fv); 
+    glProgramUniform3d = (PFNGLPROGRAMUNIFORM3DPROC) SMOL_GETGLPROC(glProgramUniform3d); 
+    glProgramUniform3dv = (PFNGLPROGRAMUNIFORM3DVPROC) SMOL_GETGLPROC(glProgramUniform3dv); 
+    glProgramUniform3ui = (PFNGLPROGRAMUNIFORM3UIPROC) SMOL_GETGLPROC(glProgramUniform3ui); 
+    glProgramUniform3uiv = (PFNGLPROGRAMUNIFORM3UIVPROC) SMOL_GETGLPROC(glProgramUniform3uiv); 
+    glProgramUniform4i = (PFNGLPROGRAMUNIFORM4IPROC) SMOL_GETGLPROC(glProgramUniform4i); 
+    glProgramUniform4iv = (PFNGLPROGRAMUNIFORM4IVPROC) SMOL_GETGLPROC(glProgramUniform4iv); 
+    glProgramUniform4f = (PFNGLPROGRAMUNIFORM4FPROC) SMOL_GETGLPROC(glProgramUniform4f); 
+    glProgramUniform4fv = (PFNGLPROGRAMUNIFORM4FVPROC) SMOL_GETGLPROC(glProgramUniform4fv); 
+    glProgramUniform4d = (PFNGLPROGRAMUNIFORM4DPROC) SMOL_GETGLPROC(glProgramUniform4d); 
+    glProgramUniform4dv = (PFNGLPROGRAMUNIFORM4DVPROC) SMOL_GETGLPROC(glProgramUniform4dv); 
+    glProgramUniform4ui = (PFNGLPROGRAMUNIFORM4UIPROC) SMOL_GETGLPROC(glProgramUniform4ui); 
+    glProgramUniform4uiv = (PFNGLPROGRAMUNIFORM4UIVPROC) SMOL_GETGLPROC(glProgramUniform4uiv); 
+    glProgramUniformMatrix2fv = (PFNGLPROGRAMUNIFORMMATRIX2FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2fv); 
+    glProgramUniformMatrix3fv = (PFNGLPROGRAMUNIFORMMATRIX3FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3fv); 
+    glProgramUniformMatrix4fv = (PFNGLPROGRAMUNIFORMMATRIX4FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4fv); 
+    glProgramUniformMatrix2dv = (PFNGLPROGRAMUNIFORMMATRIX2DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2dv); 
+    glProgramUniformMatrix3dv = (PFNGLPROGRAMUNIFORMMATRIX3DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3dv); 
+    glProgramUniformMatrix4dv = (PFNGLPROGRAMUNIFORMMATRIX4DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4dv); 
+    glProgramUniformMatrix2x3fv = (PFNGLPROGRAMUNIFORMMATRIX2X3FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x3fv); 
+    glProgramUniformMatrix3x2fv = (PFNGLPROGRAMUNIFORMMATRIX3X2FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x2fv); 
+    glProgramUniformMatrix2x4fv = (PFNGLPROGRAMUNIFORMMATRIX2X4FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x4fv); 
+    glProgramUniformMatrix4x2fv = (PFNGLPROGRAMUNIFORMMATRIX4X2FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x2fv); 
+    glProgramUniformMatrix3x4fv = (PFNGLPROGRAMUNIFORMMATRIX3X4FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x4fv); 
+    glProgramUniformMatrix4x3fv = (PFNGLPROGRAMUNIFORMMATRIX4X3FVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x3fv); 
+    glProgramUniformMatrix2x3dv = (PFNGLPROGRAMUNIFORMMATRIX2X3DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x3dv); 
+    glProgramUniformMatrix3x2dv = (PFNGLPROGRAMUNIFORMMATRIX3X2DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x2dv); 
+    glProgramUniformMatrix2x4dv = (PFNGLPROGRAMUNIFORMMATRIX2X4DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x4dv); 
+    glProgramUniformMatrix4x2dv = (PFNGLPROGRAMUNIFORMMATRIX4X2DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x2dv); 
+    glProgramUniformMatrix3x4dv = (PFNGLPROGRAMUNIFORMMATRIX3X4DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x4dv); 
+    glProgramUniformMatrix4x3dv = (PFNGLPROGRAMUNIFORMMATRIX4X3DVPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x3dv); 
+    glValidateProgramPipeline = (PFNGLVALIDATEPROGRAMPIPELINEPROC) SMOL_GETGLPROC(glValidateProgramPipeline); 
+    glGetProgramPipelineInfoLog = (PFNGLGETPROGRAMPIPELINEINFOLOGPROC) SMOL_GETGLPROC(glGetProgramPipelineInfoLog); 
+    glVertexAttribL1d = (PFNGLVERTEXATTRIBL1DPROC) SMOL_GETGLPROC(glVertexAttribL1d); 
+    glVertexAttribL2d = (PFNGLVERTEXATTRIBL2DPROC) SMOL_GETGLPROC(glVertexAttribL2d); 
+    glVertexAttribL3d = (PFNGLVERTEXATTRIBL3DPROC) SMOL_GETGLPROC(glVertexAttribL3d); 
+    glVertexAttribL4d = (PFNGLVERTEXATTRIBL4DPROC) SMOL_GETGLPROC(glVertexAttribL4d); 
+    glVertexAttribL1dv = (PFNGLVERTEXATTRIBL1DVPROC) SMOL_GETGLPROC(glVertexAttribL1dv); 
+    glVertexAttribL2dv = (PFNGLVERTEXATTRIBL2DVPROC) SMOL_GETGLPROC(glVertexAttribL2dv); 
+    glVertexAttribL3dv = (PFNGLVERTEXATTRIBL3DVPROC) SMOL_GETGLPROC(glVertexAttribL3dv); 
+    glVertexAttribL4dv = (PFNGLVERTEXATTRIBL4DVPROC) SMOL_GETGLPROC(glVertexAttribL4dv); 
+    glVertexAttribLPointer = (PFNGLVERTEXATTRIBLPOINTERPROC) SMOL_GETGLPROC(glVertexAttribLPointer); 
+    glGetVertexAttribLdv = (PFNGLGETVERTEXATTRIBLDVPROC) SMOL_GETGLPROC(glGetVertexAttribLdv); 
+    glViewportArrayv = (PFNGLVIEWPORTARRAYVPROC) SMOL_GETGLPROC(glViewportArrayv); 
+    glViewportIndexedf = (PFNGLVIEWPORTINDEXEDFPROC) SMOL_GETGLPROC(glViewportIndexedf); 
+    glViewportIndexedfv = (PFNGLVIEWPORTINDEXEDFVPROC) SMOL_GETGLPROC(glViewportIndexedfv); 
+    glScissorArrayv = (PFNGLSCISSORARRAYVPROC) SMOL_GETGLPROC(glScissorArrayv); 
+    glScissorIndexed = (PFNGLSCISSORINDEXEDPROC) SMOL_GETGLPROC(glScissorIndexed); 
+    glScissorIndexedv = (PFNGLSCISSORINDEXEDVPROC) SMOL_GETGLPROC(glScissorIndexedv); 
+    glDepthRangeArrayv = (PFNGLDEPTHRANGEARRAYVPROC) SMOL_GETGLPROC(glDepthRangeArrayv); 
+    glDepthRangeIndexed = (PFNGLDEPTHRANGEINDEXEDPROC) SMOL_GETGLPROC(glDepthRangeIndexed); 
+    glGetFloati_v = (PFNGLGETFLOATI_VPROC) SMOL_GETGLPROC(glGetFloati_v); 
+    glGetDoublei_v = (PFNGLGETDOUBLEI_VPROC) SMOL_GETGLPROC(glGetDoublei_v); 
+    glDrawArraysInstancedBaseInstance = (PFNGLDRAWARRAYSINSTANCEDBASEINSTANCEPROC) SMOL_GETGLPROC(glDrawArraysInstancedBaseInstance); 
+    glDrawElementsInstancedBaseInstance = (PFNGLDRAWELEMENTSINSTANCEDBASEINSTANCEPROC) SMOL_GETGLPROC(glDrawElementsInstancedBaseInstance); 
+    glDrawElementsInstancedBaseVertexBaseInstance = (PFNGLDRAWELEMENTSINSTANCEDBASEVERTEXBASEINSTANCEPROC) SMOL_GETGLPROC(glDrawElementsInstancedBaseVertexBaseInstance); 
+    glGetInternalformativ = (PFNGLGETINTERNALFORMATIVPROC) SMOL_GETGLPROC(glGetInternalformativ); 
+    glGetActiveAtomicCounterBufferiv = (PFNGLGETACTIVEATOMICCOUNTERBUFFERIVPROC) SMOL_GETGLPROC(glGetActiveAtomicCounterBufferiv); 
+    glBindImageTexture = (PFNGLBINDIMAGETEXTUREPROC) SMOL_GETGLPROC(glBindImageTexture); 
+    glMemoryBarrier = (PFNGLMEMORYBARRIERPROC) SMOL_GETGLPROC(glMemoryBarrier); 
+    glTexStorage1D = (PFNGLTEXSTORAGE1DPROC) SMOL_GETGLPROC(glTexStorage1D); 
+    glTexStorage2D = (PFNGLTEXSTORAGE2DPROC) SMOL_GETGLPROC(glTexStorage2D); 
+    glTexStorage3D = (PFNGLTEXSTORAGE3DPROC) SMOL_GETGLPROC(glTexStorage3D); 
+    glDrawTransformFeedbackInstanced = (PFNGLDRAWTRANSFORMFEEDBACKINSTANCEDPROC) SMOL_GETGLPROC(glDrawTransformFeedbackInstanced); 
+    glDrawTransformFeedbackStreamInstanced = (PFNGLDRAWTRANSFORMFEEDBACKSTREAMINSTANCEDPROC) SMOL_GETGLPROC(glDrawTransformFeedbackStreamInstanced); 
+    glClearBufferData = (PFNGLCLEARBUFFERDATAPROC) SMOL_GETGLPROC(glClearBufferData); 
+    glClearBufferSubData = (PFNGLCLEARBUFFERSUBDATAPROC) SMOL_GETGLPROC(glClearBufferSubData); 
+    glDispatchCompute = (PFNGLDISPATCHCOMPUTEPROC) SMOL_GETGLPROC(glDispatchCompute); 
+    glDispatchComputeIndirect = (PFNGLDISPATCHCOMPUTEINDIRECTPROC) SMOL_GETGLPROC(glDispatchComputeIndirect); 
+    glCopyImageSubData = (PFNGLCOPYIMAGESUBDATAPROC) SMOL_GETGLPROC(glCopyImageSubData); 
+    glFramebufferParameteri = (PFNGLFRAMEBUFFERPARAMETERIPROC) SMOL_GETGLPROC(glFramebufferParameteri); 
+    glGetFramebufferParameteriv = (PFNGLGETFRAMEBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetFramebufferParameteriv); 
+    glGetInternalformati64v = (PFNGLGETINTERNALFORMATI64VPROC) SMOL_GETGLPROC(glGetInternalformati64v); 
+    glInvalidateTexSubImage = (PFNGLINVALIDATETEXSUBIMAGEPROC) SMOL_GETGLPROC(glInvalidateTexSubImage); 
+    glInvalidateTexImage = (PFNGLINVALIDATETEXIMAGEPROC) SMOL_GETGLPROC(glInvalidateTexImage); 
+    glInvalidateBufferSubData = (PFNGLINVALIDATEBUFFERSUBDATAPROC) SMOL_GETGLPROC(glInvalidateBufferSubData); 
+    glInvalidateBufferData = (PFNGLINVALIDATEBUFFERDATAPROC) SMOL_GETGLPROC(glInvalidateBufferData); 
+    glInvalidateFramebuffer = (PFNGLINVALIDATEFRAMEBUFFERPROC) SMOL_GETGLPROC(glInvalidateFramebuffer); 
+    glInvalidateSubFramebuffer = (PFNGLINVALIDATESUBFRAMEBUFFERPROC) SMOL_GETGLPROC(glInvalidateSubFramebuffer); 
+    glMultiDrawArraysIndirect = (PFNGLMULTIDRAWARRAYSINDIRECTPROC) SMOL_GETGLPROC(glMultiDrawArraysIndirect); 
+    glMultiDrawElementsIndirect = (PFNGLMULTIDRAWELEMENTSINDIRECTPROC) SMOL_GETGLPROC(glMultiDrawElementsIndirect); 
+    glGetProgramInterfaceiv = (PFNGLGETPROGRAMINTERFACEIVPROC) SMOL_GETGLPROC(glGetProgramInterfaceiv); 
+    glGetProgramResourceIndex = (PFNGLGETPROGRAMRESOURCEINDEXPROC) SMOL_GETGLPROC(glGetProgramResourceIndex); 
+    glGetProgramResourceName = (PFNGLGETPROGRAMRESOURCENAMEPROC) SMOL_GETGLPROC(glGetProgramResourceName); 
+    glGetProgramResourceiv = (PFNGLGETPROGRAMRESOURCEIVPROC) SMOL_GETGLPROC(glGetProgramResourceiv); 
+    glGetProgramResourceLocation = (PFNGLGETPROGRAMRESOURCELOCATIONPROC) SMOL_GETGLPROC(glGetProgramResourceLocation); 
+    glGetProgramResourceLocationIndex = (PFNGLGETPROGRAMRESOURCELOCATIONINDEXPROC) SMOL_GETGLPROC(glGetProgramResourceLocationIndex); 
+    glShaderStorageBlockBinding = (PFNGLSHADERSTORAGEBLOCKBINDINGPROC) SMOL_GETGLPROC(glShaderStorageBlockBinding); 
+    glTexBufferRange = (PFNGLTEXBUFFERRANGEPROC) SMOL_GETGLPROC(glTexBufferRange); 
+    glTexStorage2DMultisample = (PFNGLTEXSTORAGE2DMULTISAMPLEPROC) SMOL_GETGLPROC(glTexStorage2DMultisample); 
+    glTexStorage3DMultisample = (PFNGLTEXSTORAGE3DMULTISAMPLEPROC) SMOL_GETGLPROC(glTexStorage3DMultisample); 
+    glTextureView = (PFNGLTEXTUREVIEWPROC) SMOL_GETGLPROC(glTextureView); 
+    glBindVertexBuffer = (PFNGLBINDVERTEXBUFFERPROC) SMOL_GETGLPROC(glBindVertexBuffer); 
+    glVertexAttribFormat = (PFNGLVERTEXATTRIBFORMATPROC) SMOL_GETGLPROC(glVertexAttribFormat); 
+    glVertexAttribIFormat = (PFNGLVERTEXATTRIBIFORMATPROC) SMOL_GETGLPROC(glVertexAttribIFormat); 
+    glVertexAttribLFormat = (PFNGLVERTEXATTRIBLFORMATPROC) SMOL_GETGLPROC(glVertexAttribLFormat); 
+    glVertexAttribBinding = (PFNGLVERTEXATTRIBBINDINGPROC) SMOL_GETGLPROC(glVertexAttribBinding); 
+    glVertexBindingDivisor = (PFNGLVERTEXBINDINGDIVISORPROC) SMOL_GETGLPROC(glVertexBindingDivisor); 
+    glDebugMessageControl = (PFNGLDEBUGMESSAGECONTROLPROC) SMOL_GETGLPROC(glDebugMessageControl); 
+    glDebugMessageInsert = (PFNGLDEBUGMESSAGEINSERTPROC) SMOL_GETGLPROC(glDebugMessageInsert); 
+    glDebugMessageCallback = (PFNGLDEBUGMESSAGECALLBACKPROC) SMOL_GETGLPROC(glDebugMessageCallback); 
+    glGetDebugMessageLog = (PFNGLGETDEBUGMESSAGELOGPROC) SMOL_GETGLPROC(glGetDebugMessageLog); 
+    glPushDebugGroup = (PFNGLPUSHDEBUGGROUPPROC) SMOL_GETGLPROC(glPushDebugGroup); 
+    glPopDebugGroup = (PFNGLPOPDEBUGGROUPPROC) SMOL_GETGLPROC(glPopDebugGroup); 
+    glObjectLabel = (PFNGLOBJECTLABELPROC) SMOL_GETGLPROC(glObjectLabel); 
+    glGetObjectLabel = (PFNGLGETOBJECTLABELPROC) SMOL_GETGLPROC(glGetObjectLabel); 
+    glObjectPtrLabel = (PFNGLOBJECTPTRLABELPROC) SMOL_GETGLPROC(glObjectPtrLabel); 
+    glGetObjectPtrLabel = (PFNGLGETOBJECTPTRLABELPROC) SMOL_GETGLPROC(glGetObjectPtrLabel); 
+    glBufferStorage = (PFNGLBUFFERSTORAGEPROC) SMOL_GETGLPROC(glBufferStorage); 
+    glClearTexImage = (PFNGLCLEARTEXIMAGEPROC) SMOL_GETGLPROC(glClearTexImage); 
+    glClearTexSubImage = (PFNGLCLEARTEXSUBIMAGEPROC) SMOL_GETGLPROC(glClearTexSubImage); 
+    glBindBuffersBase = (PFNGLBINDBUFFERSBASEPROC) SMOL_GETGLPROC(glBindBuffersBase); 
+    glBindBuffersRange = (PFNGLBINDBUFFERSRANGEPROC) SMOL_GETGLPROC(glBindBuffersRange); 
+    glBindTextures = (PFNGLBINDTEXTURESPROC) SMOL_GETGLPROC(glBindTextures); 
+    glBindSamplers = (PFNGLBINDSAMPLERSPROC) SMOL_GETGLPROC(glBindSamplers); 
+    glBindImageTextures = (PFNGLBINDIMAGETEXTURESPROC) SMOL_GETGLPROC(glBindImageTextures); 
+    glBindVertexBuffers = (PFNGLBINDVERTEXBUFFERSPROC) SMOL_GETGLPROC(glBindVertexBuffers); 
+    glClipControl = (PFNGLCLIPCONTROLPROC) SMOL_GETGLPROC(glClipControl); 
+    glCreateTransformFeedbacks = (PFNGLCREATETRANSFORMFEEDBACKSPROC) SMOL_GETGLPROC(glCreateTransformFeedbacks); 
+    glTransformFeedbackBufferBase = (PFNGLTRANSFORMFEEDBACKBUFFERBASEPROC) SMOL_GETGLPROC(glTransformFeedbackBufferBase); 
+    glTransformFeedbackBufferRange = (PFNGLTRANSFORMFEEDBACKBUFFERRANGEPROC) SMOL_GETGLPROC(glTransformFeedbackBufferRange); 
+    glGetTransformFeedbackiv = (PFNGLGETTRANSFORMFEEDBACKIVPROC) SMOL_GETGLPROC(glGetTransformFeedbackiv); 
+    glGetTransformFeedbacki_v = (PFNGLGETTRANSFORMFEEDBACKI_VPROC) SMOL_GETGLPROC(glGetTransformFeedbacki_v); 
+    glGetTransformFeedbacki64_v = (PFNGLGETTRANSFORMFEEDBACKI64_VPROC) SMOL_GETGLPROC(glGetTransformFeedbacki64_v); 
+    glCreateBuffers = (PFNGLCREATEBUFFERSPROC) SMOL_GETGLPROC(glCreateBuffers); 
+    glNamedBufferStorage = (PFNGLNAMEDBUFFERSTORAGEPROC) SMOL_GETGLPROC(glNamedBufferStorage); 
+    glNamedBufferData = (PFNGLNAMEDBUFFERDATAPROC) SMOL_GETGLPROC(glNamedBufferData); 
+    glNamedBufferSubData = (PFNGLNAMEDBUFFERSUBDATAPROC) SMOL_GETGLPROC(glNamedBufferSubData); 
+    glCopyNamedBufferSubData = (PFNGLCOPYNAMEDBUFFERSUBDATAPROC) SMOL_GETGLPROC(glCopyNamedBufferSubData); 
+    glClearNamedBufferData = (PFNGLCLEARNAMEDBUFFERDATAPROC) SMOL_GETGLPROC(glClearNamedBufferData); 
+    glClearNamedBufferSubData = (PFNGLCLEARNAMEDBUFFERSUBDATAPROC) SMOL_GETGLPROC(glClearNamedBufferSubData); 
+    glMapNamedBuffer = (PFNGLMAPNAMEDBUFFERPROC) SMOL_GETGLPROC(glMapNamedBuffer); 
+    glMapNamedBufferRange = (PFNGLMAPNAMEDBUFFERRANGEPROC) SMOL_GETGLPROC(glMapNamedBufferRange); 
+    glUnmapNamedBuffer = (PFNGLUNMAPNAMEDBUFFERPROC) SMOL_GETGLPROC(glUnmapNamedBuffer); 
+    glFlushMappedNamedBufferRange = (PFNGLFLUSHMAPPEDNAMEDBUFFERRANGEPROC) SMOL_GETGLPROC(glFlushMappedNamedBufferRange); 
+    glGetNamedBufferParameteriv = (PFNGLGETNAMEDBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetNamedBufferParameteriv); 
+    glGetNamedBufferParameteri64v = (PFNGLGETNAMEDBUFFERPARAMETERI64VPROC) SMOL_GETGLPROC(glGetNamedBufferParameteri64v); 
+    glGetNamedBufferPointerv = (PFNGLGETNAMEDBUFFERPOINTERVPROC) SMOL_GETGLPROC(glGetNamedBufferPointerv); 
+    glGetNamedBufferSubData = (PFNGLGETNAMEDBUFFERSUBDATAPROC) SMOL_GETGLPROC(glGetNamedBufferSubData); 
+    glCreateFramebuffers = (PFNGLCREATEFRAMEBUFFERSPROC) SMOL_GETGLPROC(glCreateFramebuffers); 
+    glNamedFramebufferRenderbuffer = (PFNGLNAMEDFRAMEBUFFERRENDERBUFFERPROC) SMOL_GETGLPROC(glNamedFramebufferRenderbuffer); 
+    glNamedFramebufferParameteri = (PFNGLNAMEDFRAMEBUFFERPARAMETERIPROC) SMOL_GETGLPROC(glNamedFramebufferParameteri); 
+    glNamedFramebufferTexture = (PFNGLNAMEDFRAMEBUFFERTEXTUREPROC) SMOL_GETGLPROC(glNamedFramebufferTexture); 
+    glNamedFramebufferTextureLayer = (PFNGLNAMEDFRAMEBUFFERTEXTURELAYERPROC) SMOL_GETGLPROC(glNamedFramebufferTextureLayer); 
+    glNamedFramebufferDrawBuffer = (PFNGLNAMEDFRAMEBUFFERDRAWBUFFERPROC) SMOL_GETGLPROC(glNamedFramebufferDrawBuffer); 
+    glNamedFramebufferDrawBuffers = (PFNGLNAMEDFRAMEBUFFERDRAWBUFFERSPROC) SMOL_GETGLPROC(glNamedFramebufferDrawBuffers); 
+    glNamedFramebufferReadBuffer = (PFNGLNAMEDFRAMEBUFFERREADBUFFERPROC) SMOL_GETGLPROC(glNamedFramebufferReadBuffer); 
+    glInvalidateNamedFramebufferData = (PFNGLINVALIDATENAMEDFRAMEBUFFERDATAPROC) SMOL_GETGLPROC(glInvalidateNamedFramebufferData); 
+    glInvalidateNamedFramebufferSubData = (PFNGLINVALIDATENAMEDFRAMEBUFFERSUBDATAPROC) SMOL_GETGLPROC(glInvalidateNamedFramebufferSubData); 
+    glClearNamedFramebufferiv = (PFNGLCLEARNAMEDFRAMEBUFFERIVPROC) SMOL_GETGLPROC(glClearNamedFramebufferiv); 
+    glClearNamedFramebufferuiv = (PFNGLCLEARNAMEDFRAMEBUFFERUIVPROC) SMOL_GETGLPROC(glClearNamedFramebufferuiv); 
+    glClearNamedFramebufferfv = (PFNGLCLEARNAMEDFRAMEBUFFERFVPROC) SMOL_GETGLPROC(glClearNamedFramebufferfv); 
+    glClearNamedFramebufferfi = (PFNGLCLEARNAMEDFRAMEBUFFERFIPROC) SMOL_GETGLPROC(glClearNamedFramebufferfi); 
+    glBlitNamedFramebuffer = (PFNGLBLITNAMEDFRAMEBUFFERPROC) SMOL_GETGLPROC(glBlitNamedFramebuffer); 
+    glCheckNamedFramebufferStatus = (PFNGLCHECKNAMEDFRAMEBUFFERSTATUSPROC) SMOL_GETGLPROC(glCheckNamedFramebufferStatus); 
+    glGetNamedFramebufferParameteriv = (PFNGLGETNAMEDFRAMEBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetNamedFramebufferParameteriv); 
+    glGetNamedFramebufferAttachmentParameteriv = (PFNGLGETNAMEDFRAMEBUFFERATTACHMENTPARAMETERIVPROC) SMOL_GETGLPROC(glGetNamedFramebufferAttachmentParameteriv); 
+    glCreateRenderbuffers = (PFNGLCREATERENDERBUFFERSPROC) SMOL_GETGLPROC(glCreateRenderbuffers); 
+    glNamedRenderbufferStorage = (PFNGLNAMEDRENDERBUFFERSTORAGEPROC) SMOL_GETGLPROC(glNamedRenderbufferStorage); 
+    glNamedRenderbufferStorageMultisample = (PFNGLNAMEDRENDERBUFFERSTORAGEMULTISAMPLEPROC) SMOL_GETGLPROC(glNamedRenderbufferStorageMultisample); 
+    glGetNamedRenderbufferParameteriv = (PFNGLGETNAMEDRENDERBUFFERPARAMETERIVPROC) SMOL_GETGLPROC(glGetNamedRenderbufferParameteriv); 
+    glCreateTextures = (PFNGLCREATETEXTURESPROC) SMOL_GETGLPROC(glCreateTextures); 
+    glTextureBuffer = (PFNGLTEXTUREBUFFERPROC) SMOL_GETGLPROC(glTextureBuffer); 
+    glTextureBufferRange = (PFNGLTEXTUREBUFFERRANGEPROC) SMOL_GETGLPROC(glTextureBufferRange); 
+    glTextureStorage1D = (PFNGLTEXTURESTORAGE1DPROC) SMOL_GETGLPROC(glTextureStorage1D); 
+    glTextureStorage2D = (PFNGLTEXTURESTORAGE2DPROC) SMOL_GETGLPROC(glTextureStorage2D); 
+    glTextureStorage3D = (PFNGLTEXTURESTORAGE3DPROC) SMOL_GETGLPROC(glTextureStorage3D); 
+    glTextureStorage2DMultisample = (PFNGLTEXTURESTORAGE2DMULTISAMPLEPROC) SMOL_GETGLPROC(glTextureStorage2DMultisample); 
+    glTextureStorage3DMultisample = (PFNGLTEXTURESTORAGE3DMULTISAMPLEPROC) SMOL_GETGLPROC(glTextureStorage3DMultisample); 
+    glTextureSubImage1D = (PFNGLTEXTURESUBIMAGE1DPROC) SMOL_GETGLPROC(glTextureSubImage1D); 
+    glTextureSubImage2D = (PFNGLTEXTURESUBIMAGE2DPROC) SMOL_GETGLPROC(glTextureSubImage2D); 
+    glTextureSubImage3D = (PFNGLTEXTURESUBIMAGE3DPROC) SMOL_GETGLPROC(glTextureSubImage3D); 
+    glCompressedTextureSubImage1D = (PFNGLCOMPRESSEDTEXTURESUBIMAGE1DPROC) SMOL_GETGLPROC(glCompressedTextureSubImage1D); 
+    glCompressedTextureSubImage2D = (PFNGLCOMPRESSEDTEXTURESUBIMAGE2DPROC) SMOL_GETGLPROC(glCompressedTextureSubImage2D); 
+    glCompressedTextureSubImage3D = (PFNGLCOMPRESSEDTEXTURESUBIMAGE3DPROC) SMOL_GETGLPROC(glCompressedTextureSubImage3D); 
+    glCopyTextureSubImage1D = (PFNGLCOPYTEXTURESUBIMAGE1DPROC) SMOL_GETGLPROC(glCopyTextureSubImage1D); 
+    glCopyTextureSubImage2D = (PFNGLCOPYTEXTURESUBIMAGE2DPROC) SMOL_GETGLPROC(glCopyTextureSubImage2D); 
+    glCopyTextureSubImage3D = (PFNGLCOPYTEXTURESUBIMAGE3DPROC) SMOL_GETGLPROC(glCopyTextureSubImage3D); 
+    glTextureParameterf = (PFNGLTEXTUREPARAMETERFPROC) SMOL_GETGLPROC(glTextureParameterf); 
+    glTextureParameterfv = (PFNGLTEXTUREPARAMETERFVPROC) SMOL_GETGLPROC(glTextureParameterfv); 
+    glTextureParameteri = (PFNGLTEXTUREPARAMETERIPROC) SMOL_GETGLPROC(glTextureParameteri); 
+    glTextureParameterIiv = (PFNGLTEXTUREPARAMETERIIVPROC) SMOL_GETGLPROC(glTextureParameterIiv); 
+    glTextureParameterIuiv = (PFNGLTEXTUREPARAMETERIUIVPROC) SMOL_GETGLPROC(glTextureParameterIuiv); 
+    glTextureParameteriv = (PFNGLTEXTUREPARAMETERIVPROC) SMOL_GETGLPROC(glTextureParameteriv); 
+    glGenerateTextureMipmap = (PFNGLGENERATETEXTUREMIPMAPPROC) SMOL_GETGLPROC(glGenerateTextureMipmap); 
+    glBindTextureUnit = (PFNGLBINDTEXTUREUNITPROC) SMOL_GETGLPROC(glBindTextureUnit); 
+    glGetTextureImage = (PFNGLGETTEXTUREIMAGEPROC) SMOL_GETGLPROC(glGetTextureImage); 
+    glGetCompressedTextureImage = (PFNGLGETCOMPRESSEDTEXTUREIMAGEPROC) SMOL_GETGLPROC(glGetCompressedTextureImage); 
+    glGetTextureLevelParameterfv = (PFNGLGETTEXTURELEVELPARAMETERFVPROC) SMOL_GETGLPROC(glGetTextureLevelParameterfv); 
+    glGetTextureLevelParameteriv = (PFNGLGETTEXTURELEVELPARAMETERIVPROC) SMOL_GETGLPROC(glGetTextureLevelParameteriv); 
+    glGetTextureParameterfv = (PFNGLGETTEXTUREPARAMETERFVPROC) SMOL_GETGLPROC(glGetTextureParameterfv); 
+    glGetTextureParameterIiv = (PFNGLGETTEXTUREPARAMETERIIVPROC) SMOL_GETGLPROC(glGetTextureParameterIiv); 
+    glGetTextureParameterIuiv = (PFNGLGETTEXTUREPARAMETERIUIVPROC) SMOL_GETGLPROC(glGetTextureParameterIuiv); 
+    glGetTextureParameteriv = (PFNGLGETTEXTUREPARAMETERIVPROC) SMOL_GETGLPROC(glGetTextureParameteriv); 
+    glCreateVertexArrays = (PFNGLCREATEVERTEXARRAYSPROC) SMOL_GETGLPROC(glCreateVertexArrays); 
+    glDisableVertexArrayAttrib = (PFNGLDISABLEVERTEXARRAYATTRIBPROC) SMOL_GETGLPROC(glDisableVertexArrayAttrib); 
+    glEnableVertexArrayAttrib = (PFNGLENABLEVERTEXARRAYATTRIBPROC) SMOL_GETGLPROC(glEnableVertexArrayAttrib); 
+    glVertexArrayElementBuffer = (PFNGLVERTEXARRAYELEMENTBUFFERPROC) SMOL_GETGLPROC(glVertexArrayElementBuffer); 
+    glVertexArrayVertexBuffer = (PFNGLVERTEXARRAYVERTEXBUFFERPROC) SMOL_GETGLPROC(glVertexArrayVertexBuffer); 
+    glVertexArrayVertexBuffers = (PFNGLVERTEXARRAYVERTEXBUFFERSPROC) SMOL_GETGLPROC(glVertexArrayVertexBuffers); 
+    glVertexArrayAttribBinding = (PFNGLVERTEXARRAYATTRIBBINDINGPROC) SMOL_GETGLPROC(glVertexArrayAttribBinding); 
+    glVertexArrayAttribFormat = (PFNGLVERTEXARRAYATTRIBFORMATPROC) SMOL_GETGLPROC(glVertexArrayAttribFormat); 
+    glVertexArrayAttribIFormat = (PFNGLVERTEXARRAYATTRIBIFORMATPROC) SMOL_GETGLPROC(glVertexArrayAttribIFormat); 
+    glVertexArrayAttribLFormat = (PFNGLVERTEXARRAYATTRIBLFORMATPROC) SMOL_GETGLPROC(glVertexArrayAttribLFormat); 
+    glVertexArrayBindingDivisor = (PFNGLVERTEXARRAYBINDINGDIVISORPROC) SMOL_GETGLPROC(glVertexArrayBindingDivisor); 
+    glGetVertexArrayiv = (PFNGLGETVERTEXARRAYIVPROC) SMOL_GETGLPROC(glGetVertexArrayiv); 
+    glGetVertexArrayIndexediv = (PFNGLGETVERTEXARRAYINDEXEDIVPROC) SMOL_GETGLPROC(glGetVertexArrayIndexediv); 
+    glGetVertexArrayIndexed64iv = (PFNGLGETVERTEXARRAYINDEXED64IVPROC) SMOL_GETGLPROC(glGetVertexArrayIndexed64iv); 
+    glCreateSamplers = (PFNGLCREATESAMPLERSPROC) SMOL_GETGLPROC(glCreateSamplers); 
+    glCreateProgramPipelines = (PFNGLCREATEPROGRAMPIPELINESPROC) SMOL_GETGLPROC(glCreateProgramPipelines); 
+    glCreateQueries = (PFNGLCREATEQUERIESPROC) SMOL_GETGLPROC(glCreateQueries); 
+    glGetQueryBufferObjecti64v = (PFNGLGETQUERYBUFFEROBJECTI64VPROC) SMOL_GETGLPROC(glGetQueryBufferObjecti64v); 
+    glGetQueryBufferObjectiv = (PFNGLGETQUERYBUFFEROBJECTIVPROC) SMOL_GETGLPROC(glGetQueryBufferObjectiv); 
+    glGetQueryBufferObjectui64v = (PFNGLGETQUERYBUFFEROBJECTUI64VPROC) SMOL_GETGLPROC(glGetQueryBufferObjectui64v); 
+    glGetQueryBufferObjectuiv = (PFNGLGETQUERYBUFFEROBJECTUIVPROC) SMOL_GETGLPROC(glGetQueryBufferObjectuiv); 
+    glMemoryBarrierByRegion = (PFNGLMEMORYBARRIERBYREGIONPROC) SMOL_GETGLPROC(glMemoryBarrierByRegion); 
+    glGetTextureSubImage = (PFNGLGETTEXTURESUBIMAGEPROC) SMOL_GETGLPROC(glGetTextureSubImage); 
+    glGetCompressedTextureSubImage = (PFNGLGETCOMPRESSEDTEXTURESUBIMAGEPROC) SMOL_GETGLPROC(glGetCompressedTextureSubImage); 
+    glGetGraphicsResetStatus = (PFNGLGETGRAPHICSRESETSTATUSPROC) SMOL_GETGLPROC(glGetGraphicsResetStatus); 
+    glGetnCompressedTexImage = (PFNGLGETNCOMPRESSEDTEXIMAGEPROC) SMOL_GETGLPROC(glGetnCompressedTexImage); 
+    glGetnTexImage = (PFNGLGETNTEXIMAGEPROC) SMOL_GETGLPROC(glGetnTexImage); 
+    glGetnUniformdv = (PFNGLGETNUNIFORMDVPROC) SMOL_GETGLPROC(glGetnUniformdv); 
+    glGetnUniformfv = (PFNGLGETNUNIFORMFVPROC) SMOL_GETGLPROC(glGetnUniformfv); 
+    glGetnUniformiv = (PFNGLGETNUNIFORMIVPROC) SMOL_GETGLPROC(glGetnUniformiv); 
+    glGetnUniformuiv = (PFNGLGETNUNIFORMUIVPROC) SMOL_GETGLPROC(glGetnUniformuiv); 
+    glReadnPixels = (PFNGLREADNPIXELSPROC) SMOL_GETGLPROC(glReadnPixels); 
+    glTextureBarrier = (PFNGLTEXTUREBARRIERPROC) SMOL_GETGLPROC(glTextureBarrier); 
+    glSpecializeShader = (PFNGLSPECIALIZESHADERPROC) SMOL_GETGLPROC(glSpecializeShader); 
+    glMultiDrawArraysIndirectCount = (PFNGLMULTIDRAWARRAYSINDIRECTCOUNTPROC) SMOL_GETGLPROC(glMultiDrawArraysIndirectCount); 
+    glMultiDrawElementsIndirectCount = (PFNGLMULTIDRAWELEMENTSINDIRECTCOUNTPROC) SMOL_GETGLPROC(glMultiDrawElementsIndirectCount); 
+    glPolygonOffsetClamp = (PFNGLPOLYGONOFFSETCLAMPPROC) SMOL_GETGLPROC(glPolygonOffsetClamp); 
+    glPrimitiveBoundingBoxARB = (PFNGLPRIMITIVEBOUNDINGBOXARBPROC) SMOL_GETGLPROC(glPrimitiveBoundingBoxARB); 
+    glGetTextureHandleARB = (PFNGLGETTEXTUREHANDLEARBPROC) SMOL_GETGLPROC(glGetTextureHandleARB); 
+    glGetTextureSamplerHandleARB = (PFNGLGETTEXTURESAMPLERHANDLEARBPROC) SMOL_GETGLPROC(glGetTextureSamplerHandleARB); 
+    glMakeTextureHandleResidentARB = (PFNGLMAKETEXTUREHANDLERESIDENTARBPROC) SMOL_GETGLPROC(glMakeTextureHandleResidentARB); 
+    glMakeTextureHandleNonResidentARB = (PFNGLMAKETEXTUREHANDLENONRESIDENTARBPROC) SMOL_GETGLPROC(glMakeTextureHandleNonResidentARB); 
+    glGetImageHandleARB = (PFNGLGETIMAGEHANDLEARBPROC) SMOL_GETGLPROC(glGetImageHandleARB); 
+    glMakeImageHandleResidentARB = (PFNGLMAKEIMAGEHANDLERESIDENTARBPROC) SMOL_GETGLPROC(glMakeImageHandleResidentARB); 
+    glMakeImageHandleNonResidentARB = (PFNGLMAKEIMAGEHANDLENONRESIDENTARBPROC) SMOL_GETGLPROC(glMakeImageHandleNonResidentARB); 
+    glUniformHandleui64ARB = (PFNGLUNIFORMHANDLEUI64ARBPROC) SMOL_GETGLPROC(glUniformHandleui64ARB); 
+    glUniformHandleui64vARB = (PFNGLUNIFORMHANDLEUI64VARBPROC) SMOL_GETGLPROC(glUniformHandleui64vARB); 
+    glProgramUniformHandleui64ARB = (PFNGLPROGRAMUNIFORMHANDLEUI64ARBPROC) SMOL_GETGLPROC(glProgramUniformHandleui64ARB); 
+    glProgramUniformHandleui64vARB = (PFNGLPROGRAMUNIFORMHANDLEUI64VARBPROC) SMOL_GETGLPROC(glProgramUniformHandleui64vARB); 
+    glIsTextureHandleResidentARB = (PFNGLISTEXTUREHANDLERESIDENTARBPROC) SMOL_GETGLPROC(glIsTextureHandleResidentARB); 
+    glIsImageHandleResidentARB = (PFNGLISIMAGEHANDLERESIDENTARBPROC) SMOL_GETGLPROC(glIsImageHandleResidentARB); 
+    glVertexAttribL1ui64ARB = (PFNGLVERTEXATTRIBL1UI64ARBPROC) SMOL_GETGLPROC(glVertexAttribL1ui64ARB); 
+    glVertexAttribL1ui64vARB = (PFNGLVERTEXATTRIBL1UI64VARBPROC) SMOL_GETGLPROC(glVertexAttribL1ui64vARB); 
+    glGetVertexAttribLui64vARB = (PFNGLGETVERTEXATTRIBLUI64VARBPROC) SMOL_GETGLPROC(glGetVertexAttribLui64vARB); 
+    glCreateSyncFromCLeventARB = (PFNGLCREATESYNCFROMCLEVENTARBPROC) SMOL_GETGLPROC(glCreateSyncFromCLeventARB); 
+    glDispatchComputeGroupSizeARB = (PFNGLDISPATCHCOMPUTEGROUPSIZEARBPROC) SMOL_GETGLPROC(glDispatchComputeGroupSizeARB); 
+    glDebugMessageControlARB = (PFNGLDEBUGMESSAGECONTROLARBPROC) SMOL_GETGLPROC(glDebugMessageControlARB); 
+    glDebugMessageInsertARB = (PFNGLDEBUGMESSAGEINSERTARBPROC) SMOL_GETGLPROC(glDebugMessageInsertARB); 
+    glDebugMessageCallbackARB = (PFNGLDEBUGMESSAGECALLBACKARBPROC) SMOL_GETGLPROC(glDebugMessageCallbackARB); 
+    glGetDebugMessageLogARB = (PFNGLGETDEBUGMESSAGELOGARBPROC) SMOL_GETGLPROC(glGetDebugMessageLogARB); 
+    glBlendEquationiARB = (PFNGLBLENDEQUATIONIARBPROC) SMOL_GETGLPROC(glBlendEquationiARB); 
+    glBlendEquationSeparateiARB = (PFNGLBLENDEQUATIONSEPARATEIARBPROC) SMOL_GETGLPROC(glBlendEquationSeparateiARB); 
+    glBlendFunciARB = (PFNGLBLENDFUNCIARBPROC) SMOL_GETGLPROC(glBlendFunciARB); 
+    glBlendFuncSeparateiARB = (PFNGLBLENDFUNCSEPARATEIARBPROC) SMOL_GETGLPROC(glBlendFuncSeparateiARB); 
+    glDrawArraysInstancedARB = (PFNGLDRAWARRAYSINSTANCEDARBPROC) SMOL_GETGLPROC(glDrawArraysInstancedARB); 
+    glDrawElementsInstancedARB = (PFNGLDRAWELEMENTSINSTANCEDARBPROC) SMOL_GETGLPROC(glDrawElementsInstancedARB); 
+    glProgramParameteriARB = (PFNGLPROGRAMPARAMETERIARBPROC) SMOL_GETGLPROC(glProgramParameteriARB); 
+    glFramebufferTextureARB = (PFNGLFRAMEBUFFERTEXTUREARBPROC) SMOL_GETGLPROC(glFramebufferTextureARB); 
+    glFramebufferTextureLayerARB = (PFNGLFRAMEBUFFERTEXTURELAYERARBPROC) SMOL_GETGLPROC(glFramebufferTextureLayerARB); 
+    glFramebufferTextureFaceARB = (PFNGLFRAMEBUFFERTEXTUREFACEARBPROC) SMOL_GETGLPROC(glFramebufferTextureFaceARB); 
+    glSpecializeShaderARB = (PFNGLSPECIALIZESHADERARBPROC) SMOL_GETGLPROC(glSpecializeShaderARB); 
+    glUniform1i64ARB = (PFNGLUNIFORM1I64ARBPROC) SMOL_GETGLPROC(glUniform1i64ARB); 
+    glUniform2i64ARB = (PFNGLUNIFORM2I64ARBPROC) SMOL_GETGLPROC(glUniform2i64ARB); 
+    glUniform3i64ARB = (PFNGLUNIFORM3I64ARBPROC) SMOL_GETGLPROC(glUniform3i64ARB); 
+    glUniform4i64ARB = (PFNGLUNIFORM4I64ARBPROC) SMOL_GETGLPROC(glUniform4i64ARB); 
+    glUniform1i64vARB = (PFNGLUNIFORM1I64VARBPROC) SMOL_GETGLPROC(glUniform1i64vARB); 
+    glUniform2i64vARB = (PFNGLUNIFORM2I64VARBPROC) SMOL_GETGLPROC(glUniform2i64vARB); 
+    glUniform3i64vARB = (PFNGLUNIFORM3I64VARBPROC) SMOL_GETGLPROC(glUniform3i64vARB); 
+    glUniform4i64vARB = (PFNGLUNIFORM4I64VARBPROC) SMOL_GETGLPROC(glUniform4i64vARB); 
+    glUniform1ui64ARB = (PFNGLUNIFORM1UI64ARBPROC) SMOL_GETGLPROC(glUniform1ui64ARB); 
+    glUniform2ui64ARB = (PFNGLUNIFORM2UI64ARBPROC) SMOL_GETGLPROC(glUniform2ui64ARB); 
+    glUniform3ui64ARB = (PFNGLUNIFORM3UI64ARBPROC) SMOL_GETGLPROC(glUniform3ui64ARB); 
+    glUniform4ui64ARB = (PFNGLUNIFORM4UI64ARBPROC) SMOL_GETGLPROC(glUniform4ui64ARB); 
+    glUniform1ui64vARB = (PFNGLUNIFORM1UI64VARBPROC) SMOL_GETGLPROC(glUniform1ui64vARB); 
+    glUniform2ui64vARB = (PFNGLUNIFORM2UI64VARBPROC) SMOL_GETGLPROC(glUniform2ui64vARB); 
+    glUniform3ui64vARB = (PFNGLUNIFORM3UI64VARBPROC) SMOL_GETGLPROC(glUniform3ui64vARB); 
+    glUniform4ui64vARB = (PFNGLUNIFORM4UI64VARBPROC) SMOL_GETGLPROC(glUniform4ui64vARB); 
+    glGetUniformi64vARB = (PFNGLGETUNIFORMI64VARBPROC) SMOL_GETGLPROC(glGetUniformi64vARB); 
+    glGetUniformui64vARB = (PFNGLGETUNIFORMUI64VARBPROC) SMOL_GETGLPROC(glGetUniformui64vARB); 
+    glGetnUniformi64vARB = (PFNGLGETNUNIFORMI64VARBPROC) SMOL_GETGLPROC(glGetnUniformi64vARB); 
+    glGetnUniformui64vARB = (PFNGLGETNUNIFORMUI64VARBPROC) SMOL_GETGLPROC(glGetnUniformui64vARB); 
+    glProgramUniform1i64ARB = (PFNGLPROGRAMUNIFORM1I64ARBPROC) SMOL_GETGLPROC(glProgramUniform1i64ARB); 
+    glProgramUniform2i64ARB = (PFNGLPROGRAMUNIFORM2I64ARBPROC) SMOL_GETGLPROC(glProgramUniform2i64ARB); 
+    glProgramUniform3i64ARB = (PFNGLPROGRAMUNIFORM3I64ARBPROC) SMOL_GETGLPROC(glProgramUniform3i64ARB); 
+    glProgramUniform4i64ARB = (PFNGLPROGRAMUNIFORM4I64ARBPROC) SMOL_GETGLPROC(glProgramUniform4i64ARB); 
+    glProgramUniform1i64vARB = (PFNGLPROGRAMUNIFORM1I64VARBPROC) SMOL_GETGLPROC(glProgramUniform1i64vARB); 
+    glProgramUniform2i64vARB = (PFNGLPROGRAMUNIFORM2I64VARBPROC) SMOL_GETGLPROC(glProgramUniform2i64vARB); 
+    glProgramUniform3i64vARB = (PFNGLPROGRAMUNIFORM3I64VARBPROC) SMOL_GETGLPROC(glProgramUniform3i64vARB); 
+    glProgramUniform4i64vARB = (PFNGLPROGRAMUNIFORM4I64VARBPROC) SMOL_GETGLPROC(glProgramUniform4i64vARB); 
+    glProgramUniform1ui64ARB = (PFNGLPROGRAMUNIFORM1UI64ARBPROC) SMOL_GETGLPROC(glProgramUniform1ui64ARB); 
+    glProgramUniform2ui64ARB = (PFNGLPROGRAMUNIFORM2UI64ARBPROC) SMOL_GETGLPROC(glProgramUniform2ui64ARB); 
+    glProgramUniform3ui64ARB = (PFNGLPROGRAMUNIFORM3UI64ARBPROC) SMOL_GETGLPROC(glProgramUniform3ui64ARB); 
+    glProgramUniform4ui64ARB = (PFNGLPROGRAMUNIFORM4UI64ARBPROC) SMOL_GETGLPROC(glProgramUniform4ui64ARB); 
+    glProgramUniform1ui64vARB = (PFNGLPROGRAMUNIFORM1UI64VARBPROC) SMOL_GETGLPROC(glProgramUniform1ui64vARB); 
+    glProgramUniform2ui64vARB = (PFNGLPROGRAMUNIFORM2UI64VARBPROC) SMOL_GETGLPROC(glProgramUniform2ui64vARB); 
+    glProgramUniform3ui64vARB = (PFNGLPROGRAMUNIFORM3UI64VARBPROC) SMOL_GETGLPROC(glProgramUniform3ui64vARB); 
+    glProgramUniform4ui64vARB = (PFNGLPROGRAMUNIFORM4UI64VARBPROC) SMOL_GETGLPROC(glProgramUniform4ui64vARB); 
+    glMultiDrawArraysIndirectCountARB = (PFNGLMULTIDRAWARRAYSINDIRECTCOUNTARBPROC) SMOL_GETGLPROC(glMultiDrawArraysIndirectCountARB); 
+    glMultiDrawElementsIndirectCountARB = (PFNGLMULTIDRAWELEMENTSINDIRECTCOUNTARBPROC) SMOL_GETGLPROC(glMultiDrawElementsIndirectCountARB); 
+    glVertexAttribDivisorARB = (PFNGLVERTEXATTRIBDIVISORARBPROC) SMOL_GETGLPROC(glVertexAttribDivisorARB); 
+    glMaxShaderCompilerThreadsARB = (PFNGLMAXSHADERCOMPILERTHREADSARBPROC) SMOL_GETGLPROC(glMaxShaderCompilerThreadsARB); 
+    glGetGraphicsResetStatusARB = (PFNGLGETGRAPHICSRESETSTATUSARBPROC) SMOL_GETGLPROC(glGetGraphicsResetStatusARB); 
+    glGetnTexImageARB = (PFNGLGETNTEXIMAGEARBPROC) SMOL_GETGLPROC(glGetnTexImageARB); 
+    glReadnPixelsARB = (PFNGLREADNPIXELSARBPROC) SMOL_GETGLPROC(glReadnPixelsARB); 
+    glGetnCompressedTexImageARB = (PFNGLGETNCOMPRESSEDTEXIMAGEARBPROC) SMOL_GETGLPROC(glGetnCompressedTexImageARB); 
+    glGetnUniformfvARB = (PFNGLGETNUNIFORMFVARBPROC) SMOL_GETGLPROC(glGetnUniformfvARB); 
+    glGetnUniformivARB = (PFNGLGETNUNIFORMIVARBPROC) SMOL_GETGLPROC(glGetnUniformivARB); 
+    glGetnUniformuivARB = (PFNGLGETNUNIFORMUIVARBPROC) SMOL_GETGLPROC(glGetnUniformuivARB); 
+    glGetnUniformdvARB = (PFNGLGETNUNIFORMDVARBPROC) SMOL_GETGLPROC(glGetnUniformdvARB); 
+    glFramebufferSampleLocationsfvARB = (PFNGLFRAMEBUFFERSAMPLELOCATIONSFVARBPROC) SMOL_GETGLPROC(glFramebufferSampleLocationsfvARB); 
+    glNamedFramebufferSampleLocationsfvARB = (PFNGLNAMEDFRAMEBUFFERSAMPLELOCATIONSFVARBPROC) SMOL_GETGLPROC(glNamedFramebufferSampleLocationsfvARB); 
+    glEvaluateDepthValuesARB = (PFNGLEVALUATEDEPTHVALUESARBPROC) SMOL_GETGLPROC(glEvaluateDepthValuesARB); 
+    glMinSampleShadingARB = (PFNGLMINSAMPLESHADINGARBPROC) SMOL_GETGLPROC(glMinSampleShadingARB); 
+    glNamedStringARB = (PFNGLNAMEDSTRINGARBPROC) SMOL_GETGLPROC(glNamedStringARB); 
+    glDeleteNamedStringARB = (PFNGLDELETENAMEDSTRINGARBPROC) SMOL_GETGLPROC(glDeleteNamedStringARB); 
+    glCompileShaderIncludeARB = (PFNGLCOMPILESHADERINCLUDEARBPROC) SMOL_GETGLPROC(glCompileShaderIncludeARB); 
+    glIsNamedStringARB = (PFNGLISNAMEDSTRINGARBPROC) SMOL_GETGLPROC(glIsNamedStringARB); 
+    glGetNamedStringARB = (PFNGLGETNAMEDSTRINGARBPROC) SMOL_GETGLPROC(glGetNamedStringARB); 
+    glGetNamedStringivARB = (PFNGLGETNAMEDSTRINGIVARBPROC) SMOL_GETGLPROC(glGetNamedStringivARB); 
+    glBufferPageCommitmentARB = (PFNGLBUFFERPAGECOMMITMENTARBPROC) SMOL_GETGLPROC(glBufferPageCommitmentARB); 
+    glNamedBufferPageCommitmentEXT = (PFNGLNAMEDBUFFERPAGECOMMITMENTEXTPROC) SMOL_GETGLPROC(glNamedBufferPageCommitmentEXT); 
+    glNamedBufferPageCommitmentARB = (PFNGLNAMEDBUFFERPAGECOMMITMENTARBPROC) SMOL_GETGLPROC(glNamedBufferPageCommitmentARB); 
+    glTexPageCommitmentARB = (PFNGLTEXPAGECOMMITMENTARBPROC) SMOL_GETGLPROC(glTexPageCommitmentARB); 
+    glTexBufferARB = (PFNGLTEXBUFFERARBPROC) SMOL_GETGLPROC(glTexBufferARB); 
+    glDepthRangeArraydvNV = (PFNGLDEPTHRANGEARRAYDVNVPROC) SMOL_GETGLPROC(glDepthRangeArraydvNV); 
+    glDepthRangeIndexeddNV = (PFNGLDEPTHRANGEINDEXEDDNVPROC) SMOL_GETGLPROC(glDepthRangeIndexeddNV); 
+    glBlendBarrierKHR = (PFNGLBLENDBARRIERKHRPROC) SMOL_GETGLPROC(glBlendBarrierKHR); 
+    glMaxShaderCompilerThreadsKHR = (PFNGLMAXSHADERCOMPILERTHREADSKHRPROC) SMOL_GETGLPROC(glMaxShaderCompilerThreadsKHR); 
+    glRenderbufferStorageMultisampleAdvancedAMD = (PFNGLRENDERBUFFERSTORAGEMULTISAMPLEADVANCEDAMDPROC) SMOL_GETGLPROC(glRenderbufferStorageMultisampleAdvancedAMD); 
+    glNamedRenderbufferStorageMultisampleAdvancedAMD = (PFNGLNAMEDRENDERBUFFERSTORAGEMULTISAMPLEADVANCEDAMDPROC) SMOL_GETGLPROC(glNamedRenderbufferStorageMultisampleAdvancedAMD); 
+    glGetPerfMonitorGroupsAMD = (PFNGLGETPERFMONITORGROUPSAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorGroupsAMD); 
+    glGetPerfMonitorCountersAMD = (PFNGLGETPERFMONITORCOUNTERSAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorCountersAMD); 
+    glGetPerfMonitorGroupStringAMD = (PFNGLGETPERFMONITORGROUPSTRINGAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorGroupStringAMD); 
+    glGetPerfMonitorCounterStringAMD = (PFNGLGETPERFMONITORCOUNTERSTRINGAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorCounterStringAMD); 
+    glGetPerfMonitorCounterInfoAMD = (PFNGLGETPERFMONITORCOUNTERINFOAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorCounterInfoAMD); 
+    glGenPerfMonitorsAMD = (PFNGLGENPERFMONITORSAMDPROC) SMOL_GETGLPROC(glGenPerfMonitorsAMD); 
+    glDeletePerfMonitorsAMD = (PFNGLDELETEPERFMONITORSAMDPROC) SMOL_GETGLPROC(glDeletePerfMonitorsAMD); 
+    glSelectPerfMonitorCountersAMD = (PFNGLSELECTPERFMONITORCOUNTERSAMDPROC) SMOL_GETGLPROC(glSelectPerfMonitorCountersAMD); 
+    glBeginPerfMonitorAMD = (PFNGLBEGINPERFMONITORAMDPROC) SMOL_GETGLPROC(glBeginPerfMonitorAMD); 
+    glEndPerfMonitorAMD = (PFNGLENDPERFMONITORAMDPROC) SMOL_GETGLPROC(glEndPerfMonitorAMD); 
+    glGetPerfMonitorCounterDataAMD = (PFNGLGETPERFMONITORCOUNTERDATAAMDPROC) SMOL_GETGLPROC(glGetPerfMonitorCounterDataAMD); 
+    glEGLImageTargetTexStorageEXT = (PFNGLEGLIMAGETARGETTEXSTORAGEEXTPROC) SMOL_GETGLPROC(glEGLImageTargetTexStorageEXT); 
+    glEGLImageTargetTextureStorageEXT = (PFNGLEGLIMAGETARGETTEXTURESTORAGEEXTPROC) SMOL_GETGLPROC(glEGLImageTargetTextureStorageEXT); 
+    glLabelObjectEXT = (PFNGLLABELOBJECTEXTPROC) SMOL_GETGLPROC(glLabelObjectEXT); 
+    glGetObjectLabelEXT = (PFNGLGETOBJECTLABELEXTPROC) SMOL_GETGLPROC(glGetObjectLabelEXT); 
+    glInsertEventMarkerEXT = (PFNGLINSERTEVENTMARKEREXTPROC) SMOL_GETGLPROC(glInsertEventMarkerEXT); 
+    glPushGroupMarkerEXT = (PFNGLPUSHGROUPMARKEREXTPROC) SMOL_GETGLPROC(glPushGroupMarkerEXT); 
+    glPopGroupMarkerEXT = (PFNGLPOPGROUPMARKEREXTPROC) SMOL_GETGLPROC(glPopGroupMarkerEXT); 
+    glMatrixLoadfEXT = (PFNGLMATRIXLOADFEXTPROC) SMOL_GETGLPROC(glMatrixLoadfEXT); 
+    glMatrixLoaddEXT = (PFNGLMATRIXLOADDEXTPROC) SMOL_GETGLPROC(glMatrixLoaddEXT); 
+    glMatrixMultfEXT = (PFNGLMATRIXMULTFEXTPROC) SMOL_GETGLPROC(glMatrixMultfEXT); 
+    glMatrixMultdEXT = (PFNGLMATRIXMULTDEXTPROC) SMOL_GETGLPROC(glMatrixMultdEXT); 
+    glMatrixLoadIdentityEXT = (PFNGLMATRIXLOADIDENTITYEXTPROC) SMOL_GETGLPROC(glMatrixLoadIdentityEXT); 
+    glMatrixRotatefEXT = (PFNGLMATRIXROTATEFEXTPROC) SMOL_GETGLPROC(glMatrixRotatefEXT); 
+    glMatrixRotatedEXT = (PFNGLMATRIXROTATEDEXTPROC) SMOL_GETGLPROC(glMatrixRotatedEXT); 
+    glMatrixScalefEXT = (PFNGLMATRIXSCALEFEXTPROC) SMOL_GETGLPROC(glMatrixScalefEXT); 
+    glMatrixScaledEXT = (PFNGLMATRIXSCALEDEXTPROC) SMOL_GETGLPROC(glMatrixScaledEXT); 
+    glMatrixTranslatefEXT = (PFNGLMATRIXTRANSLATEFEXTPROC) SMOL_GETGLPROC(glMatrixTranslatefEXT); 
+    glMatrixTranslatedEXT = (PFNGLMATRIXTRANSLATEDEXTPROC) SMOL_GETGLPROC(glMatrixTranslatedEXT); 
+    glMatrixFrustumEXT = (PFNGLMATRIXFRUSTUMEXTPROC) SMOL_GETGLPROC(glMatrixFrustumEXT); 
+    glMatrixOrthoEXT = (PFNGLMATRIXORTHOEXTPROC) SMOL_GETGLPROC(glMatrixOrthoEXT); 
+    glMatrixPopEXT = (PFNGLMATRIXPOPEXTPROC) SMOL_GETGLPROC(glMatrixPopEXT); 
+    glMatrixPushEXT = (PFNGLMATRIXPUSHEXTPROC) SMOL_GETGLPROC(glMatrixPushEXT); 
+    glClientAttribDefaultEXT = (PFNGLCLIENTATTRIBDEFAULTEXTPROC) SMOL_GETGLPROC(glClientAttribDefaultEXT); 
+    glPushClientAttribDefaultEXT = (PFNGLPUSHCLIENTATTRIBDEFAULTEXTPROC) SMOL_GETGLPROC(glPushClientAttribDefaultEXT); 
+    glTextureParameterfEXT = (PFNGLTEXTUREPARAMETERFEXTPROC) SMOL_GETGLPROC(glTextureParameterfEXT); 
+    glTextureParameterfvEXT = (PFNGLTEXTUREPARAMETERFVEXTPROC) SMOL_GETGLPROC(glTextureParameterfvEXT); 
+    glTextureParameteriEXT = (PFNGLTEXTUREPARAMETERIEXTPROC) SMOL_GETGLPROC(glTextureParameteriEXT); 
+    glTextureParameterivEXT = (PFNGLTEXTUREPARAMETERIVEXTPROC) SMOL_GETGLPROC(glTextureParameterivEXT); 
+    glTextureImage1DEXT = (PFNGLTEXTUREIMAGE1DEXTPROC) SMOL_GETGLPROC(glTextureImage1DEXT); 
+    glTextureImage2DEXT = (PFNGLTEXTUREIMAGE2DEXTPROC) SMOL_GETGLPROC(glTextureImage2DEXT); 
+    glTextureSubImage1DEXT = (PFNGLTEXTURESUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glTextureSubImage1DEXT); 
+    glTextureSubImage2DEXT = (PFNGLTEXTURESUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glTextureSubImage2DEXT); 
+    glCopyTextureImage1DEXT = (PFNGLCOPYTEXTUREIMAGE1DEXTPROC) SMOL_GETGLPROC(glCopyTextureImage1DEXT); 
+    glCopyTextureImage2DEXT = (PFNGLCOPYTEXTUREIMAGE2DEXTPROC) SMOL_GETGLPROC(glCopyTextureImage2DEXT); 
+    glCopyTextureSubImage1DEXT = (PFNGLCOPYTEXTURESUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glCopyTextureSubImage1DEXT); 
+    glCopyTextureSubImage2DEXT = (PFNGLCOPYTEXTURESUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glCopyTextureSubImage2DEXT); 
+    glGetTextureImageEXT = (PFNGLGETTEXTUREIMAGEEXTPROC) SMOL_GETGLPROC(glGetTextureImageEXT); 
+    glGetTextureParameterfvEXT = (PFNGLGETTEXTUREPARAMETERFVEXTPROC) SMOL_GETGLPROC(glGetTextureParameterfvEXT); 
+    glGetTextureParameterivEXT = (PFNGLGETTEXTUREPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetTextureParameterivEXT); 
+    glGetTextureLevelParameterfvEXT = (PFNGLGETTEXTURELEVELPARAMETERFVEXTPROC) SMOL_GETGLPROC(glGetTextureLevelParameterfvEXT); 
+    glGetTextureLevelParameterivEXT = (PFNGLGETTEXTURELEVELPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetTextureLevelParameterivEXT); 
+    glTextureImage3DEXT = (PFNGLTEXTUREIMAGE3DEXTPROC) SMOL_GETGLPROC(glTextureImage3DEXT); 
+    glTextureSubImage3DEXT = (PFNGLTEXTURESUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glTextureSubImage3DEXT); 
+    glCopyTextureSubImage3DEXT = (PFNGLCOPYTEXTURESUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glCopyTextureSubImage3DEXT); 
+    glBindMultiTextureEXT = (PFNGLBINDMULTITEXTUREEXTPROC) SMOL_GETGLPROC(glBindMultiTextureEXT); 
+    glMultiTexCoordPointerEXT = (PFNGLMULTITEXCOORDPOINTEREXTPROC) SMOL_GETGLPROC(glMultiTexCoordPointerEXT); 
+    glMultiTexEnvfEXT = (PFNGLMULTITEXENVFEXTPROC) SMOL_GETGLPROC(glMultiTexEnvfEXT); 
+    glMultiTexEnvfvEXT = (PFNGLMULTITEXENVFVEXTPROC) SMOL_GETGLPROC(glMultiTexEnvfvEXT); 
+    glMultiTexEnviEXT = (PFNGLMULTITEXENVIEXTPROC) SMOL_GETGLPROC(glMultiTexEnviEXT); 
+    glMultiTexEnvivEXT = (PFNGLMULTITEXENVIVEXTPROC) SMOL_GETGLPROC(glMultiTexEnvivEXT); 
+    glMultiTexGendEXT = (PFNGLMULTITEXGENDEXTPROC) SMOL_GETGLPROC(glMultiTexGendEXT); 
+    glMultiTexGendvEXT = (PFNGLMULTITEXGENDVEXTPROC) SMOL_GETGLPROC(glMultiTexGendvEXT); 
+    glMultiTexGenfEXT = (PFNGLMULTITEXGENFEXTPROC) SMOL_GETGLPROC(glMultiTexGenfEXT); 
+    glMultiTexGenfvEXT = (PFNGLMULTITEXGENFVEXTPROC) SMOL_GETGLPROC(glMultiTexGenfvEXT); 
+    glMultiTexGeniEXT = (PFNGLMULTITEXGENIEXTPROC) SMOL_GETGLPROC(glMultiTexGeniEXT); 
+    glMultiTexGenivEXT = (PFNGLMULTITEXGENIVEXTPROC) SMOL_GETGLPROC(glMultiTexGenivEXT); 
+    glGetMultiTexEnvfvEXT = (PFNGLGETMULTITEXENVFVEXTPROC) SMOL_GETGLPROC(glGetMultiTexEnvfvEXT); 
+    glGetMultiTexEnvivEXT = (PFNGLGETMULTITEXENVIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexEnvivEXT); 
+    glGetMultiTexGendvEXT = (PFNGLGETMULTITEXGENDVEXTPROC) SMOL_GETGLPROC(glGetMultiTexGendvEXT); 
+    glGetMultiTexGenfvEXT = (PFNGLGETMULTITEXGENFVEXTPROC) SMOL_GETGLPROC(glGetMultiTexGenfvEXT); 
+    glGetMultiTexGenivEXT = (PFNGLGETMULTITEXGENIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexGenivEXT); 
+    glMultiTexParameteriEXT = (PFNGLMULTITEXPARAMETERIEXTPROC) SMOL_GETGLPROC(glMultiTexParameteriEXT); 
+    glMultiTexParameterivEXT = (PFNGLMULTITEXPARAMETERIVEXTPROC) SMOL_GETGLPROC(glMultiTexParameterivEXT); 
+    glMultiTexParameterfEXT = (PFNGLMULTITEXPARAMETERFEXTPROC) SMOL_GETGLPROC(glMultiTexParameterfEXT); 
+    glMultiTexParameterfvEXT = (PFNGLMULTITEXPARAMETERFVEXTPROC) SMOL_GETGLPROC(glMultiTexParameterfvEXT); 
+    glMultiTexImage1DEXT = (PFNGLMULTITEXIMAGE1DEXTPROC) SMOL_GETGLPROC(glMultiTexImage1DEXT); 
+    glMultiTexImage2DEXT = (PFNGLMULTITEXIMAGE2DEXTPROC) SMOL_GETGLPROC(glMultiTexImage2DEXT); 
+    glMultiTexSubImage1DEXT = (PFNGLMULTITEXSUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glMultiTexSubImage1DEXT); 
+    glMultiTexSubImage2DEXT = (PFNGLMULTITEXSUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glMultiTexSubImage2DEXT); 
+    glCopyMultiTexImage1DEXT = (PFNGLCOPYMULTITEXIMAGE1DEXTPROC) SMOL_GETGLPROC(glCopyMultiTexImage1DEXT); 
+    glCopyMultiTexImage2DEXT = (PFNGLCOPYMULTITEXIMAGE2DEXTPROC) SMOL_GETGLPROC(glCopyMultiTexImage2DEXT); 
+    glCopyMultiTexSubImage1DEXT = (PFNGLCOPYMULTITEXSUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glCopyMultiTexSubImage1DEXT); 
+    glCopyMultiTexSubImage2DEXT = (PFNGLCOPYMULTITEXSUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glCopyMultiTexSubImage2DEXT); 
+    glGetMultiTexImageEXT = (PFNGLGETMULTITEXIMAGEEXTPROC) SMOL_GETGLPROC(glGetMultiTexImageEXT); 
+    glGetMultiTexParameterfvEXT = (PFNGLGETMULTITEXPARAMETERFVEXTPROC) SMOL_GETGLPROC(glGetMultiTexParameterfvEXT); 
+    glGetMultiTexParameterivEXT = (PFNGLGETMULTITEXPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexParameterivEXT); 
+    glGetMultiTexLevelParameterfvEXT = (PFNGLGETMULTITEXLEVELPARAMETERFVEXTPROC) SMOL_GETGLPROC(glGetMultiTexLevelParameterfvEXT); 
+    glGetMultiTexLevelParameterivEXT = (PFNGLGETMULTITEXLEVELPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexLevelParameterivEXT); 
+    glMultiTexImage3DEXT = (PFNGLMULTITEXIMAGE3DEXTPROC) SMOL_GETGLPROC(glMultiTexImage3DEXT); 
+    glMultiTexSubImage3DEXT = (PFNGLMULTITEXSUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glMultiTexSubImage3DEXT); 
+    glCopyMultiTexSubImage3DEXT = (PFNGLCOPYMULTITEXSUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glCopyMultiTexSubImage3DEXT); 
+    glEnableClientStateIndexedEXT = (PFNGLENABLECLIENTSTATEINDEXEDEXTPROC) SMOL_GETGLPROC(glEnableClientStateIndexedEXT); 
+    glDisableClientStateIndexedEXT = (PFNGLDISABLECLIENTSTATEINDEXEDEXTPROC) SMOL_GETGLPROC(glDisableClientStateIndexedEXT); 
+    glGetFloatIndexedvEXT = (PFNGLGETFLOATINDEXEDVEXTPROC) SMOL_GETGLPROC(glGetFloatIndexedvEXT); 
+    glGetDoubleIndexedvEXT = (PFNGLGETDOUBLEINDEXEDVEXTPROC) SMOL_GETGLPROC(glGetDoubleIndexedvEXT); 
+    glGetPointerIndexedvEXT = (PFNGLGETPOINTERINDEXEDVEXTPROC) SMOL_GETGLPROC(glGetPointerIndexedvEXT); 
+    glEnableIndexedEXT = (PFNGLENABLEINDEXEDEXTPROC) SMOL_GETGLPROC(glEnableIndexedEXT); 
+    glDisableIndexedEXT = (PFNGLDISABLEINDEXEDEXTPROC) SMOL_GETGLPROC(glDisableIndexedEXT); 
+    glIsEnabledIndexedEXT = (PFNGLISENABLEDINDEXEDEXTPROC) SMOL_GETGLPROC(glIsEnabledIndexedEXT); 
+    glGetIntegerIndexedvEXT = (PFNGLGETINTEGERINDEXEDVEXTPROC) SMOL_GETGLPROC(glGetIntegerIndexedvEXT); 
+    glGetBooleanIndexedvEXT = (PFNGLGETBOOLEANINDEXEDVEXTPROC) SMOL_GETGLPROC(glGetBooleanIndexedvEXT); 
+    glCompressedTextureImage3DEXT = (PFNGLCOMPRESSEDTEXTUREIMAGE3DEXTPROC) SMOL_GETGLPROC(glCompressedTextureImage3DEXT); 
+    glCompressedTextureImage2DEXT = (PFNGLCOMPRESSEDTEXTUREIMAGE2DEXTPROC) SMOL_GETGLPROC(glCompressedTextureImage2DEXT); 
+    glCompressedTextureImage1DEXT = (PFNGLCOMPRESSEDTEXTUREIMAGE1DEXTPROC) SMOL_GETGLPROC(glCompressedTextureImage1DEXT); 
+    glCompressedTextureSubImage3DEXT = (PFNGLCOMPRESSEDTEXTURESUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glCompressedTextureSubImage3DEXT); 
+    glCompressedTextureSubImage2DEXT = (PFNGLCOMPRESSEDTEXTURESUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glCompressedTextureSubImage2DEXT); 
+    glCompressedTextureSubImage1DEXT = (PFNGLCOMPRESSEDTEXTURESUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glCompressedTextureSubImage1DEXT); 
+    glGetCompressedTextureImageEXT = (PFNGLGETCOMPRESSEDTEXTUREIMAGEEXTPROC) SMOL_GETGLPROC(glGetCompressedTextureImageEXT); 
+    glCompressedMultiTexImage3DEXT = (PFNGLCOMPRESSEDMULTITEXIMAGE3DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexImage3DEXT); 
+    glCompressedMultiTexImage2DEXT = (PFNGLCOMPRESSEDMULTITEXIMAGE2DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexImage2DEXT); 
+    glCompressedMultiTexImage1DEXT = (PFNGLCOMPRESSEDMULTITEXIMAGE1DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexImage1DEXT); 
+    glCompressedMultiTexSubImage3DEXT = (PFNGLCOMPRESSEDMULTITEXSUBIMAGE3DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexSubImage3DEXT); 
+    glCompressedMultiTexSubImage2DEXT = (PFNGLCOMPRESSEDMULTITEXSUBIMAGE2DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexSubImage2DEXT); 
+    glCompressedMultiTexSubImage1DEXT = (PFNGLCOMPRESSEDMULTITEXSUBIMAGE1DEXTPROC) SMOL_GETGLPROC(glCompressedMultiTexSubImage1DEXT); 
+    glGetCompressedMultiTexImageEXT = (PFNGLGETCOMPRESSEDMULTITEXIMAGEEXTPROC) SMOL_GETGLPROC(glGetCompressedMultiTexImageEXT); 
+    glMatrixLoadTransposefEXT = (PFNGLMATRIXLOADTRANSPOSEFEXTPROC) SMOL_GETGLPROC(glMatrixLoadTransposefEXT); 
+    glMatrixLoadTransposedEXT = (PFNGLMATRIXLOADTRANSPOSEDEXTPROC) SMOL_GETGLPROC(glMatrixLoadTransposedEXT); 
+    glMatrixMultTransposefEXT = (PFNGLMATRIXMULTTRANSPOSEFEXTPROC) SMOL_GETGLPROC(glMatrixMultTransposefEXT); 
+    glMatrixMultTransposedEXT = (PFNGLMATRIXMULTTRANSPOSEDEXTPROC) SMOL_GETGLPROC(glMatrixMultTransposedEXT); 
+    glNamedBufferDataEXT = (PFNGLNAMEDBUFFERDATAEXTPROC) SMOL_GETGLPROC(glNamedBufferDataEXT); 
+    glNamedBufferSubDataEXT = (PFNGLNAMEDBUFFERSUBDATAEXTPROC) SMOL_GETGLPROC(glNamedBufferSubDataEXT); 
+    glMapNamedBufferEXT = (PFNGLMAPNAMEDBUFFEREXTPROC) SMOL_GETGLPROC(glMapNamedBufferEXT); 
+    glUnmapNamedBufferEXT = (PFNGLUNMAPNAMEDBUFFEREXTPROC) SMOL_GETGLPROC(glUnmapNamedBufferEXT); 
+    glGetNamedBufferParameterivEXT = (PFNGLGETNAMEDBUFFERPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetNamedBufferParameterivEXT); 
+    glGetNamedBufferPointervEXT = (PFNGLGETNAMEDBUFFERPOINTERVEXTPROC) SMOL_GETGLPROC(glGetNamedBufferPointervEXT); 
+    glGetNamedBufferSubDataEXT = (PFNGLGETNAMEDBUFFERSUBDATAEXTPROC) SMOL_GETGLPROC(glGetNamedBufferSubDataEXT); 
+    glProgramUniform1fEXT = (PFNGLPROGRAMUNIFORM1FEXTPROC) SMOL_GETGLPROC(glProgramUniform1fEXT); 
+    glProgramUniform2fEXT = (PFNGLPROGRAMUNIFORM2FEXTPROC) SMOL_GETGLPROC(glProgramUniform2fEXT); 
+    glProgramUniform3fEXT = (PFNGLPROGRAMUNIFORM3FEXTPROC) SMOL_GETGLPROC(glProgramUniform3fEXT); 
+    glProgramUniform4fEXT = (PFNGLPROGRAMUNIFORM4FEXTPROC) SMOL_GETGLPROC(glProgramUniform4fEXT); 
+    glProgramUniform1iEXT = (PFNGLPROGRAMUNIFORM1IEXTPROC) SMOL_GETGLPROC(glProgramUniform1iEXT); 
+    glProgramUniform2iEXT = (PFNGLPROGRAMUNIFORM2IEXTPROC) SMOL_GETGLPROC(glProgramUniform2iEXT); 
+    glProgramUniform3iEXT = (PFNGLPROGRAMUNIFORM3IEXTPROC) SMOL_GETGLPROC(glProgramUniform3iEXT); 
+    glProgramUniform4iEXT = (PFNGLPROGRAMUNIFORM4IEXTPROC) SMOL_GETGLPROC(glProgramUniform4iEXT); 
+    glProgramUniform1fvEXT = (PFNGLPROGRAMUNIFORM1FVEXTPROC) SMOL_GETGLPROC(glProgramUniform1fvEXT); 
+    glProgramUniform2fvEXT = (PFNGLPROGRAMUNIFORM2FVEXTPROC) SMOL_GETGLPROC(glProgramUniform2fvEXT); 
+    glProgramUniform3fvEXT = (PFNGLPROGRAMUNIFORM3FVEXTPROC) SMOL_GETGLPROC(glProgramUniform3fvEXT); 
+    glProgramUniform4fvEXT = (PFNGLPROGRAMUNIFORM4FVEXTPROC) SMOL_GETGLPROC(glProgramUniform4fvEXT); 
+    glProgramUniform1ivEXT = (PFNGLPROGRAMUNIFORM1IVEXTPROC) SMOL_GETGLPROC(glProgramUniform1ivEXT); 
+    glProgramUniform2ivEXT = (PFNGLPROGRAMUNIFORM2IVEXTPROC) SMOL_GETGLPROC(glProgramUniform2ivEXT); 
+    glProgramUniform3ivEXT = (PFNGLPROGRAMUNIFORM3IVEXTPROC) SMOL_GETGLPROC(glProgramUniform3ivEXT); 
+    glProgramUniform4ivEXT = (PFNGLPROGRAMUNIFORM4IVEXTPROC) SMOL_GETGLPROC(glProgramUniform4ivEXT); 
+    glProgramUniformMatrix2fvEXT = (PFNGLPROGRAMUNIFORMMATRIX2FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2fvEXT); 
+    glProgramUniformMatrix3fvEXT = (PFNGLPROGRAMUNIFORMMATRIX3FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3fvEXT); 
+    glProgramUniformMatrix4fvEXT = (PFNGLPROGRAMUNIFORMMATRIX4FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4fvEXT); 
+    glProgramUniformMatrix2x3fvEXT = (PFNGLPROGRAMUNIFORMMATRIX2X3FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x3fvEXT); 
+    glProgramUniformMatrix3x2fvEXT = (PFNGLPROGRAMUNIFORMMATRIX3X2FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x2fvEXT); 
+    glProgramUniformMatrix2x4fvEXT = (PFNGLPROGRAMUNIFORMMATRIX2X4FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x4fvEXT); 
+    glProgramUniformMatrix4x2fvEXT = (PFNGLPROGRAMUNIFORMMATRIX4X2FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x2fvEXT); 
+    glProgramUniformMatrix3x4fvEXT = (PFNGLPROGRAMUNIFORMMATRIX3X4FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x4fvEXT); 
+    glProgramUniformMatrix4x3fvEXT = (PFNGLPROGRAMUNIFORMMATRIX4X3FVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x3fvEXT); 
+    glTextureBufferEXT = (PFNGLTEXTUREBUFFEREXTPROC) SMOL_GETGLPROC(glTextureBufferEXT); 
+    glMultiTexBufferEXT = (PFNGLMULTITEXBUFFEREXTPROC) SMOL_GETGLPROC(glMultiTexBufferEXT); 
+    glTextureParameterIivEXT = (PFNGLTEXTUREPARAMETERIIVEXTPROC) SMOL_GETGLPROC(glTextureParameterIivEXT); 
+    glTextureParameterIuivEXT = (PFNGLTEXTUREPARAMETERIUIVEXTPROC) SMOL_GETGLPROC(glTextureParameterIuivEXT); 
+    glGetTextureParameterIivEXT = (PFNGLGETTEXTUREPARAMETERIIVEXTPROC) SMOL_GETGLPROC(glGetTextureParameterIivEXT); 
+    glGetTextureParameterIuivEXT = (PFNGLGETTEXTUREPARAMETERIUIVEXTPROC) SMOL_GETGLPROC(glGetTextureParameterIuivEXT); 
+    glMultiTexParameterIivEXT = (PFNGLMULTITEXPARAMETERIIVEXTPROC) SMOL_GETGLPROC(glMultiTexParameterIivEXT); 
+    glMultiTexParameterIuivEXT = (PFNGLMULTITEXPARAMETERIUIVEXTPROC) SMOL_GETGLPROC(glMultiTexParameterIuivEXT); 
+    glGetMultiTexParameterIivEXT = (PFNGLGETMULTITEXPARAMETERIIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexParameterIivEXT); 
+    glGetMultiTexParameterIuivEXT = (PFNGLGETMULTITEXPARAMETERIUIVEXTPROC) SMOL_GETGLPROC(glGetMultiTexParameterIuivEXT); 
+    glProgramUniform1uiEXT = (PFNGLPROGRAMUNIFORM1UIEXTPROC) SMOL_GETGLPROC(glProgramUniform1uiEXT); 
+    glProgramUniform2uiEXT = (PFNGLPROGRAMUNIFORM2UIEXTPROC) SMOL_GETGLPROC(glProgramUniform2uiEXT); 
+    glProgramUniform3uiEXT = (PFNGLPROGRAMUNIFORM3UIEXTPROC) SMOL_GETGLPROC(glProgramUniform3uiEXT); 
+    glProgramUniform4uiEXT = (PFNGLPROGRAMUNIFORM4UIEXTPROC) SMOL_GETGLPROC(glProgramUniform4uiEXT); 
+    glProgramUniform1uivEXT = (PFNGLPROGRAMUNIFORM1UIVEXTPROC) SMOL_GETGLPROC(glProgramUniform1uivEXT); 
+    glProgramUniform2uivEXT = (PFNGLPROGRAMUNIFORM2UIVEXTPROC) SMOL_GETGLPROC(glProgramUniform2uivEXT); 
+    glProgramUniform3uivEXT = (PFNGLPROGRAMUNIFORM3UIVEXTPROC) SMOL_GETGLPROC(glProgramUniform3uivEXT); 
+    glProgramUniform4uivEXT = (PFNGLPROGRAMUNIFORM4UIVEXTPROC) SMOL_GETGLPROC(glProgramUniform4uivEXT); 
+    glNamedProgramLocalParameters4fvEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERS4FVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameters4fvEXT); 
+    glNamedProgramLocalParameterI4iEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERI4IEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameterI4iEXT); 
+    glNamedProgramLocalParameterI4ivEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERI4IVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameterI4ivEXT); 
+    glNamedProgramLocalParametersI4ivEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERSI4IVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParametersI4ivEXT); 
+    glNamedProgramLocalParameterI4uiEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERI4UIEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameterI4uiEXT); 
+    glNamedProgramLocalParameterI4uivEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERI4UIVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameterI4uivEXT); 
+    glNamedProgramLocalParametersI4uivEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETERSI4UIVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParametersI4uivEXT); 
+    glGetNamedProgramLocalParameterIivEXT = (PFNGLGETNAMEDPROGRAMLOCALPARAMETERIIVEXTPROC) SMOL_GETGLPROC(glGetNamedProgramLocalParameterIivEXT); 
+    glGetNamedProgramLocalParameterIuivEXT = (PFNGLGETNAMEDPROGRAMLOCALPARAMETERIUIVEXTPROC) SMOL_GETGLPROC(glGetNamedProgramLocalParameterIuivEXT); 
+    glEnableClientStateiEXT = (PFNGLENABLECLIENTSTATEIEXTPROC) SMOL_GETGLPROC(glEnableClientStateiEXT); 
+    glDisableClientStateiEXT = (PFNGLDISABLECLIENTSTATEIEXTPROC) SMOL_GETGLPROC(glDisableClientStateiEXT); 
+    glGetFloati_vEXT = (PFNGLGETFLOATI_VEXTPROC) SMOL_GETGLPROC(glGetFloati_vEXT); 
+    glGetDoublei_vEXT = (PFNGLGETDOUBLEI_VEXTPROC) SMOL_GETGLPROC(glGetDoublei_vEXT); 
+    glGetPointeri_vEXT = (PFNGLGETPOINTERI_VEXTPROC) SMOL_GETGLPROC(glGetPointeri_vEXT); 
+    glNamedProgramStringEXT = (PFNGLNAMEDPROGRAMSTRINGEXTPROC) SMOL_GETGLPROC(glNamedProgramStringEXT); 
+    glNamedProgramLocalParameter4dEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETER4DEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameter4dEXT); 
+    glNamedProgramLocalParameter4dvEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETER4DVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameter4dvEXT); 
+    glNamedProgramLocalParameter4fEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETER4FEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameter4fEXT); 
+    glNamedProgramLocalParameter4fvEXT = (PFNGLNAMEDPROGRAMLOCALPARAMETER4FVEXTPROC) SMOL_GETGLPROC(glNamedProgramLocalParameter4fvEXT); 
+    glGetNamedProgramLocalParameterdvEXT = (PFNGLGETNAMEDPROGRAMLOCALPARAMETERDVEXTPROC) SMOL_GETGLPROC(glGetNamedProgramLocalParameterdvEXT); 
+    glGetNamedProgramLocalParameterfvEXT = (PFNGLGETNAMEDPROGRAMLOCALPARAMETERFVEXTPROC) SMOL_GETGLPROC(glGetNamedProgramLocalParameterfvEXT); 
+    glGetNamedProgramivEXT = (PFNGLGETNAMEDPROGRAMIVEXTPROC) SMOL_GETGLPROC(glGetNamedProgramivEXT); 
+    glGetNamedProgramStringEXT = (PFNGLGETNAMEDPROGRAMSTRINGEXTPROC) SMOL_GETGLPROC(glGetNamedProgramStringEXT); 
+    glNamedRenderbufferStorageEXT = (PFNGLNAMEDRENDERBUFFERSTORAGEEXTPROC) SMOL_GETGLPROC(glNamedRenderbufferStorageEXT); 
+    glGetNamedRenderbufferParameterivEXT = (PFNGLGETNAMEDRENDERBUFFERPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetNamedRenderbufferParameterivEXT); 
+    glNamedRenderbufferStorageMultisampleEXT = (PFNGLNAMEDRENDERBUFFERSTORAGEMULTISAMPLEEXTPROC) SMOL_GETGLPROC(glNamedRenderbufferStorageMultisampleEXT); 
+    glNamedRenderbufferStorageMultisampleCoverageEXT = (PFNGLNAMEDRENDERBUFFERSTORAGEMULTISAMPLECOVERAGEEXTPROC) SMOL_GETGLPROC(glNamedRenderbufferStorageMultisampleCoverageEXT); 
+    glCheckNamedFramebufferStatusEXT = (PFNGLCHECKNAMEDFRAMEBUFFERSTATUSEXTPROC) SMOL_GETGLPROC(glCheckNamedFramebufferStatusEXT); 
+    glNamedFramebufferTexture1DEXT = (PFNGLNAMEDFRAMEBUFFERTEXTURE1DEXTPROC) SMOL_GETGLPROC(glNamedFramebufferTexture1DEXT); 
+    glNamedFramebufferTexture2DEXT = (PFNGLNAMEDFRAMEBUFFERTEXTURE2DEXTPROC) SMOL_GETGLPROC(glNamedFramebufferTexture2DEXT); 
+    glNamedFramebufferTexture3DEXT = (PFNGLNAMEDFRAMEBUFFERTEXTURE3DEXTPROC) SMOL_GETGLPROC(glNamedFramebufferTexture3DEXT); 
+    glNamedFramebufferRenderbufferEXT = (PFNGLNAMEDFRAMEBUFFERRENDERBUFFEREXTPROC) SMOL_GETGLPROC(glNamedFramebufferRenderbufferEXT); 
+    glGetNamedFramebufferAttachmentParameterivEXT = (PFNGLGETNAMEDFRAMEBUFFERATTACHMENTPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetNamedFramebufferAttachmentParameterivEXT); 
+    glGenerateTextureMipmapEXT = (PFNGLGENERATETEXTUREMIPMAPEXTPROC) SMOL_GETGLPROC(glGenerateTextureMipmapEXT); 
+    glGenerateMultiTexMipmapEXT = (PFNGLGENERATEMULTITEXMIPMAPEXTPROC) SMOL_GETGLPROC(glGenerateMultiTexMipmapEXT); 
+    glFramebufferDrawBufferEXT = (PFNGLFRAMEBUFFERDRAWBUFFEREXTPROC) SMOL_GETGLPROC(glFramebufferDrawBufferEXT); 
+    glFramebufferDrawBuffersEXT = (PFNGLFRAMEBUFFERDRAWBUFFERSEXTPROC) SMOL_GETGLPROC(glFramebufferDrawBuffersEXT); 
+    glFramebufferReadBufferEXT = (PFNGLFRAMEBUFFERREADBUFFEREXTPROC) SMOL_GETGLPROC(glFramebufferReadBufferEXT); 
+    glGetFramebufferParameterivEXT = (PFNGLGETFRAMEBUFFERPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetFramebufferParameterivEXT); 
+    glNamedCopyBufferSubDataEXT = (PFNGLNAMEDCOPYBUFFERSUBDATAEXTPROC) SMOL_GETGLPROC(glNamedCopyBufferSubDataEXT); 
+    glNamedFramebufferTextureEXT = (PFNGLNAMEDFRAMEBUFFERTEXTUREEXTPROC) SMOL_GETGLPROC(glNamedFramebufferTextureEXT); 
+    glNamedFramebufferTextureLayerEXT = (PFNGLNAMEDFRAMEBUFFERTEXTURELAYEREXTPROC) SMOL_GETGLPROC(glNamedFramebufferTextureLayerEXT); 
+    glNamedFramebufferTextureFaceEXT = (PFNGLNAMEDFRAMEBUFFERTEXTUREFACEEXTPROC) SMOL_GETGLPROC(glNamedFramebufferTextureFaceEXT); 
+    glTextureRenderbufferEXT = (PFNGLTEXTURERENDERBUFFEREXTPROC) SMOL_GETGLPROC(glTextureRenderbufferEXT); 
+    glMultiTexRenderbufferEXT = (PFNGLMULTITEXRENDERBUFFEREXTPROC) SMOL_GETGLPROC(glMultiTexRenderbufferEXT); 
+    glVertexArrayVertexOffsetEXT = (PFNGLVERTEXARRAYVERTEXOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexOffsetEXT); 
+    glVertexArrayColorOffsetEXT = (PFNGLVERTEXARRAYCOLOROFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayColorOffsetEXT); 
+    glVertexArrayEdgeFlagOffsetEXT = (PFNGLVERTEXARRAYEDGEFLAGOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayEdgeFlagOffsetEXT); 
+    glVertexArrayIndexOffsetEXT = (PFNGLVERTEXARRAYINDEXOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayIndexOffsetEXT); 
+    glVertexArrayNormalOffsetEXT = (PFNGLVERTEXARRAYNORMALOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayNormalOffsetEXT); 
+    glVertexArrayTexCoordOffsetEXT = (PFNGLVERTEXARRAYTEXCOORDOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayTexCoordOffsetEXT); 
+    glVertexArrayMultiTexCoordOffsetEXT = (PFNGLVERTEXARRAYMULTITEXCOORDOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayMultiTexCoordOffsetEXT); 
+    glVertexArrayFogCoordOffsetEXT = (PFNGLVERTEXARRAYFOGCOORDOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayFogCoordOffsetEXT); 
+    glVertexArraySecondaryColorOffsetEXT = (PFNGLVERTEXARRAYSECONDARYCOLOROFFSETEXTPROC) SMOL_GETGLPROC(glVertexArraySecondaryColorOffsetEXT); 
+    glVertexArrayVertexAttribOffsetEXT = (PFNGLVERTEXARRAYVERTEXATTRIBOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribOffsetEXT); 
+    glVertexArrayVertexAttribIOffsetEXT = (PFNGLVERTEXARRAYVERTEXATTRIBIOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribIOffsetEXT); 
+    glEnableVertexArrayEXT = (PFNGLENABLEVERTEXARRAYEXTPROC) SMOL_GETGLPROC(glEnableVertexArrayEXT); 
+    glDisableVertexArrayEXT = (PFNGLDISABLEVERTEXARRAYEXTPROC) SMOL_GETGLPROC(glDisableVertexArrayEXT); 
+    glEnableVertexArrayAttribEXT = (PFNGLENABLEVERTEXARRAYATTRIBEXTPROC) SMOL_GETGLPROC(glEnableVertexArrayAttribEXT); 
+    glDisableVertexArrayAttribEXT = (PFNGLDISABLEVERTEXARRAYATTRIBEXTPROC) SMOL_GETGLPROC(glDisableVertexArrayAttribEXT); 
+    glGetVertexArrayIntegervEXT = (PFNGLGETVERTEXARRAYINTEGERVEXTPROC) SMOL_GETGLPROC(glGetVertexArrayIntegervEXT); 
+    glGetVertexArrayPointervEXT = (PFNGLGETVERTEXARRAYPOINTERVEXTPROC) SMOL_GETGLPROC(glGetVertexArrayPointervEXT); 
+    glGetVertexArrayIntegeri_vEXT = (PFNGLGETVERTEXARRAYINTEGERI_VEXTPROC) SMOL_GETGLPROC(glGetVertexArrayIntegeri_vEXT); 
+    glGetVertexArrayPointeri_vEXT = (PFNGLGETVERTEXARRAYPOINTERI_VEXTPROC) SMOL_GETGLPROC(glGetVertexArrayPointeri_vEXT); 
+    glMapNamedBufferRangeEXT = (PFNGLMAPNAMEDBUFFERRANGEEXTPROC) SMOL_GETGLPROC(glMapNamedBufferRangeEXT); 
+    glFlushMappedNamedBufferRangeEXT = (PFNGLFLUSHMAPPEDNAMEDBUFFERRANGEEXTPROC) SMOL_GETGLPROC(glFlushMappedNamedBufferRangeEXT); 
+    glNamedBufferStorageEXT = (PFNGLNAMEDBUFFERSTORAGEEXTPROC) SMOL_GETGLPROC(glNamedBufferStorageEXT); 
+    glClearNamedBufferDataEXT = (PFNGLCLEARNAMEDBUFFERDATAEXTPROC) SMOL_GETGLPROC(glClearNamedBufferDataEXT); 
+    glClearNamedBufferSubDataEXT = (PFNGLCLEARNAMEDBUFFERSUBDATAEXTPROC) SMOL_GETGLPROC(glClearNamedBufferSubDataEXT); 
+    glNamedFramebufferParameteriEXT = (PFNGLNAMEDFRAMEBUFFERPARAMETERIEXTPROC) SMOL_GETGLPROC(glNamedFramebufferParameteriEXT); 
+    glGetNamedFramebufferParameterivEXT = (PFNGLGETNAMEDFRAMEBUFFERPARAMETERIVEXTPROC) SMOL_GETGLPROC(glGetNamedFramebufferParameterivEXT); 
+    glProgramUniform1dEXT = (PFNGLPROGRAMUNIFORM1DEXTPROC) SMOL_GETGLPROC(glProgramUniform1dEXT); 
+    glProgramUniform2dEXT = (PFNGLPROGRAMUNIFORM2DEXTPROC) SMOL_GETGLPROC(glProgramUniform2dEXT); 
+    glProgramUniform3dEXT = (PFNGLPROGRAMUNIFORM3DEXTPROC) SMOL_GETGLPROC(glProgramUniform3dEXT); 
+    glProgramUniform4dEXT = (PFNGLPROGRAMUNIFORM4DEXTPROC) SMOL_GETGLPROC(glProgramUniform4dEXT); 
+    glProgramUniform1dvEXT = (PFNGLPROGRAMUNIFORM1DVEXTPROC) SMOL_GETGLPROC(glProgramUniform1dvEXT); 
+    glProgramUniform2dvEXT = (PFNGLPROGRAMUNIFORM2DVEXTPROC) SMOL_GETGLPROC(glProgramUniform2dvEXT); 
+    glProgramUniform3dvEXT = (PFNGLPROGRAMUNIFORM3DVEXTPROC) SMOL_GETGLPROC(glProgramUniform3dvEXT); 
+    glProgramUniform4dvEXT = (PFNGLPROGRAMUNIFORM4DVEXTPROC) SMOL_GETGLPROC(glProgramUniform4dvEXT); 
+    glProgramUniformMatrix2dvEXT = (PFNGLPROGRAMUNIFORMMATRIX2DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2dvEXT); 
+    glProgramUniformMatrix3dvEXT = (PFNGLPROGRAMUNIFORMMATRIX3DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3dvEXT); 
+    glProgramUniformMatrix4dvEXT = (PFNGLPROGRAMUNIFORMMATRIX4DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4dvEXT); 
+    glProgramUniformMatrix2x3dvEXT = (PFNGLPROGRAMUNIFORMMATRIX2X3DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x3dvEXT); 
+    glProgramUniformMatrix2x4dvEXT = (PFNGLPROGRAMUNIFORMMATRIX2X4DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix2x4dvEXT); 
+    glProgramUniformMatrix3x2dvEXT = (PFNGLPROGRAMUNIFORMMATRIX3X2DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x2dvEXT); 
+    glProgramUniformMatrix3x4dvEXT = (PFNGLPROGRAMUNIFORMMATRIX3X4DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix3x4dvEXT); 
+    glProgramUniformMatrix4x2dvEXT = (PFNGLPROGRAMUNIFORMMATRIX4X2DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x2dvEXT); 
+    glProgramUniformMatrix4x3dvEXT = (PFNGLPROGRAMUNIFORMMATRIX4X3DVEXTPROC) SMOL_GETGLPROC(glProgramUniformMatrix4x3dvEXT); 
+    glTextureBufferRangeEXT = (PFNGLTEXTUREBUFFERRANGEEXTPROC) SMOL_GETGLPROC(glTextureBufferRangeEXT); 
+    glTextureStorage1DEXT = (PFNGLTEXTURESTORAGE1DEXTPROC) SMOL_GETGLPROC(glTextureStorage1DEXT); 
+    glTextureStorage2DEXT = (PFNGLTEXTURESTORAGE2DEXTPROC) SMOL_GETGLPROC(glTextureStorage2DEXT); 
+    glTextureStorage3DEXT = (PFNGLTEXTURESTORAGE3DEXTPROC) SMOL_GETGLPROC(glTextureStorage3DEXT); 
+    glTextureStorage2DMultisampleEXT = (PFNGLTEXTURESTORAGE2DMULTISAMPLEEXTPROC) SMOL_GETGLPROC(glTextureStorage2DMultisampleEXT); 
+    glTextureStorage3DMultisampleEXT = (PFNGLTEXTURESTORAGE3DMULTISAMPLEEXTPROC) SMOL_GETGLPROC(glTextureStorage3DMultisampleEXT); 
+    glVertexArrayBindVertexBufferEXT = (PFNGLVERTEXARRAYBINDVERTEXBUFFEREXTPROC) SMOL_GETGLPROC(glVertexArrayBindVertexBufferEXT); 
+    glVertexArrayVertexAttribFormatEXT = (PFNGLVERTEXARRAYVERTEXATTRIBFORMATEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribFormatEXT); 
+    glVertexArrayVertexAttribIFormatEXT = (PFNGLVERTEXARRAYVERTEXATTRIBIFORMATEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribIFormatEXT); 
+    glVertexArrayVertexAttribLFormatEXT = (PFNGLVERTEXARRAYVERTEXATTRIBLFORMATEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribLFormatEXT); 
+    glVertexArrayVertexAttribBindingEXT = (PFNGLVERTEXARRAYVERTEXATTRIBBINDINGEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribBindingEXT); 
+    glVertexArrayVertexBindingDivisorEXT = (PFNGLVERTEXARRAYVERTEXBINDINGDIVISOREXTPROC) SMOL_GETGLPROC(glVertexArrayVertexBindingDivisorEXT); 
+    glVertexArrayVertexAttribLOffsetEXT = (PFNGLVERTEXARRAYVERTEXATTRIBLOFFSETEXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribLOffsetEXT); 
+    glTexturePageCommitmentEXT = (PFNGLTEXTUREPAGECOMMITMENTEXTPROC) SMOL_GETGLPROC(glTexturePageCommitmentEXT); 
+    glVertexArrayVertexAttribDivisorEXT = (PFNGLVERTEXARRAYVERTEXATTRIBDIVISOREXTPROC) SMOL_GETGLPROC(glVertexArrayVertexAttribDivisorEXT); 
+    glDrawArraysInstancedEXT = (PFNGLDRAWARRAYSINSTANCEDEXTPROC) SMOL_GETGLPROC(glDrawArraysInstancedEXT); 
+    glDrawElementsInstancedEXT = (PFNGLDRAWELEMENTSINSTANCEDEXTPROC) SMOL_GETGLPROC(glDrawElementsInstancedEXT); 
+    glPolygonOffsetClampEXT = (PFNGLPOLYGONOFFSETCLAMPEXTPROC) SMOL_GETGLPROC(glPolygonOffsetClampEXT); 
+    glRasterSamplesEXT = (PFNGLRASTERSAMPLESEXTPROC) SMOL_GETGLPROC(glRasterSamplesEXT); 
+    glUseShaderProgramEXT = (PFNGLUSESHADERPROGRAMEXTPROC) SMOL_GETGLPROC(glUseShaderProgramEXT); 
+    glActiveProgramEXT = (PFNGLACTIVEPROGRAMEXTPROC) SMOL_GETGLPROC(glActiveProgramEXT); 
+    glCreateShaderProgramEXT = (PFNGLCREATESHADERPROGRAMEXTPROC) SMOL_GETGLPROC(glCreateShaderProgramEXT); 
+    glFramebufferFetchBarrierEXT = (PFNGLFRAMEBUFFERFETCHBARRIEREXTPROC) SMOL_GETGLPROC(glFramebufferFetchBarrierEXT); 
+    glWindowRectanglesEXT = (PFNGLWINDOWRECTANGLESEXTPROC) SMOL_GETGLPROC(glWindowRectanglesEXT); 
+    glApplyFramebufferAttachmentCMAAINTEL = (PFNGLAPPLYFRAMEBUFFERATTACHMENTCMAAINTELPROC) SMOL_GETGLPROC(glApplyFramebufferAttachmentCMAAINTEL); 
+    glBeginPerfQueryINTEL = (PFNGLBEGINPERFQUERYINTELPROC) SMOL_GETGLPROC(glBeginPerfQueryINTEL); 
+    glCreatePerfQueryINTEL = (PFNGLCREATEPERFQUERYINTELPROC) SMOL_GETGLPROC(glCreatePerfQueryINTEL); 
+    glDeletePerfQueryINTEL = (PFNGLDELETEPERFQUERYINTELPROC) SMOL_GETGLPROC(glDeletePerfQueryINTEL); 
+    glEndPerfQueryINTEL = (PFNGLENDPERFQUERYINTELPROC) SMOL_GETGLPROC(glEndPerfQueryINTEL); 
+    glGetFirstPerfQueryIdINTEL = (PFNGLGETFIRSTPERFQUERYIDINTELPROC) SMOL_GETGLPROC(glGetFirstPerfQueryIdINTEL); 
+    glGetNextPerfQueryIdINTEL = (PFNGLGETNEXTPERFQUERYIDINTELPROC) SMOL_GETGLPROC(glGetNextPerfQueryIdINTEL); 
+    glGetPerfCounterInfoINTEL = (PFNGLGETPERFCOUNTERINFOINTELPROC) SMOL_GETGLPROC(glGetPerfCounterInfoINTEL); 
+    glGetPerfQueryDataINTEL = (PFNGLGETPERFQUERYDATAINTELPROC) SMOL_GETGLPROC(glGetPerfQueryDataINTEL); 
+    glGetPerfQueryIdByNameINTEL = (PFNGLGETPERFQUERYIDBYNAMEINTELPROC) SMOL_GETGLPROC(glGetPerfQueryIdByNameINTEL); 
+    glGetPerfQueryInfoINTEL = (PFNGLGETPERFQUERYINFOINTELPROC) SMOL_GETGLPROC(glGetPerfQueryInfoINTEL); 
+    glFramebufferParameteriMESA = (PFNGLFRAMEBUFFERPARAMETERIMESAPROC) SMOL_GETGLPROC(glFramebufferParameteriMESA); 
+    glGetFramebufferParameterivMESA = (PFNGLGETFRAMEBUFFERPARAMETERIVMESAPROC) SMOL_GETGLPROC(glGetFramebufferParameterivMESA); 
+    glMultiDrawArraysIndirectBindlessNV = (PFNGLMULTIDRAWARRAYSINDIRECTBINDLESSNVPROC) SMOL_GETGLPROC(glMultiDrawArraysIndirectBindlessNV); 
+    glMultiDrawElementsIndirectBindlessNV = (PFNGLMULTIDRAWELEMENTSINDIRECTBINDLESSNVPROC) SMOL_GETGLPROC(glMultiDrawElementsIndirectBindlessNV); 
+    glMultiDrawArraysIndirectBindlessCountNV = (PFNGLMULTIDRAWARRAYSINDIRECTBINDLESSCOUNTNVPROC) SMOL_GETGLPROC(glMultiDrawArraysIndirectBindlessCountNV); 
+    glMultiDrawElementsIndirectBindlessCountNV = (PFNGLMULTIDRAWELEMENTSINDIRECTBINDLESSCOUNTNVPROC) SMOL_GETGLPROC(glMultiDrawElementsIndirectBindlessCountNV); 
+    glGetTextureHandleNV = (PFNGLGETTEXTUREHANDLENVPROC) SMOL_GETGLPROC(glGetTextureHandleNV); 
+    glGetTextureSamplerHandleNV = (PFNGLGETTEXTURESAMPLERHANDLENVPROC) SMOL_GETGLPROC(glGetTextureSamplerHandleNV); 
+    glMakeTextureHandleResidentNV = (PFNGLMAKETEXTUREHANDLERESIDENTNVPROC) SMOL_GETGLPROC(glMakeTextureHandleResidentNV); 
+    glMakeTextureHandleNonResidentNV = (PFNGLMAKETEXTUREHANDLENONRESIDENTNVPROC) SMOL_GETGLPROC(glMakeTextureHandleNonResidentNV); 
+    glGetImageHandleNV = (PFNGLGETIMAGEHANDLENVPROC) SMOL_GETGLPROC(glGetImageHandleNV); 
+    glMakeImageHandleResidentNV = (PFNGLMAKEIMAGEHANDLERESIDENTNVPROC) SMOL_GETGLPROC(glMakeImageHandleResidentNV); 
+    glMakeImageHandleNonResidentNV = (PFNGLMAKEIMAGEHANDLENONRESIDENTNVPROC) SMOL_GETGLPROC(glMakeImageHandleNonResidentNV); 
+    glUniformHandleui64NV = (PFNGLUNIFORMHANDLEUI64NVPROC) SMOL_GETGLPROC(glUniformHandleui64NV); 
+    glUniformHandleui64vNV = (PFNGLUNIFORMHANDLEUI64VNVPROC) SMOL_GETGLPROC(glUniformHandleui64vNV); 
+    glProgramUniformHandleui64NV = (PFNGLPROGRAMUNIFORMHANDLEUI64NVPROC) SMOL_GETGLPROC(glProgramUniformHandleui64NV); 
+    glProgramUniformHandleui64vNV = (PFNGLPROGRAMUNIFORMHANDLEUI64VNVPROC) SMOL_GETGLPROC(glProgramUniformHandleui64vNV); 
+    glIsTextureHandleResidentNV = (PFNGLISTEXTUREHANDLERESIDENTNVPROC) SMOL_GETGLPROC(glIsTextureHandleResidentNV); 
+    glIsImageHandleResidentNV = (PFNGLISIMAGEHANDLERESIDENTNVPROC) SMOL_GETGLPROC(glIsImageHandleResidentNV); 
+    glBlendParameteriNV = (PFNGLBLENDPARAMETERINVPROC) SMOL_GETGLPROC(glBlendParameteriNV); 
+    glBlendBarrierNV = (PFNGLBLENDBARRIERNVPROC) SMOL_GETGLPROC(glBlendBarrierNV); 
+    glViewportPositionWScaleNV = (PFNGLVIEWPORTPOSITIONWSCALENVPROC) SMOL_GETGLPROC(glViewportPositionWScaleNV); 
+    glCreateStatesNV = (PFNGLCREATESTATESNVPROC) SMOL_GETGLPROC(glCreateStatesNV); 
+    glDeleteStatesNV = (PFNGLDELETESTATESNVPROC) SMOL_GETGLPROC(glDeleteStatesNV); 
+    glIsStateNV = (PFNGLISSTATENVPROC) SMOL_GETGLPROC(glIsStateNV); 
+    glStateCaptureNV = (PFNGLSTATECAPTURENVPROC) SMOL_GETGLPROC(glStateCaptureNV); 
+    glGetCommandHeaderNV = (PFNGLGETCOMMANDHEADERNVPROC) SMOL_GETGLPROC(glGetCommandHeaderNV); 
+    glGetStageIndexNV = (PFNGLGETSTAGEINDEXNVPROC) SMOL_GETGLPROC(glGetStageIndexNV); 
+    glDrawCommandsNV = (PFNGLDRAWCOMMANDSNVPROC) SMOL_GETGLPROC(glDrawCommandsNV); 
+    glDrawCommandsAddressNV = (PFNGLDRAWCOMMANDSADDRESSNVPROC) SMOL_GETGLPROC(glDrawCommandsAddressNV); 
+    glDrawCommandsStatesNV = (PFNGLDRAWCOMMANDSSTATESNVPROC) SMOL_GETGLPROC(glDrawCommandsStatesNV); 
+    glDrawCommandsStatesAddressNV = (PFNGLDRAWCOMMANDSSTATESADDRESSNVPROC) SMOL_GETGLPROC(glDrawCommandsStatesAddressNV); 
+    glCreateCommandListsNV = (PFNGLCREATECOMMANDLISTSNVPROC) SMOL_GETGLPROC(glCreateCommandListsNV); 
+    glDeleteCommandListsNV = (PFNGLDELETECOMMANDLISTSNVPROC) SMOL_GETGLPROC(glDeleteCommandListsNV); 
+    glIsCommandListNV = (PFNGLISCOMMANDLISTNVPROC) SMOL_GETGLPROC(glIsCommandListNV); 
+    glListDrawCommandsStatesClientNV = (PFNGLLISTDRAWCOMMANDSSTATESCLIENTNVPROC) SMOL_GETGLPROC(glListDrawCommandsStatesClientNV); 
+    glCommandListSegmentsNV = (PFNGLCOMMANDLISTSEGMENTSNVPROC) SMOL_GETGLPROC(glCommandListSegmentsNV); 
+    glCompileCommandListNV = (PFNGLCOMPILECOMMANDLISTNVPROC) SMOL_GETGLPROC(glCompileCommandListNV); 
+    glCallCommandListNV = (PFNGLCALLCOMMANDLISTNVPROC) SMOL_GETGLPROC(glCallCommandListNV); 
+    glBeginConditionalRenderNV = (PFNGLBEGINCONDITIONALRENDERNVPROC) SMOL_GETGLPROC(glBeginConditionalRenderNV); 
+    glEndConditionalRenderNV = (PFNGLENDCONDITIONALRENDERNVPROC) SMOL_GETGLPROC(glEndConditionalRenderNV); 
+    glSubpixelPrecisionBiasNV = (PFNGLSUBPIXELPRECISIONBIASNVPROC) SMOL_GETGLPROC(glSubpixelPrecisionBiasNV); 
+    glConservativeRasterParameterfNV = (PFNGLCONSERVATIVERASTERPARAMETERFNVPROC) SMOL_GETGLPROC(glConservativeRasterParameterfNV); 
+    glConservativeRasterParameteriNV = (PFNGLCONSERVATIVERASTERPARAMETERINVPROC) SMOL_GETGLPROC(glConservativeRasterParameteriNV); 
+    glDepthRangedNV = (PFNGLDEPTHRANGEDNVPROC) SMOL_GETGLPROC(glDepthRangedNV); 
+    glClearDepthdNV = (PFNGLCLEARDEPTHDNVPROC) SMOL_GETGLPROC(glClearDepthdNV); 
+    glDepthBoundsdNV = (PFNGLDEPTHBOUNDSDNVPROC) SMOL_GETGLPROC(glDepthBoundsdNV); 
+    glDrawVkImageNV = (PFNGLDRAWVKIMAGENVPROC) SMOL_GETGLPROC(glDrawVkImageNV); 
+    glGetVkProcAddrNV = (PFNGLGETVKPROCADDRNVPROC) SMOL_GETGLPROC(glGetVkProcAddrNV); 
+    glWaitVkSemaphoreNV = (PFNGLWAITVKSEMAPHORENVPROC) SMOL_GETGLPROC(glWaitVkSemaphoreNV); 
+    glSignalVkSemaphoreNV = (PFNGLSIGNALVKSEMAPHORENVPROC) SMOL_GETGLPROC(glSignalVkSemaphoreNV); 
+    glSignalVkFenceNV = (PFNGLSIGNALVKFENCENVPROC) SMOL_GETGLPROC(glSignalVkFenceNV); 
+    glFragmentCoverageColorNV = (PFNGLFRAGMENTCOVERAGECOLORNVPROC) SMOL_GETGLPROC(glFragmentCoverageColorNV); 
+    glCoverageModulationTableNV = (PFNGLCOVERAGEMODULATIONTABLENVPROC) SMOL_GETGLPROC(glCoverageModulationTableNV); 
+    glGetCoverageModulationTableNV = (PFNGLGETCOVERAGEMODULATIONTABLENVPROC) SMOL_GETGLPROC(glGetCoverageModulationTableNV); 
+    glCoverageModulationNV = (PFNGLCOVERAGEMODULATIONNVPROC) SMOL_GETGLPROC(glCoverageModulationNV); 
+    glRenderbufferStorageMultisampleCoverageNV = (PFNGLRENDERBUFFERSTORAGEMULTISAMPLECOVERAGENVPROC) SMOL_GETGLPROC(glRenderbufferStorageMultisampleCoverageNV); 
+    glUniform1i64NV = (PFNGLUNIFORM1I64NVPROC) SMOL_GETGLPROC(glUniform1i64NV); 
+    glUniform2i64NV = (PFNGLUNIFORM2I64NVPROC) SMOL_GETGLPROC(glUniform2i64NV); 
+    glUniform3i64NV = (PFNGLUNIFORM3I64NVPROC) SMOL_GETGLPROC(glUniform3i64NV); 
+    glUniform4i64NV = (PFNGLUNIFORM4I64NVPROC) SMOL_GETGLPROC(glUniform4i64NV); 
+    glUniform1i64vNV = (PFNGLUNIFORM1I64VNVPROC) SMOL_GETGLPROC(glUniform1i64vNV); 
+    glUniform2i64vNV = (PFNGLUNIFORM2I64VNVPROC) SMOL_GETGLPROC(glUniform2i64vNV); 
+    glUniform3i64vNV = (PFNGLUNIFORM3I64VNVPROC) SMOL_GETGLPROC(glUniform3i64vNV); 
+    glUniform4i64vNV = (PFNGLUNIFORM4I64VNVPROC) SMOL_GETGLPROC(glUniform4i64vNV); 
+    glUniform1ui64NV = (PFNGLUNIFORM1UI64NVPROC) SMOL_GETGLPROC(glUniform1ui64NV); 
+    glUniform2ui64NV = (PFNGLUNIFORM2UI64NVPROC) SMOL_GETGLPROC(glUniform2ui64NV); 
+    glUniform3ui64NV = (PFNGLUNIFORM3UI64NVPROC) SMOL_GETGLPROC(glUniform3ui64NV); 
+    glUniform4ui64NV = (PFNGLUNIFORM4UI64NVPROC) SMOL_GETGLPROC(glUniform4ui64NV); 
+    glUniform1ui64vNV = (PFNGLUNIFORM1UI64VNVPROC) SMOL_GETGLPROC(glUniform1ui64vNV); 
+    glUniform2ui64vNV = (PFNGLUNIFORM2UI64VNVPROC) SMOL_GETGLPROC(glUniform2ui64vNV); 
+    glUniform3ui64vNV = (PFNGLUNIFORM3UI64VNVPROC) SMOL_GETGLPROC(glUniform3ui64vNV); 
+    glUniform4ui64vNV = (PFNGLUNIFORM4UI64VNVPROC) SMOL_GETGLPROC(glUniform4ui64vNV); 
+    glGetUniformi64vNV = (PFNGLGETUNIFORMI64VNVPROC) SMOL_GETGLPROC(glGetUniformi64vNV); 
+    glProgramUniform1i64NV = (PFNGLPROGRAMUNIFORM1I64NVPROC) SMOL_GETGLPROC(glProgramUniform1i64NV); 
+    glProgramUniform2i64NV = (PFNGLPROGRAMUNIFORM2I64NVPROC) SMOL_GETGLPROC(glProgramUniform2i64NV); 
+    glProgramUniform3i64NV = (PFNGLPROGRAMUNIFORM3I64NVPROC) SMOL_GETGLPROC(glProgramUniform3i64NV); 
+    glProgramUniform4i64NV = (PFNGLPROGRAMUNIFORM4I64NVPROC) SMOL_GETGLPROC(glProgramUniform4i64NV); 
+    glProgramUniform1i64vNV = (PFNGLPROGRAMUNIFORM1I64VNVPROC) SMOL_GETGLPROC(glProgramUniform1i64vNV); 
+    glProgramUniform2i64vNV = (PFNGLPROGRAMUNIFORM2I64VNVPROC) SMOL_GETGLPROC(glProgramUniform2i64vNV); 
+    glProgramUniform3i64vNV = (PFNGLPROGRAMUNIFORM3I64VNVPROC) SMOL_GETGLPROC(glProgramUniform3i64vNV); 
+    glProgramUniform4i64vNV = (PFNGLPROGRAMUNIFORM4I64VNVPROC) SMOL_GETGLPROC(glProgramUniform4i64vNV); 
+    glProgramUniform1ui64NV = (PFNGLPROGRAMUNIFORM1UI64NVPROC) SMOL_GETGLPROC(glProgramUniform1ui64NV); 
+    glProgramUniform2ui64NV = (PFNGLPROGRAMUNIFORM2UI64NVPROC) SMOL_GETGLPROC(glProgramUniform2ui64NV); 
+    glProgramUniform3ui64NV = (PFNGLPROGRAMUNIFORM3UI64NVPROC) SMOL_GETGLPROC(glProgramUniform3ui64NV); 
+    glProgramUniform4ui64NV = (PFNGLPROGRAMUNIFORM4UI64NVPROC) SMOL_GETGLPROC(glProgramUniform4ui64NV); 
+    glProgramUniform1ui64vNV = (PFNGLPROGRAMUNIFORM1UI64VNVPROC) SMOL_GETGLPROC(glProgramUniform1ui64vNV); 
+    glProgramUniform2ui64vNV = (PFNGLPROGRAMUNIFORM2UI64VNVPROC) SMOL_GETGLPROC(glProgramUniform2ui64vNV); 
+    glProgramUniform3ui64vNV = (PFNGLPROGRAMUNIFORM3UI64VNVPROC) SMOL_GETGLPROC(glProgramUniform3ui64vNV); 
+    glProgramUniform4ui64vNV = (PFNGLPROGRAMUNIFORM4UI64VNVPROC) SMOL_GETGLPROC(glProgramUniform4ui64vNV); 
+    glGetInternalformatSampleivNV = (PFNGLGETINTERNALFORMATSAMPLEIVNVPROC) SMOL_GETGLPROC(glGetInternalformatSampleivNV); 
+    glGetMemoryObjectDetachedResourcesuivNV = (PFNGLGETMEMORYOBJECTDETACHEDRESOURCESUIVNVPROC) SMOL_GETGLPROC(glGetMemoryObjectDetachedResourcesuivNV); 
+    glResetMemoryObjectParameterNV = (PFNGLRESETMEMORYOBJECTPARAMETERNVPROC) SMOL_GETGLPROC(glResetMemoryObjectParameterNV); 
+    glTexAttachMemoryNV = (PFNGLTEXATTACHMEMORYNVPROC) SMOL_GETGLPROC(glTexAttachMemoryNV); 
+    glBufferAttachMemoryNV = (PFNGLBUFFERATTACHMEMORYNVPROC) SMOL_GETGLPROC(glBufferAttachMemoryNV); 
+    glTextureAttachMemoryNV = (PFNGLTEXTUREATTACHMEMORYNVPROC) SMOL_GETGLPROC(glTextureAttachMemoryNV); 
+    glNamedBufferAttachMemoryNV = (PFNGLNAMEDBUFFERATTACHMEMORYNVPROC) SMOL_GETGLPROC(glNamedBufferAttachMemoryNV); 
+    glBufferPageCommitmentMemNV = (PFNGLBUFFERPAGECOMMITMENTMEMNVPROC) SMOL_GETGLPROC(glBufferPageCommitmentMemNV); 
+    glTexPageCommitmentMemNV = (PFNGLTEXPAGECOMMITMENTMEMNVPROC) SMOL_GETGLPROC(glTexPageCommitmentMemNV); 
+    glNamedBufferPageCommitmentMemNV = (PFNGLNAMEDBUFFERPAGECOMMITMENTMEMNVPROC) SMOL_GETGLPROC(glNamedBufferPageCommitmentMemNV); 
+    glTexturePageCommitmentMemNV = (PFNGLTEXTUREPAGECOMMITMENTMEMNVPROC) SMOL_GETGLPROC(glTexturePageCommitmentMemNV); 
+    glDrawMeshTasksNV = (PFNGLDRAWMESHTASKSNVPROC) SMOL_GETGLPROC(glDrawMeshTasksNV); 
+    glDrawMeshTasksIndirectNV = (PFNGLDRAWMESHTASKSINDIRECTNVPROC) SMOL_GETGLPROC(glDrawMeshTasksIndirectNV); 
+    glMultiDrawMeshTasksIndirectNV = (PFNGLMULTIDRAWMESHTASKSINDIRECTNVPROC) SMOL_GETGLPROC(glMultiDrawMeshTasksIndirectNV); 
+    glMultiDrawMeshTasksIndirectCountNV = (PFNGLMULTIDRAWMESHTASKSINDIRECTCOUNTNVPROC) SMOL_GETGLPROC(glMultiDrawMeshTasksIndirectCountNV); 
+    glGenPathsNV = (PFNGLGENPATHSNVPROC) SMOL_GETGLPROC(glGenPathsNV); 
+    glDeletePathsNV = (PFNGLDELETEPATHSNVPROC) SMOL_GETGLPROC(glDeletePathsNV); 
+    glIsPathNV = (PFNGLISPATHNVPROC) SMOL_GETGLPROC(glIsPathNV); 
+    glPathCommandsNV = (PFNGLPATHCOMMANDSNVPROC) SMOL_GETGLPROC(glPathCommandsNV); 
+    glPathCoordsNV = (PFNGLPATHCOORDSNVPROC) SMOL_GETGLPROC(glPathCoordsNV); 
+    glPathSubCommandsNV = (PFNGLPATHSUBCOMMANDSNVPROC) SMOL_GETGLPROC(glPathSubCommandsNV); 
+    glPathSubCoordsNV = (PFNGLPATHSUBCOORDSNVPROC) SMOL_GETGLPROC(glPathSubCoordsNV); 
+    glPathStringNV = (PFNGLPATHSTRINGNVPROC) SMOL_GETGLPROC(glPathStringNV); 
+    glPathGlyphsNV = (PFNGLPATHGLYPHSNVPROC) SMOL_GETGLPROC(glPathGlyphsNV); 
+    glPathGlyphRangeNV = (PFNGLPATHGLYPHRANGENVPROC) SMOL_GETGLPROC(glPathGlyphRangeNV); 
+    glWeightPathsNV = (PFNGLWEIGHTPATHSNVPROC) SMOL_GETGLPROC(glWeightPathsNV); 
+    glCopyPathNV = (PFNGLCOPYPATHNVPROC) SMOL_GETGLPROC(glCopyPathNV); 
+    glInterpolatePathsNV = (PFNGLINTERPOLATEPATHSNVPROC) SMOL_GETGLPROC(glInterpolatePathsNV); 
+    glTransformPathNV = (PFNGLTRANSFORMPATHNVPROC) SMOL_GETGLPROC(glTransformPathNV); 
+    glPathParameterivNV = (PFNGLPATHPARAMETERIVNVPROC) SMOL_GETGLPROC(glPathParameterivNV); 
+    glPathParameteriNV = (PFNGLPATHPARAMETERINVPROC) SMOL_GETGLPROC(glPathParameteriNV); 
+    glPathParameterfvNV = (PFNGLPATHPARAMETERFVNVPROC) SMOL_GETGLPROC(glPathParameterfvNV); 
+    glPathParameterfNV = (PFNGLPATHPARAMETERFNVPROC) SMOL_GETGLPROC(glPathParameterfNV); 
+    glPathDashArrayNV = (PFNGLPATHDASHARRAYNVPROC) SMOL_GETGLPROC(glPathDashArrayNV); 
+    glPathStencilFuncNV = (PFNGLPATHSTENCILFUNCNVPROC) SMOL_GETGLPROC(glPathStencilFuncNV); 
+    glPathStencilDepthOffsetNV = (PFNGLPATHSTENCILDEPTHOFFSETNVPROC) SMOL_GETGLPROC(glPathStencilDepthOffsetNV); 
+    glStencilFillPathNV = (PFNGLSTENCILFILLPATHNVPROC) SMOL_GETGLPROC(glStencilFillPathNV); 
+    glStencilStrokePathNV = (PFNGLSTENCILSTROKEPATHNVPROC) SMOL_GETGLPROC(glStencilStrokePathNV); 
+    glStencilFillPathInstancedNV = (PFNGLSTENCILFILLPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glStencilFillPathInstancedNV); 
+    glStencilStrokePathInstancedNV = (PFNGLSTENCILSTROKEPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glStencilStrokePathInstancedNV); 
+    glPathCoverDepthFuncNV = (PFNGLPATHCOVERDEPTHFUNCNVPROC) SMOL_GETGLPROC(glPathCoverDepthFuncNV); 
+    glCoverFillPathNV = (PFNGLCOVERFILLPATHNVPROC) SMOL_GETGLPROC(glCoverFillPathNV); 
+    glCoverStrokePathNV = (PFNGLCOVERSTROKEPATHNVPROC) SMOL_GETGLPROC(glCoverStrokePathNV); 
+    glCoverFillPathInstancedNV = (PFNGLCOVERFILLPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glCoverFillPathInstancedNV); 
+    glCoverStrokePathInstancedNV = (PFNGLCOVERSTROKEPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glCoverStrokePathInstancedNV); 
+    glGetPathParameterivNV = (PFNGLGETPATHPARAMETERIVNVPROC) SMOL_GETGLPROC(glGetPathParameterivNV); 
+    glGetPathParameterfvNV = (PFNGLGETPATHPARAMETERFVNVPROC) SMOL_GETGLPROC(glGetPathParameterfvNV); 
+    glGetPathCommandsNV = (PFNGLGETPATHCOMMANDSNVPROC) SMOL_GETGLPROC(glGetPathCommandsNV); 
+    glGetPathCoordsNV = (PFNGLGETPATHCOORDSNVPROC) SMOL_GETGLPROC(glGetPathCoordsNV); 
+    glGetPathDashArrayNV = (PFNGLGETPATHDASHARRAYNVPROC) SMOL_GETGLPROC(glGetPathDashArrayNV); 
+    glGetPathMetricsNV = (PFNGLGETPATHMETRICSNVPROC) SMOL_GETGLPROC(glGetPathMetricsNV); 
+    glGetPathMetricRangeNV = (PFNGLGETPATHMETRICRANGENVPROC) SMOL_GETGLPROC(glGetPathMetricRangeNV); 
+    glGetPathSpacingNV = (PFNGLGETPATHSPACINGNVPROC) SMOL_GETGLPROC(glGetPathSpacingNV); 
+    glIsPointInFillPathNV = (PFNGLISPOINTINFILLPATHNVPROC) SMOL_GETGLPROC(glIsPointInFillPathNV); 
+    glIsPointInStrokePathNV = (PFNGLISPOINTINSTROKEPATHNVPROC) SMOL_GETGLPROC(glIsPointInStrokePathNV); 
+    glGetPathLengthNV = (PFNGLGETPATHLENGTHNVPROC) SMOL_GETGLPROC(glGetPathLengthNV); 
+    glPointAlongPathNV = (PFNGLPOINTALONGPATHNVPROC) SMOL_GETGLPROC(glPointAlongPathNV); 
+    glMatrixLoad3x2fNV = (PFNGLMATRIXLOAD3X2FNVPROC) SMOL_GETGLPROC(glMatrixLoad3x2fNV); 
+    glMatrixLoad3x3fNV = (PFNGLMATRIXLOAD3X3FNVPROC) SMOL_GETGLPROC(glMatrixLoad3x3fNV); 
+    glMatrixLoadTranspose3x3fNV = (PFNGLMATRIXLOADTRANSPOSE3X3FNVPROC) SMOL_GETGLPROC(glMatrixLoadTranspose3x3fNV); 
+    glMatrixMult3x2fNV = (PFNGLMATRIXMULT3X2FNVPROC) SMOL_GETGLPROC(glMatrixMult3x2fNV); 
+    glMatrixMult3x3fNV = (PFNGLMATRIXMULT3X3FNVPROC) SMOL_GETGLPROC(glMatrixMult3x3fNV); 
+    glMatrixMultTranspose3x3fNV = (PFNGLMATRIXMULTTRANSPOSE3X3FNVPROC) SMOL_GETGLPROC(glMatrixMultTranspose3x3fNV); 
+    glStencilThenCoverFillPathNV = (PFNGLSTENCILTHENCOVERFILLPATHNVPROC) SMOL_GETGLPROC(glStencilThenCoverFillPathNV); 
+    glStencilThenCoverStrokePathNV = (PFNGLSTENCILTHENCOVERSTROKEPATHNVPROC) SMOL_GETGLPROC(glStencilThenCoverStrokePathNV); 
+    glStencilThenCoverFillPathInstancedNV = (PFNGLSTENCILTHENCOVERFILLPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glStencilThenCoverFillPathInstancedNV); 
+    glStencilThenCoverStrokePathInstancedNV = (PFNGLSTENCILTHENCOVERSTROKEPATHINSTANCEDNVPROC) SMOL_GETGLPROC(glStencilThenCoverStrokePathInstancedNV); 
+    glPathGlyphIndexRangeNV = (PFNGLPATHGLYPHINDEXRANGENVPROC) SMOL_GETGLPROC(glPathGlyphIndexRangeNV); 
+    glPathGlyphIndexArrayNV = (PFNGLPATHGLYPHINDEXARRAYNVPROC) SMOL_GETGLPROC(glPathGlyphIndexArrayNV); 
+    glPathMemoryGlyphIndexArrayNV = (PFNGLPATHMEMORYGLYPHINDEXARRAYNVPROC) SMOL_GETGLPROC(glPathMemoryGlyphIndexArrayNV); 
+    glProgramPathFragmentInputGenNV = (PFNGLPROGRAMPATHFRAGMENTINPUTGENNVPROC) SMOL_GETGLPROC(glProgramPathFragmentInputGenNV); 
+    glGetProgramResourcefvNV = (PFNGLGETPROGRAMRESOURCEFVNVPROC) SMOL_GETGLPROC(glGetProgramResourcefvNV); 
+    glFramebufferSampleLocationsfvNV = (PFNGLFRAMEBUFFERSAMPLELOCATIONSFVNVPROC) SMOL_GETGLPROC(glFramebufferSampleLocationsfvNV); 
+    glNamedFramebufferSampleLocationsfvNV = (PFNGLNAMEDFRAMEBUFFERSAMPLELOCATIONSFVNVPROC) SMOL_GETGLPROC(glNamedFramebufferSampleLocationsfvNV); 
+    glResolveDepthValuesNV = (PFNGLRESOLVEDEPTHVALUESNVPROC) SMOL_GETGLPROC(glResolveDepthValuesNV); 
+    glScissorExclusiveNV = (PFNGLSCISSOREXCLUSIVENVPROC) SMOL_GETGLPROC(glScissorExclusiveNV); 
+    glScissorExclusiveArrayvNV = (PFNGLSCISSOREXCLUSIVEARRAYVNVPROC) SMOL_GETGLPROC(glScissorExclusiveArrayvNV); 
+    glMakeBufferResidentNV = (PFNGLMAKEBUFFERRESIDENTNVPROC) SMOL_GETGLPROC(glMakeBufferResidentNV); 
+    glMakeBufferNonResidentNV = (PFNGLMAKEBUFFERNONRESIDENTNVPROC) SMOL_GETGLPROC(glMakeBufferNonResidentNV); 
+    glIsBufferResidentNV = (PFNGLISBUFFERRESIDENTNVPROC) SMOL_GETGLPROC(glIsBufferResidentNV); 
+    glMakeNamedBufferResidentNV = (PFNGLMAKENAMEDBUFFERRESIDENTNVPROC) SMOL_GETGLPROC(glMakeNamedBufferResidentNV); 
+    glMakeNamedBufferNonResidentNV = (PFNGLMAKENAMEDBUFFERNONRESIDENTNVPROC) SMOL_GETGLPROC(glMakeNamedBufferNonResidentNV); 
+    glIsNamedBufferResidentNV = (PFNGLISNAMEDBUFFERRESIDENTNVPROC) SMOL_GETGLPROC(glIsNamedBufferResidentNV); 
+    glGetBufferParameterui64vNV = (PFNGLGETBUFFERPARAMETERUI64VNVPROC) SMOL_GETGLPROC(glGetBufferParameterui64vNV); 
+    glGetNamedBufferParameterui64vNV = (PFNGLGETNAMEDBUFFERPARAMETERUI64VNVPROC) SMOL_GETGLPROC(glGetNamedBufferParameterui64vNV); 
+    glGetIntegerui64vNV = (PFNGLGETINTEGERUI64VNVPROC) SMOL_GETGLPROC(glGetIntegerui64vNV); 
+    glUniformui64NV = (PFNGLUNIFORMUI64NVPROC) SMOL_GETGLPROC(glUniformui64NV); 
+    glUniformui64vNV = (PFNGLUNIFORMUI64VNVPROC) SMOL_GETGLPROC(glUniformui64vNV); 
+    glGetUniformui64vNV = (PFNGLGETUNIFORMUI64VNVPROC) SMOL_GETGLPROC(glGetUniformui64vNV); 
+    glProgramUniformui64NV = (PFNGLPROGRAMUNIFORMUI64NVPROC) SMOL_GETGLPROC(glProgramUniformui64NV); 
+    glProgramUniformui64vNV = (PFNGLPROGRAMUNIFORMUI64VNVPROC) SMOL_GETGLPROC(glProgramUniformui64vNV); 
+    glBindShadingRateImageNV = (PFNGLBINDSHADINGRATEIMAGENVPROC) SMOL_GETGLPROC(glBindShadingRateImageNV); 
+    glGetShadingRateImagePaletteNV = (PFNGLGETSHADINGRATEIMAGEPALETTENVPROC) SMOL_GETGLPROC(glGetShadingRateImagePaletteNV); 
+    glGetShadingRateSampleLocationivNV = (PFNGLGETSHADINGRATESAMPLELOCATIONIVNVPROC) SMOL_GETGLPROC(glGetShadingRateSampleLocationivNV); 
+    glShadingRateImageBarrierNV = (PFNGLSHADINGRATEIMAGEBARRIERNVPROC) SMOL_GETGLPROC(glShadingRateImageBarrierNV); 
+    glShadingRateImagePaletteNV = (PFNGLSHADINGRATEIMAGEPALETTENVPROC) SMOL_GETGLPROC(glShadingRateImagePaletteNV); 
+    glShadingRateSampleOrderNV = (PFNGLSHADINGRATESAMPLEORDERNVPROC) SMOL_GETGLPROC(glShadingRateSampleOrderNV); 
+    glShadingRateSampleOrderCustomNV = (PFNGLSHADINGRATESAMPLEORDERCUSTOMNVPROC) SMOL_GETGLPROC(glShadingRateSampleOrderCustomNV); 
+    glTextureBarrierNV = (PFNGLTEXTUREBARRIERNVPROC) SMOL_GETGLPROC(glTextureBarrierNV); 
+    glVertexAttribL1i64NV = (PFNGLVERTEXATTRIBL1I64NVPROC) SMOL_GETGLPROC(glVertexAttribL1i64NV); 
+    glVertexAttribL2i64NV = (PFNGLVERTEXATTRIBL2I64NVPROC) SMOL_GETGLPROC(glVertexAttribL2i64NV); 
+    glVertexAttribL3i64NV = (PFNGLVERTEXATTRIBL3I64NVPROC) SMOL_GETGLPROC(glVertexAttribL3i64NV); 
+    glVertexAttribL4i64NV = (PFNGLVERTEXATTRIBL4I64NVPROC) SMOL_GETGLPROC(glVertexAttribL4i64NV); 
+    glVertexAttribL1i64vNV = (PFNGLVERTEXATTRIBL1I64VNVPROC) SMOL_GETGLPROC(glVertexAttribL1i64vNV); 
+    glVertexAttribL2i64vNV = (PFNGLVERTEXATTRIBL2I64VNVPROC) SMOL_GETGLPROC(glVertexAttribL2i64vNV); 
+    glVertexAttribL3i64vNV = (PFNGLVERTEXATTRIBL3I64VNVPROC) SMOL_GETGLPROC(glVertexAttribL3i64vNV); 
+    glVertexAttribL4i64vNV = (PFNGLVERTEXATTRIBL4I64VNVPROC) SMOL_GETGLPROC(glVertexAttribL4i64vNV); 
+    glVertexAttribL1ui64NV = (PFNGLVERTEXATTRIBL1UI64NVPROC) SMOL_GETGLPROC(glVertexAttribL1ui64NV); 
+    glVertexAttribL2ui64NV = (PFNGLVERTEXATTRIBL2UI64NVPROC) SMOL_GETGLPROC(glVertexAttribL2ui64NV); 
+    glVertexAttribL3ui64NV = (PFNGLVERTEXATTRIBL3UI64NVPROC) SMOL_GETGLPROC(glVertexAttribL3ui64NV); 
+    glVertexAttribL4ui64NV = (PFNGLVERTEXATTRIBL4UI64NVPROC) SMOL_GETGLPROC(glVertexAttribL4ui64NV); 
+    glVertexAttribL1ui64vNV = (PFNGLVERTEXATTRIBL1UI64VNVPROC) SMOL_GETGLPROC(glVertexAttribL1ui64vNV); 
+    glVertexAttribL2ui64vNV = (PFNGLVERTEXATTRIBL2UI64VNVPROC) SMOL_GETGLPROC(glVertexAttribL2ui64vNV); 
+    glVertexAttribL3ui64vNV = (PFNGLVERTEXATTRIBL3UI64VNVPROC) SMOL_GETGLPROC(glVertexAttribL3ui64vNV); 
+    glVertexAttribL4ui64vNV = (PFNGLVERTEXATTRIBL4UI64VNVPROC) SMOL_GETGLPROC(glVertexAttribL4ui64vNV); 
+    glGetVertexAttribLi64vNV = (PFNGLGETVERTEXATTRIBLI64VNVPROC) SMOL_GETGLPROC(glGetVertexAttribLi64vNV); 
+    glGetVertexAttribLui64vNV = (PFNGLGETVERTEXATTRIBLUI64VNVPROC) SMOL_GETGLPROC(glGetVertexAttribLui64vNV); 
+    glVertexAttribLFormatNV = (PFNGLVERTEXATTRIBLFORMATNVPROC) SMOL_GETGLPROC(glVertexAttribLFormatNV); 
+    glBufferAddressRangeNV = (PFNGLBUFFERADDRESSRANGENVPROC) SMOL_GETGLPROC(glBufferAddressRangeNV); 
+    glVertexFormatNV = (PFNGLVERTEXFORMATNVPROC) SMOL_GETGLPROC(glVertexFormatNV); 
+    glNormalFormatNV = (PFNGLNORMALFORMATNVPROC) SMOL_GETGLPROC(glNormalFormatNV); 
+    glColorFormatNV = (PFNGLCOLORFORMATNVPROC) SMOL_GETGLPROC(glColorFormatNV); 
+    glIndexFormatNV = (PFNGLINDEXFORMATNVPROC) SMOL_GETGLPROC(glIndexFormatNV); 
+    glTexCoordFormatNV = (PFNGLTEXCOORDFORMATNVPROC) SMOL_GETGLPROC(glTexCoordFormatNV); 
+    glEdgeFlagFormatNV = (PFNGLEDGEFLAGFORMATNVPROC) SMOL_GETGLPROC(glEdgeFlagFormatNV); 
+    glSecondaryColorFormatNV = (PFNGLSECONDARYCOLORFORMATNVPROC) SMOL_GETGLPROC(glSecondaryColorFormatNV); 
+    glFogCoordFormatNV = (PFNGLFOGCOORDFORMATNVPROC) SMOL_GETGLPROC(glFogCoordFormatNV); 
+    glVertexAttribFormatNV = (PFNGLVERTEXATTRIBFORMATNVPROC) SMOL_GETGLPROC(glVertexAttribFormatNV); 
+    glVertexAttribIFormatNV = (PFNGLVERTEXATTRIBIFORMATNVPROC) SMOL_GETGLPROC(glVertexAttribIFormatNV); 
+    glGetIntegerui64i_vNV = (PFNGLGETINTEGERUI64I_VNVPROC) SMOL_GETGLPROC(glGetIntegerui64i_vNV); 
+    glViewportSwizzleNV = (PFNGLVIEWPORTSWIZZLENVPROC) SMOL_GETGLPROC(glViewportSwizzleNV); 
+    glFramebufferTextureMultiviewOVR = (PFNGLFRAMEBUFFERTEXTUREMULTIVIEWOVRPROC) SMOL_GETGLPROC(glFramebufferTextureMultiviewOVR); 
+  }
+}
+#endif // SMOL_ENGINE_IMPLEMENTATION
